% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.r
\name{annotate}
\alias{annotate}
\title{Annotations supporting NPC}
\usage{
annotate(
  geom,
  x = NULL,
  y = NULL,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  xend = NULL,
  yend = NULL,
  npcx = NULL,
  npcy = NULL,
  ...,
  na.rm = FALSE
)
}
\arguments{
\item{geom}{character Name of geom to use for annotation.}

\item{x, y, xmin, ymin, xmax, ymax, xend, yend, npcx, npcy}{numeric Positioning
aesthetics - you must specify at least one of these.}

\item{...}{Other named arguments passed on to \code{layer()}. These are often
aesthetics, used to set an aesthetic to a fixed value, like colour = "red"
or size = 3. They may also be parameters to the paired geom/stat.}

\item{na.rm}{logical If \code{FALSE}, the default, missing values are removed
with a warning. If TRUE, missing values are silently removed.}
}
\description{
A revised version of \code{annotate()} from package 'ggplot2' adding support
for \code{npcx} and \code{npcy} position aesthetics, allowing use of the
geometries defined in the current package such as \code{geom_text_npc()}.
When package 'ggpmisc' is loaded this definitions overrides that in package
'ggplot2'.
}
\details{
Note that all position aesthetics are scaled (i.e., they will
  expand the limits of the plot so they are visible), but all other
  aesthetics are set. This means that layers created with this function will
  never affect the legend.
}
\examples{

p <- ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point()
p + annotate("text_npc", npcx = .9, npcy = .9, label = "Some text")
p + annotate("text_npc", npcx = "right", npcy = "top", label = "Some text")
p + annotate("label_npc", npcx = c(.1, .9), npcy = c(.1, .9),
             label = c("A", "B"))

}
