% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-eq-label.R
\name{build_eq.x.rhs}
\alias{build_eq.x.rhs}
\alias{build_lhs}
\title{Left and right hand sides of model equations}
\usage{
build_eq.x.rhs(output.type = "expression", orientation = "x")

build_lhs(output.type = "expression", orientation = "x")
}
\arguments{
\item{output.type}{character One of "expression", "latex", "tex", "text",
"tikz", "markdown", "marquee".}

\item{orientation}{character \code{"x"} or \code{"y"}, indicating
the aesthetic onto which the explanatory variable is mapped.}
}
\value{
A \code{character} string.

A \code{character} string.
}
\description{
Left and right hand sides of model equations
}
\keyword{internal}
