% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-manip.R
\name{delete_layers}
\alias{delete_layers}
\alias{move_layers}
\alias{show_layers}
\alias{edit_layers}
\title{Layer manipulation}
\usage{
delete_layers(x, match_type, idx = NULL)

move_layers(x, match_type, position = "top", idx = NULL)

show_layers(x, match_type, idx = NULL)

edit_layers(x, match_type, idx = NULL, position = 0L, action)
}
\arguments{
\item{x}{an object of class \code{gg} to be operated upon.}

\item{match_type}{The name of the ggproto object class for the geom(s),
position(s) or stat(s) to be removed.}

\item{idx}{integer vector If not \code{NULL}, the default, it overrides
matching by class and applies the operation to the layers selected by the
indexing into the list of layers.}

\item{action}{character One of "delete", "move", or "show".}
}
\value{
An edited copy of x.

An integer vector with indexes to matched layers.
}
\description{
Delete or move one or more layers in a ggplot object.
}
\note{
These functions must be used with care as they select all layers matching
  the provided geom, position or stat ggproto object class. Layers added with
  a stat do use a geom, and vice versa.
}
\examples{
library(ggplot2)

df <- data.frame(
  gp = factor(rep(letters[1:3], each = 10)),
  y = rnorm(30)
)
p <- ggplot(df, aes(gp, y)) +
     geom_point() +
     stat_summary(fun.data = "mean_se", colour = "red")
p
delete_layers(p, "GeomPoint")
delete_layers(p, "StatSummary")
move_layers(p, "GeomPoint", "top")
move_layers(p, "GeomPointrange", "bottom")
move_layers(p, "StatSummary", "bottom")
move_layers(p, "GeomPointrange", 1L)
show_layers(p, "GeomPoint")

}
\references{
\url{https://stackoverflow.com/questions/13407236/remove-a-layer-from-a-ggplot2-chart}
}
\keyword{internal}
