% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-manip.R
\name{edit_layers}
\alias{edit_layers}
\title{Edit plot layer}
\usage{
edit_layers(x, match_type = NULL, position = 0L, idx = NULL, action)
}
\arguments{
\item{x}{an object of class \code{gg} to be operated upon.}

\item{match_type}{The name of the ggproto object class for the geom(s),
position(s) or stat(s) matching that of the layers to be operated upon.}

\item{idx}{integer vector Index(es) into the list of layers, used to select
the layers to be operated upon.}

\item{action}{character One of "delete", "move", or "which".}
}
\description{
Delete or move one or more layers in a ggplot object.
}
\keyword{internal}
