% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-fm.R
\name{extract_weights}
\alias{extract_weights}
\title{Extract prior and fitted weights}
\usage{
extract_weights(fm, n.row)
}
\arguments{
\item{fm}{a fitted model object of a supported class.}

\item{n.row}{interger The expected length of the weights vectors to extract.}
}
\value{
A list with two named members: \code{rob.weight.vals} the weights
  actually used to weight residuals, either user supplied or computed, and
  \code{weight.vals} the prior weights passed as argument. When not available
  the vectors are filled with \code{NA_real_} values.
}
\description{
Extract the prior and fitted weights from a fitted model object.
}
\note{
Called by \code{\link{stat_fit_residuals}()} and
  \code{\link{stat_fit_deviations}()}.
}
\keyword{internal}
