% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-continuous-logFC.R
\name{FC_name}
\alias{FC_name}
\title{Fold change- axis labels}
\usage{
FC_name(
  name = "Abundance\%unit",
  log.base = FALSE,
  format = getOption("photobiology.math", default = "R.expression")
)
}
\arguments{
\item{name}{character The axis-label or name of the scale.}

\item{log.base}{integer The base used for the logarithm transformation.}

\item{format}{character string, "R", "R.expression", "R.character", or
"LaTeX".}
}
\value{
a character string or an R expression.
}
\description{
Generate axis labels for different log-fold change on different log bases.
}
\examples{

FC_name()
FC_name(format = "R.expression")
FC_name(format = "LaTeX")
FC_name(name = "Metabolite concentration")
FC_name(log.base = 10)
FC_name(log.base = 0)

}
\keyword{internal}
