% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-fm.R
\name{fit_models_internal}
\alias{fit_models_internal}
\title{Apply model fit methods to data}
\usage{
fit_models_internal(
  data,
  method,
  method.name,
  method.args,
  n.min,
  formula,
  fit.seed,
  orientation,
  level = 0.95,
  accept.rq = TRUE
)
}
\arguments{
\item{data}{data.frame containing the variables in the model.}

\item{method, method.name}{function and character, respectively.}

\item{method.args}{named list with additional arguments. Not \code{data}
or \code{weights} which are always passed through aesthetic mappings.}

\item{n.min}{integer Minimum number of distinct values in the explanatory
variable (on the rhs of formula) for fitting to the attempted.}

\item{formula}{a formula object. Using aesthetic names \code{x} and \code{y}
instead of original variable names.}

\item{fit.seed}{RNG seed argument passed to
\code{\link[base:Random]{set.seed}()}. Defaults to \code{NA}, indicating
that \code{set.seed()} should not be called.}

\item{orientation}{character Either "x" or "y" controlling the default for
\code{formula}. The letter indicates the aesthetic considered the
explanatory variable in the model fit.}

\item{level}{numeric Value in 0..1 used for SMA and MA fits.}

\item{accept.rq}{logical Accept quantile regression fits with 'quantreg' or
warn when encountered.}
}
\value{
A list with three named members: \code{fm} the fitted model object
  and \code{method.args}, the arguments passed to the model fit function as a
  nested named list, \code{fit.seed} the seed used and \code{method.name},
  the name of the model fit function passed as arguemnt, which can differ
  from the class of \code{fm}.
}
\description{
Fit models using different methods translating some arguments
to make possible use of consistent arguments across calls to
stats.
}
\note{
Called by \code{\link{stat_fit_residuals}()},
  \code{\link{stat_fit_deviations}()} and \code{\link{stat_fit_fitted}()}.
}
\keyword{internal}
