% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-point-rug.r
\name{geom_x_margin_points}
\alias{geom_x_margin_points}
\alias{geom_y_margin_points}
\title{Rug-like dot plots in the margins}
\usage{
geom_x_margin_points(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., sides = "b", dot.shift = 0.015,
  na.rm = FALSE, show.legend = FALSE, inherit.aes = TRUE)

geom_y_margin_points(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., sides = "l", dot.shift = 0.015,
  na.rm = FALSE, show.legend = FALSE, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. Only needs
to be set at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{stat}{The statistical transformation to use on the data for this layer,
as a string.}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}

\item{sides}{A string that controls which sides of the plot the rugs appear on.
It can be set to a string containing any of `"trbl"`, for top, right,
bottom, and left.}

\item{dot.shift}{numeric value expressed in npc units for the shift of the
rug points inwards from the edge of the plotting area.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with a
warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}
}
\description{
A rug-like plot is a compact visualisation designed to supplement a 2d display
with the two 1d marginal annotations. Rug-like plots highligth individual
cases so are best used for marginal annotations.
}
\examples{
p <- ggplot(mtcars, aes(wt, mpg)) +
  geom_point()
p
p + geom_x_margin_points(data = data.frame(x = c(2.5, 4.5)),
                    mapping = aes(x))
p + geom_x_margin_points(sides="t")    # Rug on top only
p + geom_x_margin_points(shape = "triangle down filled", colour = "blue")
p + geom_x_margin_points(sides="trbl", shape = "circle small") # All four sides

# Use jittering to avoid overplotting for smaller datasets
ggplot(mpg, aes(displ, cty)) +
  geom_point() +
  geom_x_margin_points()

}
