% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predefined-mappings.R
\docType{data}
\name{label.sep}
\alias{label.sep}
\alias{eq_label}
\alias{eq.n_label}
\alias{rr.n_label}
\alias{rr.p_label}
\alias{adj.rr.n_label}
\alias{adj.rr.p_label}
\alias{eq.rr.n_label}
\alias{eq.adj.rr.n_label}
\alias{eq.p.n_label}
\alias{F.p.n_label}
\alias{F.p_label}
\title{Predefined label mappings}
\format{
An object of class \code{character} of length 1.

An object of class \code{uneval} of length 1.

An object of class \code{uneval} of length 1.

An object of class \code{uneval} of length 1.

An object of class \code{uneval} of length 1.

An object of class \code{uneval} of length 1.

An object of class \code{uneval} of length 1.

An object of class \code{uneval} of length 1.

An object of class \code{uneval} of length 1.

An object of class \code{uneval} of length 1.

An object of class \code{uneval} of length 1.

An object of class \code{uneval} of length 1.
}
\usage{
label.sep

eq_label

eq.n_label

rr.n_label

rr.p_label

adj.rr.n_label

adj.rr.p_label

eq.rr.n_label

eq.adj.rr.n_label

eq.p.n_label

F.p.n_label

F.p_label
}
\description{
Mappings of model-fit-derived text or expressions to the label aesthetic.
}
\examples{
# generate artificial data
set.seed(4321)
x <- 1:100
y <- (x + x^2 + x^3) + rnorm(length(x), mean = 0, sd = mean(x^3) / 4)
my.data <- data.frame(x = x,
                      y = y * 1e-5,
                      group = c("A", "B"),
                      y2 = y * 1e-5 + c(2, 0))

# give a name to a formula
formula <- y ~ poly(x, 3, raw = TRUE)

# using predefined label mappings
ggplot(my.data, aes(x, y2, colour = group)) +
  geom_point() +
  stat_poly_line(formula = formula) +
  stat_poly_eq(rr.n_label, formula = formula)

ggplot(my.data, aes(x, y2, colour = group)) +
  geom_point() +
  stat_poly_line(formula = formula) +
  stat_poly_eq(eq.adj.rr.n_label, formula = formula, size = 3)

ggplot(my.data, aes(x, y2, colour = group)) +
  geom_point() +
  stat_poly_line(formula = formula) +
  stat_poly_eq(F.p_label, formula = formula)

ggplot(my.data, aes(x, y)) +
  stat_quant_band(formula = formula) +
  stat_quant_eq(eq.n_label, formula = formula, size = 3) +
  geom_point()

ggplot(my.data, aes(x, y)) +
  stat_ma_line() +
  stat_ma_eq(eq.n_label) +
  geom_point()

}
\keyword{datasets}
