% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moved-to-gginnards.R
\name{Moved}
\alias{Moved}
\alias{geom_null}
\alias{stat_debug_group}
\alias{stat_debug_panel}
\alias{geom_debug}
\alias{append_layers}
\alias{bottom_layer}
\alias{delete_layers}
\alias{extract_layers}
\alias{move_layers}
\alias{num_layers}
\alias{shift_layers}
\alias{top_layer}
\alias{which_layers}
\title{Moved to package 'gginnards'}
\description{
Some stats, geoms and the plot layer manipulation functions have been
moved from package 'ggpmisc' to a separate new package called 'gginnards'.
}
\details{
To continue using any of these functions and methods, simply run at
  the R prompt or add to your script \code{library(gginnards)}, after
  installing package 'gginnards'.
}
\seealso{
\code{\link[gginnards]{gginnards-package}},
  \code{\link[gginnards]{geom_null}},
  \code{\link[gginnards]{stat_debug_group}},
  \code{\link[gginnards]{stat_debug_panel}},
  \code{\link[gginnards]{geom_debug}} and
  \code{\link[gginnards]{delete_layers}}.
}
