% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{quadrant_example.df}
\alias{quadrant_example.df}
\title{Example gene expression data}
\format{A \code{data.frame} object with 6088 rows and 6 variables}
\usage{
quadrant_example.df
}
\description{
A dataset containing reshaped and simplified output from an
  analysis of data from RNAseq done with package edgeR.
}
\examples{
names(quadrant_example.df)
head(quadrant_example.df)

}
\seealso{
Other Spectral data examples: \code{\link{volcano_example.df}}
}
\concept{Spectral data examples}
\keyword{datasets}
