% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed.R
\name{stat_quadrat_counts}
\alias{stat_quadrat_counts}
\title{Renamed to stat_quadrant_counts}
\description{
Function \code{stat_quadrant_counts()} was initially named
  \code{stat_quadrat_counts()} by mistake, and later its name corrected
  to \code{stat_quadrant_counts()}.
}
