% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-continuous-Pvalue.r
\name{reverselog_trans}
\alias{reverselog_trans}
\title{Reverse log transformation}
\usage{
reverselog_trans(base = exp(1))
}
\arguments{
\item{base}{numeric Base of logarithm}
}
\description{
Reverse log transformation with support for different logarithm
bases as used for transforming \emph{P}-values.
}
\details{
The transformation is given by
\deqn{-log[base](x)}

For example, this transformation is used with \code{base = 10} to convert
\eqn{P = 10^{-6}} into 6.
}
\keyword{internal}
