% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-apply.R
\name{stat_apply_group}
\alias{stat_apply_group}
\alias{stat_apply_panel}
\title{Apply a function to x or y values}
\usage{
stat_apply_group(mapping = NULL, data = NULL, geom = "line",
  .fun.x = NULL, .fun.x.args = list(), .fun.y = NULL,
  .fun.y.args = list(), position = "identity", na.rm = FALSE,
  show.legend = FALSE, inherit.aes = TRUE, ...)

stat_apply_panel(mapping = NULL, data = NULL, geom = "line",
  .fun.x = NULL, .fun.x.args = list(), .fun.y = NULL,
  .fun.y.args = list(), position = "identity", na.rm = FALSE,
  show.legend = FALSE, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. Only needs
to be set
 at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{.fun.x, .fun.y}{function to be applied or the name of the function to be
applied as a character string. One and only one of these parameters should
be passed a non-null argument.}

\item{.fun.x.args, .fun.y.args}{additional arguments to be passed to
the function as a named list.}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{na.rm}{a logical value indicating whether NA values should be
stripped before the computation proceeds.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2]{borders}}.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This can
include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}
}
\description{
\code{stat_apply_group} and \code{stat_apply_panel} apply functions to data.
}
\details{
The function to be applied is expected to be vectorized and to
  return a vector of the same length. The vector mapped to the x or y
  aesthetic is passed as the first positional argument to the call. The
  function must accept as first argument a vector or list that matches the
  data.
}
\note{
This stat is at early stages of development and its interface may
  change at any time, or it may be even removed from the package without
  previous notice.
}
\section{Computed variables}{

One x or y of is replaced by the vector returned by the applied function.
\describe{
  \item{x}{x-value as returned by \code{.fun.x}}
  \item{y}{y-value as returned by \code{.fun.y}}
}
}

\examples{

library(ggplot2)
df <- data.frame(X = rep(1:20,2),
                 Y = runif(40),
                 category = rep(c("A","B"), each = 20))
ggplot(df, aes(x = X, y = Y, colour = category)) +
  stat_apply_group(.fun.y = cumsum)
ggplot(df, aes(x = X, y = Y, colour = category)) +
  geom_point() +
  stat_apply_group(.fun.y = runmed, .fun.y.args = list(k = 5))
ggplot(df, aes(x = X, y = Y, colour = category)) +
  stat_apply_panel(.fun.y = function(x) {(x - min(x)) / (max(x) - min(x))})

}
\references{
Answers question "R ggplot on-the-fly calculation by grouping variable" at
\url{https://stackoverflow.com/questions/51412522}.
}
