% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-cor-zzzz.R
\name{stat_corr}
\alias{stat_corr}
\title{Annotate plot with correlation test}
\usage{
stat_corr(
  mapping = NULL,
  data = NULL,
  geom = "text_npc",
  position = "identity",
  ...,
  method = "pearson",
  alternative = "two.sided",
  exact = NULL,
  conf.level = 0.95,
  continuity = FALSE,
  small.r = FALSE,
  small.p = FALSE,
  coef.keep.zeros = TRUE,
  r.digits = 2,
  t.digits = 3,
  p.digits = 3,
  label.x = "left",
  label.y = "top",
  hstep = 0,
  vstep = NULL,
  output.type = NULL,
  na.rm = FALSE,
  parse = NULL,
  show.legend = FALSE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. Only needs to be
set at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset, only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points on this
layer}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}

\item{method}{character One of "pearson", "kendall" or "spearman".}

\item{alternative}{character One of "two.sided", "less" or "greater".}

\item{exact}{logical Whether an exact p-value should be computed. Used for
Kendall's tau and Spearman's rho.}

\item{conf.level}{numeric Confidence level for the returned confidence
interval (only if \code{method = "pearson"} which is the default).}

\item{continuity}{logical If TRUE , a continuity correction is used for
Kendall's tau and Spearman's rho when not computed exactly.}

\item{small.r, small.p}{logical Flags to switch use of lower case r and p for
coefficient of correlation (only for \code{method = "pearson"}) and p-value.}

\item{coef.keep.zeros}{logical Keep or drop trailing zeros when formatting
the fitted coefficients and F-value.}

\item{r.digits, t.digits, p.digits}{integer Number of digits after the decimal point to
use for R, r.squared, tau or rho and P-value in labels.}

\item{label.x, label.y}{\code{numeric} with range 0..1 "normalized parent
coordinates" (npc units) or character if using \code{geom_text_npc()} or
\code{geom_label_npc()}. If using \code{geom_text()} or \code{geom_label()}
numeric in native data units. If too short they will be recycled.}

\item{hstep, vstep}{numeric in npc units, the horizontal and vertical step
used between labels for different groups.}

\item{output.type}{character One of "expression", "LaTeX", "text",
"markdown" or "numeric".}

\item{na.rm}{a logical indicating whether NA values should be stripped before
the computation proceeds.}

\item{parse}{logical Passed to the geom. If \code{TRUE}, the labels will be
parsed into expressions and displayed as described in \code{?plotmath}.
Default is \code{TRUE} if \code{output.type = "expression"} and
\code{FALSE} otherwise.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}
}
\description{
\code{stat_corr} applies \code{stats::cor.test()} respecting grouping
with \code{method = "pearson"} default but
alternatively using \code{"kendall"} or \code{"spearman"} methods. It
generates labels for correlation coefficients and p-value, coefficient of
determination (R^2) for method "pearson" and number of observations.
}
\details{
This statistic can be used to annotate a plot with the correlation
  coefficient and the outcome of its test of significance. It supports
  Pearson, Kendall and Spearman methods to compute correlation. This stat
  generates labels as R expressions by default but LaTeX (use TikZ device),
  markdown (use package 'ggtext') and plain text are also supported, as well
  as numeric values for user-generated text labels. The value of \code{parse}
  is set automatically based on \code{output-type}, but if you assemble
  labels that need parsing from \code{numeric} output, the default needs to
  be overridden.

  A ggplot statistic receives as \code{data} a data frame that is not the one
  passed as argument by the user, but instead a data frame with the variables
  mapped to aesthetics. \code{cor.test()} is always applied to the variabled
  mapped to the \code{x} and \code{y} aesthetics, so the scales used for
  \code{x} and \code{y} should both be continuous scales rather than
  discrete.
}
\examples{
# generate artificial data
set.seed(4321)
x <- (1:100) / 10
y <- x + rnorm(length(x))
my.data <- data.frame(x = x,
                      y = y,
                      y.desc = - y,
                      group = c("A", "B"))

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_corr()

ggplot(my.data, aes(x, y.desc)) +
  geom_point() +
  stat_corr(label.x = "right")

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_corr(aes(label = paste(after_stat(r.label),
                              after_stat(p.value.label),
                              after_stat(n.label),
                              sep = "*\"; \"*")))

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_corr(small.r = TRUE)

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_corr(aes(label = paste(after_stat(r.label),
                              after_stat(p.value.label),
                              after_stat(n.label),
                              sep = "*\"; \"*")),
            method = "kendall")

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_corr(method = "kendall")

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_corr(method = "spearman")

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_corr(aes(label = paste(after_stat(r.label),
                              after_stat(p.value.label),
                              after_stat(n.label),
                              sep = "*\"; \"*")),
            method = "spearman")

}
\seealso{
\code{\link[stats]{cor.test}} for details on the computations.
}
