% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-correlation.R
\name{stat_correlation}
\alias{stat_correlation}
\title{Correlation test annotations}
\usage{
stat_correlation(
  mapping = NULL,
  data = NULL,
  geom = "text_npc",
  position = "identity",
  ...,
  method = "pearson",
  n.min = 2L,
  alternative = "two.sided",
  exact = NULL,
  r.conf.level = ifelse(method == "pearson", 0.95, NA),
  continuity = FALSE,
  fit.seed = NA,
  small.r = getOption("ggpmisc.small.r", default = FALSE),
  small.p = getOption("ggpmisc.small.p", default = FALSE),
  coef.keep.zeros = TRUE,
  r.digits = 2,
  t.digits = 3,
  p.digits = 3,
  CI.brackets = c("[", "]"),
  label.x = "left",
  label.y = "top",
  hstep = 0,
  vstep = NULL,
  output.type = NULL,
  boot.R = ifelse(method == "pearson", 0, 999),
  na.rm = FALSE,
  parse = NULL,
  show.legend = FALSE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}()}. Only needs to be set at the layer level if you
are overriding the plot defaults.}

\item{data}{A layer specific dataset, only needed if you want to override the
plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points on this
layer.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}

\item{method}{character One of "pearson", "kendall" or "spearman".}

\item{n.min}{integer Minimum number of distinct values in the explanatory
variable (on the rhs of formula) for fitting to the attempted.}

\item{alternative}{character One of "two.sided", "less" or "greater".}

\item{exact}{logical Whether an exact p-value should be computed. Used for
Kendall's tau and Spearman's rho.}

\item{r.conf.level}{numeric Confidence level for the returned confidence
interval. If set to \code{NA} computation of CI is skipped.}

\item{continuity}{logical If TRUE , a continuity correction is used for
Kendall's tau and Spearman's rho when not computed exactly.}

\item{fit.seed}{RNG seed argument passed to
\code{\link[base:Random]{set.seed}()}. Defaults to \code{NA}, indicating
that \code{set.seed()} should not be called.}

\item{small.r, small.p}{logical Flags to switch use of lower case r and p for
coefficient of correlation (only for \code{method = "pearson"}) and
p-value.}

\item{coef.keep.zeros}{logical Keep or drop trailing zeros when formatting
the correlation coefficients and t-value, z-value or S-value (see note
below).}

\item{r.digits, t.digits, p.digits}{integer Number of digits after the decimal
point to use for R, r.squared, tau or rho and P-value in labels. If
\code{Inf}, use exponential notation with three decimal places.}

\item{CI.brackets}{character vector of length 2. The opening and closing
brackets used for the CI label.}

\item{label.x, label.y}{\code{numeric} with range 0..1 "normalized parent
coordinates" (npc units) or character if using \code{geom_text_npc()} or
\code{geom_label_npc()}. If using \code{geom_text()} or \code{geom_label()}
numeric in native data units. If too short they will be recycled.}

\item{hstep, vstep}{numeric in npc units, the horizontal and vertical step
used between labels for different groups.}

\item{output.type}{character One of "expression", "text", "markdown",
"marquee", "latex", "latex.eqn", "latex.deqn" or "numeric".}

\item{boot.R}{interger The number of bootstrap resamples. Set to zero for no
bootstrap estimates for the CI.}

\item{na.rm}{a logical indicating whether NA values should be stripped before
the computation proceeds.}

\item{parse}{logical Passed to the geom. If \code{TRUE}, the labels will be
parsed into expressions and displayed as described in
\code{\link[grDevices]{plotmath}}. Default is \code{TRUE} if
\code{output.type = "expression"} and \code{FALSE} otherwise.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}
}
\description{
Statistic \code{stat_correlation()} applies \code{stats::cor.test()}
respecting grouping with \code{method = "pearson"} default but alternatively
using \code{"kendall"} or \code{"spearman"} methods. It adds textual labels
to a plot.
}
\details{
This statistic can be used to annotate a plot with the correlation
  coefficient and the outcome of its test of significance. It supports
  Pearson, Kendall and Spearman methods to compute correlation. This
  statistic generates labels as R expressions by default but LaTeX (use TikZ
  device), markdown (use package 'ggtext') and plain text are also supported,
  as well as numeric values for user-generated text labels. The character
  labels include the symbol describing the quantity together with the numeric
  value. For the confidence interval (CI) the default is to follow the APA
  recommendation of using square brackets. As the CI is computed by
  bootstrapping, \code{fit.seed} if different to \code{NA} immediately before
  this computation.

  The value of \code{parse} is set automatically based on \code{output-type},
  but if you assemble labels that need parsing from \code{numeric} output,
  the default needs to be overridden. By default the value of
  \code{output.type} is guessed from the name of the geometry.

  A ggplot statistic receives as \code{data} a data frame that is not the one
  passed as argument by the user, but instead a data frame with the variables
  mapped to aesthetics. \code{cor.test()} is always applied to the variables
  mapped to the \code{x} and \code{y} aesthetics, so the scales used for
  \code{x} and \code{y} should both be continuous scales rather than
  discrete.
}
\note{
Currently \code{coef.keep.zeros} is ignored, with trailing zeros always
  retained in the \code{character} labels returned but not protected from
  being dropped by R when these \code{character} strings are parsed into
  plotmath expressions (i.e., when \code{output.type = "expression"}).
}
\section{Computed variables}{


 If output.type is \code{"numeric"} the returned
  tibble contains the columns listed below with variations depending on the
  \code{method}. If the model fit function used does not return a value, the
  variable is set to \code{NA_real_}.
\describe{
  \item{x,npcx}{x position}
  \item{y,npcy}{y position}
  \item{r, and cor, tau or rho}{numeric values for correlation coefficient estimates}
  \item{t.value and its df, z.value or S.value }{numeric values for statistic estimates}
  \item{p.value, n}{numeric values.}
  \item{r.conf.level}{numeric value, as fraction of one.}
  \item{r.confint.low}{Confidence interval limit for \code{r}.}
  \item{r.confint.high}{Confidence interval limit for \code{r}.}
  \item{grp.label}{Set according to mapping in \code{aes}.}
  \item{method.label}{Set according \code{method} used.}
  \item{method, test}{character values}}

If output.type different from \code{"numeric"} the returned tibble contains
in addition to the columns listed above those listed below. If the numeric
value is missing the label is set to \code{character(0L)}.

\describe{
  \item{r.label, and cor.label, tau.label or rho.label}{Correlation coefficient as a character string.}
  \item{t.value.label, z.value.label or S.value.label}{t-value and degrees of freedom, z-value or S-value as a character string.}
  \item{p.value.label}{P-value for test against zero, as a character string.}
  \item{r.confint.label, and cor.conint.label, tau.confint.label or rho.confint.label}{Confidence interval for \code{r} (only with \code{method = "pearson"}).}
  \item{n.label}{Number of observations used in the fit, as a character string.}
  \item{grp.label}{Set according to mapping in \code{aes}, as a character string.}}

To explore the computed values returned for a given input we suggest the use
of \code{\link[gginnards]{geom_debug}} as shown in the last examples below.
}

\section{Position of labels}{

  When data are grouped by mapping a factor to an aesthetic, e.g.,
  \code{colour}, \code{shape} and/or \code{linetype} the model is fitted
  separately to each group, and for each group a whole set of labels is
  generated. If the argument passed to \code{label.y} is a vector of length
  1, this value determines the position of the equation and/or other labels
  for the first group, and the positions of the labels for the remaining
  groups are generated by adding \code{vspace} based on the group number.
  If the argument passed to \code{label.y} is a vector of length > 1, it is
  used unchanged, possibly extended by recycling, ignoring \code{vstep}.

  If the labels are rotated by 90 degrees then the automatic stepping is
  best based on \code{hstep} with \code{vstep = 0}. Similarly as described
  above, if \code{label.x} is a vector of length > 1, it is
  used unchanged, possibly extended by recycling, ignoring \code{hstep}.

  When using facets and with a grouping that does not repeat in each panel,
  the automatic positioning in most cases will not be the desired one. Manual
  positioning using a vector of length > 1 for \code{label.x} and/or
  \code{label.y} is the currently available workaround.
}

\section{Output types}{
 The formatting of character strings to be displayed
in plots are marked as mathematical equations. Depending on the geom used,
the mark-up needs to be encoded differently, or in some cases mark-up not
applied.
\describe{
  \item{\code{"expression"}}{The labels are encoded as character strings to be parsed into R's plotmath expressions.}
  \item{\code{"LaTeX", "TeX", "tikz", "latex"}}{The labels are encoded as 'LaTeX' maths equations, without the "fences" for switching in math mode.}
  \item{\code{"latex.eqn"}}{Same as \code{"latex"} but enclosed in single \code{$}, i.e., as in-line maths.}
  \item{\code{"latex.deqn"}}{Same as \code{"latex"} but enclosed in double \code{$$}, i.e., as display maths.}
  \item{\code{"markdown"}}{The labels are encoded as character strings using markdown syntax, with some embedded HTML.}
  \item{\code{"marquee"}}{The labels are encoded as character strings using markdown syntax, with 'marquee' supported spans.}
  \item{\code{"text"}}{The labels are plain ASCII character strings.}
  \item{\code{"numeric"}}{No labels are generated. This value is accepted by the statistics, but not by the label formatting functions.}
  \item{\code{NULL}}{The value used depends on the argument passed to \code{geom}.}}

If \code{geom = "latex"} (package 'xdvir') the output type used is
\code{"latex.eqn"}. If \code{geom = "richtext"} (package 'ggtext') or
\code{geom = "textbox"} (package 'ggtext') the output type used is
\code{"markdown"}. If \code{geom = "marquee"} (package 'marquee') the output
type used is \code{"marquee"}. For all other values of \code{geom} the default
is \code{"expression"}. Invalid values as argument trigger an error.
}

\section{Which variables are available for mapping?}{

  Computed variables and their names can vary depending on the \code{method}
  used to fit a model or the \code{output.type} in use. They can also depend
  for a given \code{method} on other arguments passed when fitting a model or
  extracting estimates and other computed values. In many cases, when values
  are not available, the variables are filled with \code{NA} values.

  In the statistics returning formatted strings for use as annotations, a
  message is issued by default in interactive R sessions, listing the short
  names for available formatted labels as recognized by functions
  \code{\link{use_label}()} and \code{\link{f_use_label}()}, except when
  \code{output.type = "numeric"} is passed, in which case the names of all
  variables accessible by \code{after_stat()} within a call to \code{aes()}
  are listed. This default (\code{"nicknames"}) can be changed by setting R option
  \code{"ggpmisc.stat.vars.message"} to one of \code{"names"},
  \code{"colnames"} or \code{"none"}.

  In the statistics that plot a prediction or more generally mainly return
  numeric variables, a message is issued by default in interactive R
  sessions, listing the names of all variables accessible by
  \code{after_stat()} within a call to \code{aes()} with at least some
  non-missing values. This default (\code{"colnames"}) can be changed by
  setting R option \code{"ggpmisc.stat.vars.message"} to \code{"none"}.
}

\examples{
# generate artificial data
set.seed(4321)
x <- (1:100) / 10
y <- x + rnorm(length(x))
my.data <- data.frame(x = x,
                      y = y,
                      y.desc = - y,
                      group = c("A", "B"))

# by default only R is displayed
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_correlation()

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_correlation(small.r = TRUE)

ggplot(my.data, aes(x, y.desc)) +
  geom_point() +
  stat_correlation(label.x = "right")

# non-default methods
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_correlation(method = "kendall")

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_correlation(method = "spearman")

# use_label() can map a user selected label
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_correlation(use_label("R2"))

# use_label() can assemble and map a combined label
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_correlation(use_label("R", "P", "n", "method"))

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_correlation(use_label("R", "R.CI"))

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_correlation(use_label("R", "R.CI"),
                   r.conf.level = 0.95)

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_correlation(use_label("R", "R.CI"),
                   method = "kendall",
                   r.conf.level = 0.95)

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_correlation(use_label("R", "R.CI"),
                   method = "spearman",
                   r.conf.level = 0.95)

# f_use_label() provides additional flexibility
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_correlation(
    f_use_label("R", "R.CI",
      format = "\"Estimate: \"*\%s*\" with confidence \"*\%s"),
      method = "spearman", r.conf.level = 0.95)

# manually assemble and map a specific label using paste() and aes()
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_correlation(aes(label =
                         after_stat(
                           paste(r.label, p.value.label, n.label,
                                 sep = "*\", \"*"))))

# manually format and map a specific label using sprintf() and aes()
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_correlation(aes(label =
                         after_stat(
                           sprintf("\%s*\" with \"*\%s*\" for \"*\%s",
                                   r.label, p.value.label, t.value.label))))

# Inspecting the returned data using geom_debug_group()
# This provides a quick way of printing the returned data frame.

gginnards.installed <- requireNamespace("gginnards", quietly = TRUE)

if (gginnards.installed)
  library(gginnards)

# the whole of computed data
if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    geom_point() +
    stat_correlation(geom = "debug_group")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    geom_point() +
    stat_correlation(geom = "debug_group", method = "pearson")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    geom_point() +
    stat_correlation(geom = "debug_group", output.type = "numeric")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    geom_point() +
    stat_correlation(geom = "debug_group", output.type = "markdown")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    geom_point() +
    stat_correlation(geom = "debug_group", output.type = "LaTeX")

}
\seealso{
\code{\link[stats]{cor.test}} for details on the computations.
}
\concept{statistics for correlation-test annotation}
\section{Aesthetics}{

\code{stat_correlation()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → inferred \cr
 • \tab \code{grp.label} \tab   \cr
 • \tab \code{hjust} \tab → \code{"inward"} \cr
 • \tab \code{label} \tab → \code{after_stat(r.label)} \cr
 • \tab \code{npcx} \tab → \code{after_stat(npcx)} \cr
 • \tab \code{npcy} \tab → \code{after_stat(npcy)} \cr
 • \tab \code{vjust} \tab → \code{"inward"} \cr
}

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}
