% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-distrmix-line.R
\name{stat_distrmix_line}
\alias{stat_distrmix_line}
\title{Predicted line from distribution mixture model fit}
\usage{
stat_distrmix_line(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  ...,
  orientation = "x",
  method = "normalmixEM",
  se = NULL,
  fit.seed = NA,
  fm.values = FALSE,
  n = min(100 + 50 * k, 300),
  fullrange = TRUE,
  level = 0.95,
  method.args = list(),
  k = 2,
  free.mean = TRUE,
  free.sd = TRUE,
  components = "all",
  n.min = 10L * k,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}}. Only needs to be set at the layer level if you
are overriding the plot defaults.}

\item{data}{A layer specific dataset, only needed if you want to override the
plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points on this
layer.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}

\item{orientation}{character Either "x" or "y", the mapping of the values
to which the mixture model is to be fitetd. NOT YET IMPLEMENTED!}

\item{method}{function or character If character, "normalmixEM"
or the name of a model fit function are accepted, possibly followed by the
fit function's \code{method} argument separated by a colon. The function
must return a model fit object of class \code{mixEM}.}

\item{se}{Currently ignored.}

\item{fit.seed}{RNG seed argument passed to \code{\link[base:Random]{set.seed}()}.
Defaults to \code{NA}, which means that \code{set.seed()} will not be
called.}

\item{fm.values}{logical Add parameter estimates and their standard errors
to the returned values (`FALSE` by default.)}

\item{n}{Number of points at which to evaluate the model prediction.}

\item{fullrange}{Should the prediction span the combined range of the scale
and of the fitted distributions, or just span the range of the data?}

\item{level}{Level of confidence interval to use (0.95 by default).}

\item{method.args}{named list with additional arguments.}

\item{k}{integer Number of mixture components to fit.}

\item{free.mean, free.sd}{logical If TRUE, allow the fitted \code{mean} and/or
fitted \code{sd} to vary among the component Normal distributions.}

\item{components}{character One of \code{"all"}, \code{"sum"}, or
\code{"members"} select which densities are returned.}

\item{n.min}{integer Minimum number of distinct values in the mapped
variable for fitting to the attempted.}

\item{na.rm}{a logical indicating whether NA values should be stripped before
the computation proceeds.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}
}
\value{
The value returned by the statistic is a data frame, with \code{n}
  rows of predicted density for each component of the mixture plus their
  sum and the corresponding vector of \code{x} values. Optionally it will
  also include additional values related to the model fit.
}
\description{
\code{stat_distrmix_line()} fits a Normal mixture model, by default with
\code{\link[mixtools]{normalmixEM}()}. Predicted values are
computed and, by default, plotted.
}
\details{
This statistic is similar to \code{\link[ggplot2]{stat_density}} but
  instead of fitting a single distribution it can fit a mixture of two or
  more Normal distributions, using an approach related to clustering.
  Defaults are consistent between \code{stat_distrmix_line()} and
  \code{stat_distrmix_eq()}. Parameter \code{fit.seed} if not \code{NA} is used
  in a call to \code{set.seed()} immediately before calling the model fit
  function. As the fitting procedure makes use of the (pseudo-)random number
  generator (RNG), convergence can depend on it, and in such cases setting
  \code{fit.seed} to the same value in \code{\link{stat_distrmix_line}()} and in
  \code{\link{stat_distrmix_eq}()} can ensure consistency, and more
  generally, reproducibility.

  A mixture model as described above, is fitted for \code{k >= 2}, while
  \code{k == 1} is treated as a special case and a Normal distribution fitted
  with function \code{\link[MASS]{fitdistr}()}. In this case the SE values
  are exact estimates.
}
\section{Computed variables}{
 \code{stat_distrmix_line()} provides the following
  variables, some of which depend on the orientation:
  \describe{\item{density}{predicted density values}
  \item{x}{the \code{n} values for the quantiles}
  \item{component}{A factor indexing the components and/or their sum}}

  If \code{fm.values = TRUE} is passed then columns with diagnosis and
  parameters estimates are added, with the same value in each row within a
  group:
  \describe{\item{converged}{\code{logical} indicating if convergence was achieved}
  \item{n}{\code{numeric} the number of \code{x} values}
  \item{.size}{\code{numeric} the number of \code{density} values}
  \item{fm.class}{\code{character} the most derived class of the fitted model object}
  \item{fm.method}{\code{character} the method, as given by the \code{ft}
  field of the fitted model objects}}
  This is wasteful and disabled by default, but provides a simple and robust
  approach to achieve effects like colouring or hiding of the model fit line
  by group depending on the outcome of model fitting.
}

\examples{
ggplot(faithful, aes(x = waiting)) +
  stat_distrmix_line()

# ggplot(faithful, aes(y = waiting)) +
#  stat_distrmix_line(orientation = "y")

ggplot(faithful, aes(x = waiting)) +
  stat_distrmix_line(components = "sum")

ggplot(faithful, aes(x = waiting)) +
  stat_distrmix_line(components = "members")

ggplot(faithful, aes(x = waiting)) +
 geom_histogram(aes(y = after_stat(density)), bins = 20) +
 stat_distrmix_line(aes(colour = after_stat(component),
                         fill = after_stat(component)),
                     geom = "area", linewidth = 1, alpha = 0.25, se = FALSE)

ggplot(faithful, aes(x = waiting)) +
 stat_distrmix_line(aes(colour = after_stat(component),
                         fill = after_stat(component)),
                     geom = "area", linewidth = 1, alpha = 0.25,
                     components = "members", se = FALSE)

ggplot(faithful, aes(x = waiting)) +
 stat_distrmix_line(geom = "area", linewidth = 1, alpha = 0.25,
                     colour = "black", outline.type = "upper",
                     components = "sum", se = FALSE)

# special case of no mixture
ggplot(subset(faithful, waiting > 66), aes(x = waiting)) +
  stat_distrmix_line(k = 1)

# Inspecting the returned data using geom_debug_group()
gginnards.installed <- requireNamespace("gginnards", quietly = TRUE)

if (gginnards.installed)
  library(gginnards)

if (gginnards.installed)
  ggplot(faithful, aes(x = waiting)) +
    stat_distrmix_line(geom = "debug_group", components = "all")

if (gginnards.installed)
  ggplot(faithful, aes(x = waiting)) +
    stat_distrmix_line(geom = "debug_group", components = "sum")

if (gginnards.installed)
  ggplot(faithful, aes(x = waiting)) +
    stat_distrmix_line(geom = "debug_group", components = "members")

if (gginnards.installed)
  ggplot(faithful, aes(x = waiting)) +
    stat_distrmix_line(geom = "debug_group", fm.values = TRUE)

}
\seealso{
Other ggplot statistics for mixture model fits.: 
\code{\link{stat_distrmix_eq}()}
}
\concept{ggplot statistics for mixture model fits.}
\section{Aesthetics}{

\code{stat_distrmix_line()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}} \emph{or} \code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → \code{after_stat(component)} \cr
}

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}
