% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try-data-frame.R
\name{try_data_frame}
\alias{try_data_frame}
\title{Convert an R object to a data frame}
\usage{
try_data_frame(x, time.resolution = "second", as.numeric = FALSE)
}
\arguments{
\item{x}{An R object}

\item{time.resolution}{character The time unit to which the returned time
values will be rounded.}

\item{as.numeric}{logical If TRUE convert time to numeric, expressed as
fractional calendar years.}
}
\value{
A dataframe.
}
\description{
This functions tries to convert any R object into a data.frame object.
If \code{x} is already a data.frame, it is returned as is. If it is
a list or a vector it is converted by means of \code{as.data.frame()}.
If of any other type, a conversion into an object of class \code{xts} is
attempted by means of \code{try.xts()} and if succesful the \code{xts}
object is converted into a data frame with a variable \code{time}
containing times as \code{POSIXct} and the remaining data columns with
the time series data. In this conversion row names are stripped.
}
\note{
This function can be used to easily plot time series data with ggplot2.
}
\examples{
library(xts)
try_data_frame(lynx)
try_data_frame(lynx, "year")
try_data_frame(austres)
try_data_frame(austres, "quarter")
try_data_frame(cars)
try_data_frame(photobiology::sun.spct)

}

