% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{volcano_example.df}
\alias{volcano_example.df}
\title{Example gene expression data}
\format{A \code{data.frame} object with 1218 rows and 5 variables}
\usage{
volcano_example.df
}
\description{
A dataset containing reshaped and simplified output from an
  analysis of data from RNAseq done with package edgeR.
}
\examples{
names(volcano_example.df)
head(volcano_example.df)

}
\seealso{
Other Spectral data examples: \code{\link{quadrant_example.df}}
}
\concept{Spectral data examples}
\keyword{datasets}
