% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-fm.R
\name{check_limit_to}
\alias{check_limit_to}
\title{Check limit.to and fullrange arguments}
\usage{
check_limit_to(fullrange, limit.to = NULL, orientation = "x")
}
\arguments{
\item{fullrange}{logical Should the fit prediction span the full
range of the plot, or just the range of the explanatory variable?}

\item{limit.to}{character or numeric If character one of \code{""},
\code{"x"}, \code{"y"} or \code{"xy"}. Should the fit prediction be
constrained to the range of the variables mapped to \code{x} and/or
\code{y} in each data group? If numeric, the \emph{new data} values to use
for the explanatory variable when computing the predicted line and
confidence band. When set, \code{limit.to} silently overrides
\code{fullrange}!}

\item{orientation}{character Either "x" or "y" controlling the default for
\code{formula}. The letter indicates the aesthetic considered the
explanatory variable in the model fit.}
}
\description{
Implement backwards compatibility and support override of
\code{fullrange} by \code{limit.to}.
}
\keyword{internal}
