% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{check_output_type}
\alias{check_output_type}
\title{Validate output type}
\usage{
check_output_type(
  output.type,
  geom = "text",
  supported.types = c("expression", "text", "markdown", "marquee", "numeric", "latex",
    "latex.eqn", "latex.deqn")
)
}
\arguments{
\item{output.type}{character User-set argument or default from stat}

\item{geom}{character The name of the geom that will be used to render the
labels.}

\item{supported.types}{character vector of accepted values for user input.}
}
\value{
If \code{output.type} is \code{NULL} a suitable value based on the
  name of the geom is returned, defaulting to "expression". If not
  \code{NULL}, the value is passed through unchanged.
}
\description{
Replace \code{NULL} \code{output.type} based on \code{geom} and validate
other values. Convert synonyms and change into lower case mal-formed
input.
}
\section{Output types}{
 The formatting of character strings to be displayed
in plots are marked as mathematical equations. Depending on the geom used,
the mark-up needs to be encoded differently, or in some cases mark-up not
applied.
\describe{
  \item{\code{"expression"}}{The labels are encoded as character strings to be parsed into R's plotmath expressions.}
  \item{\code{"LaTeX", "TeX", "tikz", "latex"}}{The labels are encoded as 'LaTeX' maths equations, without the "fences" for switching in math mode.}
  \item{\code{"latex.eqn"}}{Same as \code{"latex"} but enclosed in single \code{$}, i.e., as in-line maths.}
  \item{\code{"latex.deqn"}}{Same as \code{"latex"} but enclosed in double \code{$$}, i.e., as display maths.}
  \item{\code{"markdown"}}{The labels are encoded as character strings using markdown syntax, with some embedded HTML.}
  \item{\code{"marquee"}}{The labels are encoded as character strings using markdown syntax, with 'marquee' supported spans.}
  \item{\code{"text"}}{The labels are plain ASCII character strings.}
  \item{\code{"numeric"}}{No labels are generated. This value is accepted by the statistics, but not by the label formatting functions.}
  \item{\code{NULL}}{The value used, \code{expression}, \code{latex.eqn} or \code{markup} depends on the argument passed to \code{geom}.}}

If \code{geom = "latex"} (package 'xdvir') the output type used is
\code{"latex.eqn"}. If \code{geom = "richtext"} (package 'ggtext') or
\code{geom = "textbox"} (package 'ggtext') the output type used is
\code{"markdown"}. If \code{geom = "marquee"} (package 'marquee') the output
type used is \code{"marquee"}. For all other values of \code{geom} the default
is \code{"expression"} unless the user passes an argument. Invalid values as
argument trigger an Error.
}

\examples{
check_output_type(NULL)
check_output_type("text")
check_output_type(NULL, geom = "text")
check_output_type(NULL, geom = "latex")

}
