% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-poly-formula.R
\name{check_poly_formula}
\alias{check_poly_formula}
\title{Validate model formula as a polynomial}
\usage{
check_poly_formula(
  formula,
  x.name = "x",
  warn.incr.poly.text = "'formula' not an increasing polynomial: 'eq.label' set to NA!",
  warn.transf.rhs.txt =
    paste0("'formula' rhs includes transformations requiring an argument for ",
    "'eq.x.rhs': 'eq.label' set to NA!."),
  warn.transf.lhs.txt =
    paste0("'formula' lhs includes transformations requiring an argument for ",
    "'eq.with.lhs': 'eq.label' set to NA!."),
  warn.spline.rhs.txt = paste0("'formula' rhs includes a spline base function: ",
    " 'eq.label' set to NA!."),
  warn.as.is.txt = paste0("Power (^) terms in 'formula' for a polynomial need to ",
    "be protected by 'I()': 'eq.label' set to NA!."),
  warn.poly.raw.txt = paste0("'poly()' in 'formula' has to be passed 'raw = TRUE': ",
    "'eq.label' set to NA!"),
  stop.pow.poly.text =
    "Both 'poly()' and power (^) terms in 'formula', use one or the other.",
  check.transf.rhs = TRUE,
  check.transf.lhs = TRUE
)
}
\arguments{
\item{formula}{A model formula in \code{x.name}.}

\item{x.name}{character The name of the explanatory variable in the formula.}

\item{warn.incr.poly.text, warn.transf.lhs.txt, warn.transf.rhs.txt, warn.spline.rhs.txt, warn.as.is.txt, warn.poly.raw.txt, stop.pow.poly.text}{character Text for warnings and errors. If \code{NULL} the warning is not
issued and failure reported silently to the caller.}

\item{check.transf.rhs, check.transf.lhs}{logical flag enabling test for
transformation of variables.}
}
\value{
A logical, TRUE if the formula describes an increasing polynomial
  suitable for conversion into a text label, and FALSE otherwise. When
  validation fails, warnings are issued describing the problem encountered.
}
\description{
Analyse a model formula to determine if it describes a polynomial with
terms in order of increasing powers, and fulfils the expectations of the
algorithm used to generate the equation-label.
}
\details{
The assumption is that this function will be called from within a
  'ggplot2' compatible layer function, and that model formulas will always
  have a single explanatory variable, variables will be \code{x} and
  \code{y}. Its behaviour is undefined or erroneous in other cases. This
  validation check cannot be demonstrated to be always correct, as it is
  difficult to test, or even list all possible variations of supported vs.
  unsupported formulas.

  Many valid model formulas that can be successfully fitted, cannot
  be automatically converted into correct character labels by functions in
  'ggpmisc'. Thus, this function triggers a
  warning in case of a test failure, not an error, returning a logic value. If
  this value is \code{FALSE}, the statistics in 'ggpmisc' skip the generation
  of an equation label, setting it to \code{NA}, while returning other
  character labels and the numeric estimates
  of the fitted coefficients. Thus, the stats can be used also
  with models that are not polynomials or containing transformations, but in
  this case the model equation when needed needs to be assembled in user's
  code within a call to `aes()`.

  Model formulas with and without an intercept term are accepted as valid, as
  \code{+0}, \code{-1} and \code{+1} are accepted. If a single \code{as.is}
  power term is included or if arithmetic (\code{sqrt()}, \code{exp()},
  \code{log()}), or trigonometric functions (\code{cos()}, \code{sin()},
  \code{tan()}, etc.) are encountered a warning is issued about the need to
  pass a matching argument to parameter \code{eq.x.rhs} of the statistic.

  If two or more terms are \code{as.is} (\code{I( )} protected) powers
  (\code{^}), they are expected to be in increasing order with no missing
  intermediate power terms.

  Spline base functions \code{ns()}, \code{bs()} and \code{lspline()} are
  searched for and flagged, while \code{poly()} is accepted. If \code{poly()}
  is used in the model formula, a single term is expected. When calling
  function \code{poly()}, \code{raw = TRUE} must be passed to obtain suitable
  estimates for the fitted coefficients, and this is also checked.

  When the formula \emph{rhs} contains more than one power term, all power
  terms defined using \code{^} must be protected as \code{"as.is"}
  \code{I()}, as otherwise they are not powers but instead part of the
  formula specification.

  If the warning text is \code{NULL} or \code{character(0)} no warning is
  issued, but the test is done. In contrast, \code{check.transf.rhs,check =
  FALSE} and \code{transf.lhs = FALSE} skip these two tests on with
  \code{raw = TRUE}.
}
\examples{
# polynomials
check_poly_formula(y ~ 1)
check_poly_formula(y ~ x)
check_poly_formula(y ~ x^3)
check_poly_formula(y ~ x + 0)
check_poly_formula(y ~ x - 1)
check_poly_formula(y ~ x + 1)
check_poly_formula(y ~ x + I(x^2))
check_poly_formula(y ~ 1 + x + I(x^2))
check_poly_formula(y ~ x + I(x^2) + I(x^3))
check_poly_formula(y ~ I(x) + I(x^2) + I(x^3))

# transformations on x, first degree polynomials
check_poly_formula(y ~ sqrt(x))
check_poly_formula(y ~ log(x))
check_poly_formula(y ~ I(x^2))

# incomplete or terms in decreasing/mixed order
check_poly_formula(y ~ I(x^2) + x)
check_poly_formula(y ~ I(x^2) + I(x^3))
check_poly_formula(y ~ I(x^2) + I(x^4))
check_poly_formula(y ~ x + I(x^3) + I(x^2))

# polynomials using poly()
check_poly_formula(y ~ poly(x, 2, raw = TRUE)) # label o.k.
check_poly_formula(y ~ poly(x, 2)) # orthogonal polynomial -> bad label

# spline terms not accepted
check_poly_formula(y ~ ns(x))
check_poly_formula(y ~ bs(x))
check_poly_formula(y ~ lspline(x))

}
