% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-eq-label.R
\name{coefs2poly_eq}
\alias{coefs2poly_eq}
\title{Format a polynomial as an equation}
\usage{
coefs2poly_eq(
  coefs,
  coef.digits = 3L,
  coef.keep.zeros = TRUE,
  eq.x.rhs = "x",
  lhs = "y~`=`~",
  output.type = "expression",
  decimal.mark = "."
)
}
\arguments{
\item{coefs}{numeric}

\item{coef.digits}{integer}

\item{coef.keep.zeros}{logical}

\item{eq.x.rhs}{character}

\item{lhs}{character}

\item{output.type}{character One of "expression", "latex", "tex", "text",
"tikz", "markdown".}

\item{decimal.mark}{character}
}
\value{
A \code{character} string.
}
\description{
Uses a vector of coefficients from a model fit of a polynomial to build
the fitted model equation with embedded coefficient estimates.
}
\examples{
coefs2poly_eq(1:5)

}
