% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-label.R
\name{decode_label_names}
\alias{decode_label_names}
\title{Decode short names and variant names of labels}
\usage{
decode_label_names(label.names = NULL)
}
\arguments{
\item{label.names}{list or vector of character strings with mixed names.}
}
\description{
Decode short names and variant names of labels
}
\keyword{internal}
