% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-continuous-omics.r
\name{FC_format}
\alias{FC_format}
\alias{FC_plain}
\title{Formatter for fold change tick labels}
\usage{
FC_format(log.base.labels = 2, log.base.data = 2, digits = 3, ...)

FC_plain(x, log.base.data = 2, log.base.labels = 2, digits = 3, ...)
}
\arguments{
\item{log.base.labels}{of labels.}

\item{log.base.data}{of data.}

\item{digits}{number of significant digits to show.}

\item{...}{other arguments passed on to \code{\link{format}}.}

\item{x}{a numeric vector to format.}
}
\value{
a function with single parameter x, a numeric vector, that
  returns a character vector.
}
\description{
Generate tick labels discounting a log transformation in data, and
optionally applying a different log transformation or no transformation.
}
\examples{

FC_format(2, 10)(1:10)
FC_format(2, FALSE)(c(1/4, 1/2, 1,2,4,8))
FC_format(10, FALSE)(c(1, 10, 100, 1000))
FC_format(FALSE, 10)(c(-1, 1, 2, 3))

}
\keyword{internal}
