% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmodel2-util.R
\name{fit_lmodel2_internal}
\alias{fit_lmodel2_internal}
\title{Fit an lmodel2 model}
\usage{
fit_lmodel2_internal(
  data,
  method,
  method.name,
  method.args,
  n.min,
  formula,
  range.y,
  range.x,
  fit.seed,
  orientation,
  nperm
)
}
\arguments{
\item{data}{data.frame containing the variables in the model.}

\item{method, method.name}{function and character, respectively.}

\item{method.args}{named list with additional arguments. Not \code{data}
or \code{weights} which are always passed through aesthetic mappings.}

\item{n.min}{integer Minimum number of distinct values in the explanatory
variable (on the rhs of formula) for fitting to the attempted.}

\item{formula}{a formula object. Using aesthetic names \code{x} and \code{y}
instead of original variable names.}

\item{range.y, range.x}{character Pass "relative" or "interval" if method
"RMA" is to be computed.}

\item{fit.seed}{RNG seed argument passed to
\code{\link[base:Random]{set.seed}()}. Defaults to \code{NA}, indicating
that \code{set.seed()} should not be called.}

\item{orientation}{character Either "x" or "y" controlling the default for
\code{formula}. The letter indicates the aesthetic considered the
explanatory variable in the model fit.}

\item{nperm}{integer Number of permutation used to estimate significance.}
}
\value{
A list with three named members: \code{fm} the fitted model object
  and \code{method.args}, the arguments passed to the model fit function as a
  nested named list, \code{fit.seed} the seed used and \code{method.name},
  the name of the model fit function passed as arguemnt, which can differ
  from the class of \code{fm}.
}
\description{
Fit models using 'lmoldel2' translating some arguments
to make possible use of consistent arguments across calls to
stats.
}
\note{
Called by \code{\link{stat_ma_line}()},
  and \code{\link{stat_ma_eq}()}.
}
\keyword{internal}
