% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-fit-residuals.R
\name{fit_models_internal}
\alias{fit_models_internal}
\title{Fit a model to extract residuals or fitted values}
\usage{
fit_models_internal(
  data,
  method,
  method.name,
  method.args,
  n.min,
  formula,
  fit.seed,
  orientation
)
}
\description{
Fit a model to extract residuals or fitted values
}
\keyword{internal}
