% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-linked-text.r
\name{geom_linked_text}
\alias{geom_linked_text}
\alias{geom_linked_label}
\title{Linked Text}
\usage{
geom_linked_text(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  min.segment.length = 0.5,
  arrow = NULL,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_linked_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  min.segment.length = 0.5,
  arrow = NULL,
  label.padding = grid::unit(0.25, "lines"),
  label.r = grid::unit(0.15, "lines"),
  label.size = 0.25,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by
\code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. If specified and
\code{inherit.aes = TRUE} (the default), is combined with the default
mapping at the top level of the plot. You only need to supply
\code{mapping} if there isn't a mapping defined for the plot.}

\item{data}{A data frame. If specified, overrides the default data frame
defined at the top level of the plot.}

\item{stat}{The statistical transformation to use on the data for this layer,
as a string.}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. There
  are three types of arguments you can use here:

  \itemize{ \item Aesthetics: to set an aesthetic to a fixed value, like
  \code{colour = "red"} or \code{size = 3}. \item Other arguments to the
  layer, for example you override the default \code{stat} associated with the
  layer. \item Other arguments passed on to the stat. }}

\item{parse}{If \code{TRUE}, the labels will be parsed into expressions and
displayed as described in \code{?plotmath}.}

\item{nudge_x, nudge_y}{Horizontal and vertical adjustments to nudge the
starting position of each text label. The units for \code{nudge_x} and
\code{nudge_y} are the same as for the data units on the x-axis and y-axis.}

\item{min.segment.length}{Skip drawing segments shorter than this, as unit or
number. Defaults to 0.5. (Default unit is lines, but other units can be
specified by passing \code{unit(x, "units")}).}

\item{arrow}{specification for arrow heads, as created by
\code{\link[grid]{arrow}}}

\item{check_overlap}{If \code{TRUE}, text that overlaps previous text in the
same layer will not be plotted. \code{check_overlap} takes place at draw
time and in the order of the data, thus its action depends of the size at
which the plot is drawn.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with a
warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}

\item{label.padding}{Amount of padding around label. Defaults to 0.25 lines.}

\item{label.r}{Radius of rounded corners. Defaults to 0.15 lines.}

\item{label.size}{Size of label border, in mm.}
}
\description{
Text geoms are useful for labeling plots. `geom_linked_text()`
  adds text to the plot and for nudged positions links the original location
  to the nudeged text with a segment. `geom_linked_label()` draws a rectangle
  behind the text, making it easier to read on a crowded background.
}
\details{
Note that when you resize a plot, text labels stay the same size,
  even though the size of the plot area changes. This happens because the
  "width" and "height" of a text element are 0. Obviously, text labels do
  have height and width, but they are physical units, not data units. For the
  same reason, stacking and dodging text will not work by default, and axis
  limits are not automatically expanded to include all text.

  `geom_linked_text()` and `geom_linked_label()` add labels for each row in
  the data, even if coordinates x, y are set to single values in the call to
  `geom_linked_label()` or `geom_linked_text()`.

  By default these geoms use `position_nudge_center()` which is backwards
  compatible with `position_nudge()` from 'ggplot2' but provides additional
  control on the direction of the nudging. In contrast to `position_nudge()`,
  `position_nudge_center()` and `position_nudge_line()` preserve the original
  coordinates.
}
\section{`geom_linked_label()`}{
 Currently `geom_linked_label()` does not support
  the `angle` aesthetic and is considerably slower than `geom_linked_text()`. The
  `fill` aesthetic controls the background colour of the label.
}

\section{Alignment}{
 You can modify text alignment with the `vjust` and
  `hjust` aesthetics. These can either be a number between 0 (right/bottom)
  and 1 (top/left) or a character (`"left"`, `"middle"`, `"right"`,
  `"bottom"`, `"center"`, `"top"`). There are two special alignments:
  `"inward"` and `"outward"`. Inward always aligns text towards the center,
  and outward aligns it away from the center.
}

