% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-debug.R, R/stat-debug-group.R, R/stat-debug-panel.R, R/stat-dens2d-filter.r, R/stat-dens2d-labels.r, R/stat-fit-broom.R, R/stat-fit-deviations.R, R/stat-fit-residuals.R, R/stat-peaks.R, R/stat-poly-eq.R
\docType{data}
\name{GeomNull}
\alias{GeomDebug}
\alias{GeomNull}
\alias{StatDebugGroup}
\alias{StatDebugPanel}
\alias{StatDens2dFilter}
\alias{StatDens2dLabels}
\alias{StatFitAugment}
\alias{StatFitDeviations}
\alias{StatFitGlance}
\alias{StatFitResiduals}
\alias{StatPeaks}
\alias{StatPolyEq}
\alias{StatValleys}
\alias{Stats}
\alias{fit_augment_compute_group_fun}
\alias{fit_glance_compute_group_fun}
\alias{peaks_compute_group_fun}
\alias{poly_eq_compute_group_fun}
\alias{residuals_compute_group_fun}
\alias{valleys_compute_group_fun}
\title{\code{Stat*} Objects}
\description{
All \code{stat_*} functions (like \code{stat_bin}) return a layer that
contains a \code{Stat*} object (like \code{StatBin}). The \code{Stat*}
object is responsible for rendering the data in the plot.
}
\details{
Each of the \code{Stat*} objects is a \code{\link[ggplot2]{ggproto}} object, descended
from the top-level \code{Stat}, and each implements various methods and
fields. To create a new type of Stat object, you typically will want to
implement one or more of the following:
}
\seealso{
\code{\link[ggplot2]{ggplot2-ggproto}}
}
\keyword{datasets}
\keyword{internal}

