% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-fit-broom-tb.R, R/stat-fit-broom.R,
%   R/stat-fit-deviations.R, R/stat-fit-residuals.R, R/stat-peaks.R,
%   R/stat-poly-eq.R, R/stat-quant-eq.R
\docType{data}
\name{fit_tb_compute_panel_fun}
\alias{fit_tb_compute_panel_fun}
\alias{StatFitTb}
\alias{fit_glance_compute_group_fun}
\alias{StatFitGlance}
\alias{fit_augment_compute_group_fun}
\alias{StatFitAugment}
\alias{fit_tidy_compute_group_fun}
\alias{StatFitTidy}
\alias{StatFitDeviations}
\alias{residuals_compute_group_fun}
\alias{StatFitResiduals}
\alias{peaks_compute_group_fun}
\alias{valleys_compute_group_fun}
\alias{Stats}
\alias{StatPeaks}
\alias{StatValleys}
\alias{poly_eq_compute_group_fun}
\alias{StatPolyEq}
\alias{quant_eq_compute_group_fun}
\alias{StatQuantEq}
\title{\code{Stat*} Objects}
\description{
All \code{stat_*} functions (like \code{stat_bin}) return a layer that
contains a \code{Stat*} object (like \code{StatBin}). The \code{Stat*}
object is responsible for rendering the data in the plot.
}
\details{
Each of the \code{Stat*} objects is a \code{\link[ggplot2]{ggproto}} object,
derived from the top-level \code{Stat}, and each implements various methods
and fields. To create a new type of Stat object, you typically will want to
implement one or more of the following:
}
\seealso{
\code{\link[ggplot2]{ggplot2-ggproto}}
}
\keyword{datasets}
\keyword{internal}
