% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-grob.R, R/geom-label-npc.r,
%   R/geom-margin-arrow.r, R/geom-margin-grob.r, R/geom-margin-point.r,
%   R/geom-plot.R, R/geom-quadrant-lines.R, R/geom-table.R, R/geom-text-npc.r,
%   R/stat-apply.R, R/stat-dens2d-filter.r, R/stat-dens2d-labels.r,
%   R/stat-fit-broom-tb.R, R/stat-fit-broom.R, R/stat-fit-deviations.R,
%   R/stat-fit-residuals.R, R/stat-fmt-tb.R, R/stat-peaks.R, R/stat-poly-eq.R,
%   R/stat-quadrant-counts.R
\docType{data}
\name{grob_draw_panel_fun}
\alias{grob_draw_panel_fun}
\alias{GeomGrob}
\alias{grobnpc_draw_panel_fun}
\alias{GeomGrobNpc}
\alias{GeomLabelNpc}
\alias{GeomXMarginArrow}
\alias{GeomYMarginArrow}
\alias{GeomXMarginGrob}
\alias{GeomYMarginGrob}
\alias{GeomXMarginPoint}
\alias{GeomYMarginPoint}
\alias{gplot_draw_panel_fun}
\alias{GeomPlot}
\alias{gplotnpc_draw_panel_fun}
\alias{GeomPlotNpc}
\alias{GeomQuadrantLines}
\alias{GeomVHLines}
\alias{gtb_draw_panel_fun}
\alias{GeomTable}
\alias{gtbnpc_draw_panel_fun}
\alias{GeomTableNpc}
\alias{GeomTextNpc}
\alias{stat_apply_fun}
\alias{Stats}
\alias{StatApplyGroup}
\alias{StatApplyPanel}
\alias{StatDens2dFilter}
\alias{StatDens2dFilterG}
\alias{StatDens2dLabels}
\alias{fit_tb_compute_panel_fun}
\alias{StatFitTb}
\alias{fit_glance_compute_group_fun}
\alias{StatFitGlance}
\alias{fit_augment_compute_group_fun}
\alias{StatFitAugment}
\alias{fit_tidy_compute_group_fun}
\alias{StatFitTidy}
\alias{StatFitDeviations}
\alias{residuals_compute_group_fun}
\alias{StatFitResiduals}
\alias{fmt_tb_compute_group_fun}
\alias{StatFmtTb}
\alias{peaks_compute_group_fun}
\alias{valleys_compute_group_fun}
\alias{StatPeaks}
\alias{StatValleys}
\alias{poly_eq_compute_group_fun}
\alias{StatPolyEq}
\alias{compute_counts_fun}
\alias{StatQuadrantCounts}
\title{\code{Stat*} Objects}
\description{
All \code{stat_*} functions (like \code{stat_bin}) return a layer that
contains a \code{Stat*} object (like \code{StatBin}). The \code{Stat*}
object is responsible for rendering the data in the plot.

All \code{stat_*} functions (like \code{stat_bin}) return a layer that
contains a \code{Stat*} object (like \code{StatBin}). The \code{Stat*}
object is responsible for rendering the data in the plot.

All \code{stat_*} functions (like \code{stat_bin}) return a layer that
contains a \code{Stat*} object (like \code{StatBin}). The \code{Stat*}
object is responsible for rendering the data in the plot.
}
\details{
Each of the \code{Stat*} objects is a \code{\link[ggplot2]{ggproto}} object, descended
from the top-level \code{Stat}, and each implements various methods and
fields. To create a new type of Stat object, you typically will want to
implement one or more of the following:

Each of the \code{Stat*} objects is a \code{\link[ggplot2]{ggproto}} object, descended
from the top-level \code{Stat}, and each implements various methods and
fields. To create a new type of Stat object, you typically will want to
implement one or more of the following:

Each of the \code{Stat*} objects is a \code{\link[ggplot2]{ggproto}} object,
derived from the top-level \code{Stat}, and each implements various methods
and fields. To create a new type of Stat object, you typically will want to
implement one or more of the following:
}
\seealso{
\code{\link[ggplot2]{ggplot2-ggproto}}

\code{\link[ggplot2]{ggplot2-ggproto}}

\code{\link[ggplot2]{ggplot2-ggproto}}
}
\keyword{datasets}
\keyword{internal}
