% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpmisc.R
\docType{package}
\name{ggpmisc-package}
\alias{ggpmisc}
\alias{ggpmisc-package}
\title{Miscelaneous Extensions to ggplot2}
\description{
This package provides additional stats for ggplot.
}
\details{
Pakage \code{ggpmisc} is a package with extensions to ggplot2.
It defines statistics for finding peaks and valleys, constructing labels
for polynomial equations fitted with \code{lm()}, and for debugging
or discovering the data that \code{compute_group} and \code{compute_panel}
are passed within ggplot statistics. It also defines function
\code{try_data_frame()} which attempts to convert time series objects
supported by \code{\link[xts]{try.xts}()} into data frame objects
usable as \code{data} argument to \code{ggplot()}.
}
\note{
The signatures of \code{stat_peaks()} and \code{stat_valleys()} are
identical to those of \code{\link[ggspectra]{stat_peaks}} and
\code{\link[ggspectra]{stat_valleys}} from package
\code{\link[photobiology]{photobiology}}
but the variables returned are a subset as values related to light spectra
are missing. Furthermore the stats from package \code{ggpmisc} work
correctly when the x aesthetic uses a date or datetime scale, while those
from package \code{photobiology} do not generate correct labels in
this case.
}
\author{
Pedro J. Aphalo
}
\references{
\code{ggplot2} web site at \url{http://ggplot2.org/}\cr
\code{ggplot2} source code at \url{https://github.com/hadley/ggplot2}
}

