% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpmisc-package.r
\docType{package}
\name{ggpmisc-package}
\alias{ggpmisc}
\alias{ggpmisc-package}
\title{Miscelaneous extensions to ggplot2}
\description{
Pakage \code{ggpmisc} is a package with extensions to ggplot2.
It defines statistics for finding peaks and valleys, constructing labels
for polynomial equations fitted with \code{lm()}, and for debugging
or discovering the data that \code{compute_group} and \code{compute_panel}
are passed within ggplot statistics.
}
\details{
\tabular{ll}{
Package: \tab ggpmisc\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 2016-01-23\cr
License: \tab GPL (>= 3.0)\cr
URL: \tab \url{http://www.r4photobiology.info},\cr
}
}
\note{
The signatures of \code{stat_peaks()} and \code{stat_valleys()} are
identical to those of \code{\link[ggspectra]{stat_peaks}} and
\code{\link[ggspectra]{stat_valleys}} from package
\code{\link[photobiology]{photobiology}}
but the variables returned are a subset as values related to light spectra
are missing.
}
\author{
Pedro J. Aphalo
}
\references{
\code{ggplot2} web site at \url{http://ggplot2.org/}\cr
\code{ggplot2} source code at \url{https://github.com/hadley/ggplot2}
}
\keyword{misc}

