% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-fm.R
\name{guess_orientation}
\alias{guess_orientation}
\title{Guess the orientation from model formula}
\usage{
guess_orientation(
  orientation = NULL,
  formula = NULL,
  default.formula = y ~ x,
  formula.on.x = FALSE
)
}
\arguments{
\item{orientation}{character \code{"x"} or \code{"y"}.}

\item{formula}{model formula based on x and y.}

\item{formula.on.x}{logical Flip x and y in formula, used when the x
and y in data are not flipped in the compute function.}
}
\value{
Named list with members formula and orientation.
}
\description{
Guess the orientation from model formula
}
\details{
Set defaults for both \code{orientation} and \code{formula},
  ensuring consistency both when x and y and swapped in data and when they
  are not.
}
\keyword{internal}
