% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-colour-outcome.r
\name{outcome2factor}
\alias{outcome2factor}
\title{Convert numeric ternary outcomes into a factor}
\usage{
outcome2factor(x, n.levels = 3L)
}
\arguments{
\item{x}{a numeric vector of -1, 0, and +1 values, indicating down
regulation, uncertain response or upregulation.}

\item{n.levels}{numeric Number of levels to create.}
}
\description{
Convert numeric ternary outcomes into a factor
}
\details{
This function converts the numerically encoded values into a factor
  with the three levels \code{"down"}, \code{"uncertain"} and \code{"up"}, or
  into a factor with two levels \code{de} and \code{uncertain} as expected
  by scales \code{scale_colour_outcome()}, \code{scale_fill_outcome()} and
  \code{scale_shape_outcome()}.
}
\note{
When \code{n.levels = 2} both -1 and +1 are merged to the same level
  with label \code{"de"}.
}
\examples{

outcome2factor(c(-1, 1, 0, 1))

}
\seealso{
Other scales for omics data: \code{\link{scale_colour_outcome}},
  \code{\link{scale_shape_outcome}},
  \code{\link{scale_x_logFC}}, \code{\link{scale_y_Pvalue}}
}
\concept{scales for omics data}
