% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-continuous-Pvalue.r
\name{reverselog_trans}
\alias{reverselog_trans}
\title{Reverse log transformation}
\usage{
reverselog_trans(base = exp(1))
}
\arguments{
\item{base}{numeric Base of logarithm}
}
\description{
Reverse log transformation
}
\keyword{P-value}
\keyword{define}
\keyword{for}
\keyword{internal}
\keyword{labels}
\keyword{needed}
\keyword{tick}
\keyword{transformation}
