% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{show_colnames}
\alias{show_colnames}
\title{Message listing non-missing columns}
\usage{
show_colnames(z, show.what = NULL, stat.name = "stat")
}
\arguments{
\item{z}{data.frame from which to extract names of columns.}

\item{show.what}{character Replace column names by short nicknames in the
message text.}

\item{stat.name}{character Name to be included in the message text.}
}
\description{
Issue a message listing column names that are available for mapping to
aesthetics.
}
\details{
Columns of \code{z} with names ending in \code{.label} and
  containing non-missing values are reported after replacing column names by
  the shortnames used in function \code{\link{use_label}()} and
  \code{f_use_label()}.

  Computed variables and their names can vary depending on the
  \code{method} used to fit a model or the \code{output.type} in use. They
  can also depend for a given \code{method} on other arguments passed when
  fitting a model or extracting estimates and other computed values. A
  message is issued listing the short names for formatted labels as
  recognized by functions \code{\link{use_label}()} and
  \code{\link{use_label}()}.
}
\keyword{internal}
