% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{show_colnames}
\alias{show_colnames}
\title{Message listing non-missing columns}
\usage{
show_colnames(z, show.what = "colnames")
}
\arguments{
\item{z}{data.frame from which to extract names of columns.}

\item{show.what}{character Replace column names by short nicknames in the
message text.}
}
\description{
Issue a message listing column names that are available for mapping to
aesthetics.
}
\details{
Columns of \code{z} with names ending in \code{.label} and
  containing non-missing values are reported after replacing column names by
  the shortnames used in function \code{\link{use_label}()} and
  \code{f_use_label()}.
}
\keyword{internal}
