% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{show_labels}
\alias{show_labels}
\title{Message listing non-missing labels}
\usage{
show_labels(z, show.nicknames = TRUE)
}
\arguments{
\item{z}{data.frame from which to extract names of columns.}

\item{show.nicknames}{logical Replace column names by short nicknames in the
message text.}
}
\description{
Issue a message listing labels that are available using their "nicknames".
}
\details{
Columns of \code{z} with names ending in \code{.label} and
  containing non-missing values are reported after replacing column names by
  the shortnames used in function \code{\link{use_label}()} and
  \code{f_use_label()}.
}
\keyword{internal}
