% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-distrmix-eq.R, R/stat-distrmix-line.R
\name{stat_distrmix_eq}
\alias{stat_distrmix_eq}
\alias{stat_distrmix_line}
\alias{stat_distrmix_area}
\title{Mixture model prediction and annotations}
\usage{
stat_distrmix_eq(
  mapping = NULL,
  data = NULL,
  geom = "text_npc",
  position = "identity",
  ...,
  orientation = NA,
  method = "normalmixEM",
  method.args = list(),
  n.min = 10L * k,
  level = 0.95,
  k = 2,
  free.mean = TRUE,
  free.sd = TRUE,
  se = FALSE,
  fit.seed = NA,
  fm.values = TRUE,
  components = NULL,
  eq.with.lhs = TRUE,
  eq.digits = 2,
  label.x = "left",
  label.y = "top",
  hstep = 0,
  vstep = NULL,
  output.type = NULL,
  na.rm = FALSE,
  parse = NULL,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_distrmix_line(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  ...,
  orientation = NA,
  method = "normalmixEM",
  se = NULL,
  quantiles = NA,
  fit.seed = NA,
  fm.values = FALSE,
  n = min(100 + 50 * k, 300),
  fullrange = TRUE,
  level = 0.95,
  method.args = list(),
  k = 2,
  free.mean = TRUE,
  free.sd = TRUE,
  components = "all",
  n.min = 10L * k,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_distrmix_area(
  mapping = NULL,
  data = NULL,
  geom = "area",
  position = "identity",
  ...,
  orientation = NA,
  method = "normalmixEM",
  se = NULL,
  quantiles = NA,
  fit.seed = NA,
  fm.values = FALSE,
  n = min(100 + 50 * k, 300),
  fullrange = TRUE,
  level = 0.95,
  method.args = list(),
  k = 2,
  free.mean = TRUE,
  free.sd = TRUE,
  components = "sum",
  n.min = 10L * k,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}()}. Only needs to be set at the layer level if you
are overriding the plot defaults.}

\item{data}{A layer specific dataset, only needed if you want to override the
plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points on this
layer.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}

\item{orientation}{character Either "x" or "y" controlling the aesthetic to
which the density model is fit. With the default \code{orientation = NA}
the orientation used is based on the mapping and nearly always correct.}

\item{method}{function or character If character, \code{"normalmixEM"} or the
name of a model fit function are accepted, possibly followed by the fit
function's method argument separated by a colon. The function must return a
model fit object of class \code{"mixEM"}.}

\item{method.args}{named list with additional arguments. Not \code{data}
or \code{weights} which are always passed through aesthetic mappings.}

\item{n.min}{integer Minimum number of distinct values in the variable for
fitting to the attempted. The default depends on \code{k}.}

\item{level}{Level of confidence interval to use (0.95 by default).}

\item{k}{integer Number of mixture components to fit.}

\item{free.mean, free.sd}{logical If TRUE, allow the fitted \code{mean} and/or
fitted \code{sd} to vary among the component Normal distributions.}

\item{se}{logical If \code{TRUE} standard errors for the parameter estimates
are returned in addition to the parameter estimates.}

\item{fit.seed}{RNG seed argument passed to
\code{\link[base:Random]{set.seed}()}. Defaults to \code{NA}, indicating
that \code{set.seed()} should not be called.}

\item{fm.values}{logical Add parameter estimates and their standard errors
to the returned values (`FALSE` by default.)}

\item{components}{character One of \code{"all"}, \code{"sum"}, or
\code{"members"} select which densities are returned.}

\item{eq.with.lhs}{If \code{character} the string is pasted to the front of
the equation label before parsing or a \code{logical} (see note).}

\item{eq.digits}{integer Number of digits after the decimal point to
use for parameters in labels. If \code{Inf}, use exponential
notation with three decimal places.}

\item{label.x, label.y}{\code{numeric} with range 0..1 "normalized parent
coordinates" (npc units) or character if using \code{geom_text_npc()} or
\code{geom_label_npc()}. If using \code{geom_text()} or \code{geom_label()}
numeric in native data units. If too short they will be recycled.}

\item{hstep, vstep}{numeric in npc units, the horizontal and vertical step
used between labels for different groups.}

\item{output.type}{character One of "expression", "text", "markdown",
"marquee", "latex", "latex.eqn", "latex.deqn" or "numeric".}

\item{na.rm}{a logical indicating whether NA values should be stripped before
the computation proceeds.}

\item{parse}{logical Passed to the geom. If \code{TRUE}, the labels will be
parsed into expressions and displayed as described in
\code{\link[grDevices]{plotmath}}. Default is \code{TRUE} if
\code{output.type = "expression"} and \code{FALSE} otherwise.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}

\item{quantiles}{numeric The boundaries for tails given by two
\emph{P}-values in the range 0 to 1, with \code{NA} being equivalent to
\code{quantiles = c(0, 1)}. If a longer vector is passed as argument, its
range is used.}

\item{n}{Number of points at which to predict with the fitted model.}

\item{fullrange}{logical Should the fit prediction span the full
range of the plot, or just the range of the explanatory variable?}
}
\value{
The value returned by \code{stat_distrmix_line()} is a data frame,
  with \code{n} rows of predictions or possibly fewer in the case when
  \code{fullrange = FALSE} is passed in the call. Predictions for each
  component of the mixture plus for their sum are returned in long form.

  The value returned by \code{stat_distrmix_eq()} is a data frame, with one
  row of estimates for each group of data in the plot.

  Both statistics optionally also return additional values related to the
  model fit.
}
\description{
Statistics \code{stat_distrmix_line()} and \code{stat_distrmix_eq()} fit a
Normal mixture model. \code{stat_distrmix_line()} adds a layer with
prediction lines or areas, optionally with tail quantiles highlighted, and
\code{stat_distrmix_eq()} adds a layer with textual labels of parameter
estimates to a plot.
}
\details{
\code{stat_distrmix_line()} and \code{stat_distrmix_area()} are
  similar to \code{\link[ggplot2]{stat_density}} but they fit the Normal
  distribution to observations. In addition to a single Normal distribution
  they can fit a mixture of two or more Normal distributions, using an
  approach related to clustering. Defaults related to how fitting is done are
  consistent between \code{stat_distrmix_line()}, \code{stat_distrmix_area()}
  and \code{stat_distrmix_eq()}. \code{stat_distrmix_eq()} can be used to add
  matched textual annotations, while \code{stat_distrmix_line()} and
  \code{stat_distrmix_area()} only differ in their default arguments,
  including the \code{geom}.

  If \code{k >= 2} a mixture-of-Normals model is fitted with
  \code{\link[mixtools]{normalmixEM}()}, while if \code{k == 1} a single
  Normal distribution is fitted with function \code{\link[MASS]{fitdistr}()}.
  Only for \code{k == 1} the SE values are exact estimates.

  In \code{stat_distrmix_line()}, predictions are computed to cover >= 0.999 of
  the integral in all cases, trimming to the range of the data with
  \code{fullrange = FALSE} is done as the last step in the computations. This
  ensures correct estimates of the cumulated density (CDF) and of tail
  quantiles, whose locations are estimated based on the CDF.

  Parameter \code{fit.seed} if not \code{NA} is used in a call to
  \code{set.seed()} immediately before calling the model fit function. As the
  fitting procedure makes use of the (pseudo-)random number generator (RNG),
  convergence can depend on it, and in such cases setting \code{fit.seed} to
  the same value in \code{stat_distrmix_line()} and in
  \code{stat_distrmix_eq()} can ensure consistency, and more generally,
  reproducibility.

  The minimum number of observations with distinct values in the data
  variable can be set through parameter \code{n.min}. The default depends on
  \code{k}, the number of components in the mix. Model fits with too few
  observations are unreliable, thus, using larger values of \code{n.min} than
  the default is wise. The value of \code{n}, instead, sets the number of
  predicted values, which affects the smoothness of the plotted curve and
  the accuracy with which the location of the quantiles is predicted.
}
\section{Variables computed by \code{stat_distrmix_line()}}{


  Some of the variables can have missing values or depend on
  \code{orientation} and/or \code{method}. A message is issued listing
  the column names containing non-missing values.

  \describe{\item{x}{\code{n} or fewer values}
  \item{component}{A factor indexing the components and/or their sum}
  \item{density}{predicted density values}
  \item{cum.density}{predicted cumulated density values}
  \item{is.tail}{logical, \code{TRUE} if in a tail exceeding the values of the \code{quantiles}}}

  If \code{fm.values = TRUE} is passed then columns with diagnosis and
  parameters estimates are added, with the same value in each row within a
  group:
  \describe{\item{converged}{\code{logical} indicating if convergence was achieved}
  \item{n}{\code{numeric} the number of \code{x} values}
  \item{.size}{\code{numeric} the number of \code{density} values}
  \item{fm.class}{\code{character} the most derived class of the fitted model object}
  \item{fm.method}{\code{character} the method, as given by the \code{ft}
  field of the fitted model objects}}

  This provides a simple and robust
  approach to achieve effects like colouring or hiding annotations
  by group depending on the outcome of model fitting.
}

\section{Variables computed by \code{stat_distrmix_eq()}}{


  Computed variables and their names can vary depending on the \code{method}
  used to fit a model or the \code{output.type} in use. They can also depend
  for a given \code{method} on other arguments passed when fitting a model or
  extracting estimates and other computed values. A message is issued listing
  the short names for formatted labels as recognized by functions
  \code{\link{use_label}()} and \code{\link{f_use_label}()}, except when
  \code{output.type = "numeric"} is passed, in which case the
  names of all variables accessible by \code{after_stat()} within a call to
  \code{aes()} are listed.

  Some of the variables depend on the orientation:
  \describe{\item{x}{the location of text labels}
  \item{y}{the location of text labels}
  \item{eq.label}{\code{character} string for equations}
  \item{n.label}{\code{character} string for number of observations}
  \item{method.label}{\code{character} string for model fit method}
  \item{lambda}{\code{numeric} the estimate of the contribution of the
 component of the mixture towards the joint density}
  \item{mu}{\code{numeric} the estimate of the mean}
  \item{sigma}{\code{numeric} the estimate of the standard deviation}
  \item{component}{A factor indexing the components of the mixture and/or
  their sum}}

  If \code{SE = TRUE} is passed then columns with standard errors for the
  parameter estimates:
  \describe{\item{lambda.se}{\code{numeric} the estimate of the contribution
  of the component of the mixture towards the joint density}
  \item{mu.se}{\code{numeric} the estimate of the mean}
  \item{sigma.se}{\code{numeric} the estimate of the standard deviation}}

  If \code{fm.values = TRUE} the same additional columns are returned as by
  \code{stat_distrmix_eq()}. This is wasteful of storage space as values are
  stored in multiple copies and, thus, disabled by default. However, it
  provides a simple and robust approach to achieve effects like colouring or
  hiding of the model fit line by group depending on the outcome of model
  fitting.
}

\section{Output types}{
 The formatting of character strings to be displayed
in plots are marked as mathematical equations. Depending on the geom used,
the mark-up needs to be encoded differently, or in some cases mark-up not
applied.
\describe{
  \item{\code{"expression"}}{The labels are encoded as character strings to be parsed into R's plotmath expressions.}
  \item{\code{"LaTeX", "TeX", "tikz", "latex"}}{The labels are encoded as 'LaTeX' maths equations, without the "fences" for switching in math mode.}
  \item{\code{"latex.eqn"}}{Same as \code{"latex"} but enclosed in single \code{$}, i.e., as in-line maths.}
  \item{\code{"latex.deqn"}}{Same as \code{"latex"} but enclosed in double \code{$$}, i.e., as display maths.}
  \item{\code{"markdown"}}{The labels are encoded as character strings using markdown syntax, with some embedded HTML.}
  \item{\code{"marquee"}}{The labels are encoded as character strings using markdown syntax, with 'marquee' supported spans.}
  \item{\code{"text"}}{The labels are plain ASCII character strings.}
  \item{\code{"numeric"}}{No labels are generated. This value is accepted by the statistics, but not by the label formatting functions.}
  \item{\code{NULL}}{The value used depends on the argument passed to \code{geom}.}}

If \code{geom = "latex"} (package 'xdvir') the output type used is
\code{"latex.eqn"}. If \code{geom = "richtext"} (package 'ggtext') or
\code{geom = "textbox"} (package 'ggtext') the output type used is
\code{"markdown"}. If \code{geom = "marquee"} (package 'marquee') the output
type used is \code{"marquee"}. For all other values of \code{geom} the default
is \code{"expression"}. Invalid values as argument trigger an error.
}

\section{Position of labels}{

  When data are grouped by mapping a factor to an aesthetic, e.g.,
  \code{colour}, \code{shape} and/or \code{linetype} the model is fitted
  separately to each group, and for each group a whole set of labels is
  generated. If the argument passed to \code{label.y} is a vector of length
  1, this value determines the position of the equation and/or other labels
  for the first group, and the positions of the labels for the remaining
  groups are generated by adding \code{vspace} based on the group number.
  If the argument passed to \code{label.y} is a vector of length > 1, it is
  used unchanged, possibly extended by recycling, ignoring \code{vstep}.

  If the labels are rotated by 90 degrees then the automatic stepping is
  best based on \code{hstep} with \code{vstep = 0}. Similarly as described
  above, if \code{label.x} is a vector of length > 1, it is
  used unchanged, possibly extended by recycling, ignoring \code{hstep}.

  When using facets and with a grouping that does not repeat in each panel,
  the automatic positioning in most cases will not be the desired one. Manual
  positioning using a vector of length > 1 for \code{label.x} and/or
  \code{label.y} is the currently available workaround.
}

\examples{
ggplot(faithful, aes(x = waiting)) +
  stat_distrmix_line() +
  stat_distrmix_eq()

ggplot(faithful, aes(x = waiting)) +
  stat_distrmix_area() +
  stat_distrmix_eq()

ggplot(faithful, aes(x = waiting)) +
  stat_distrmix_area(aes(fill = after_stat(is.tail)),
                     quantiles = c(0.025, 0.975),
                     show.legend = FALSE)

ggplot(faithful, aes(x = waiting)) +
  stat_distrmix_line(components = "sum") +
  stat_distrmix_eq(use_label("eq", "n", "method"))

ggplot(faithful, aes(x = waiting)) +
  stat_distrmix_line(components = "sum") +
  stat_distrmix_eq(geom = "label_npc")

ggplot(faithful, aes(x = waiting)) +
  stat_distrmix_line(components = "sum") +
  stat_distrmix_eq(geom = "text", label.x = "center", label.y = "bottom")

ggplot(faithful, aes(x = waiting)) +
  stat_distrmix_line(components = "sum") +
  stat_distrmix_eq(geom = "text", hjust = "inward")

ggplot(faithful, aes(x = waiting)) +
  stat_distrmix_line(components = "members") +
  stat_distrmix_eq(components = "members")

ggplot(faithful, aes(x = waiting)) +
  stat_distrmix_line(components = "members") +
  stat_distrmix_eq(components = "members", se = TRUE)

ggplot(faithful, aes(y = waiting)) +
  stat_distrmix_line(components = "sum") +
  stat_distrmix_eq(label.x = "right")

ggplot(faithful, aes(x = waiting)) +
  geom_histogram(aes(y = after_stat(density)), bins = 20) +
  stat_distrmix_line(aes(colour = after_stat(component),
                         fill = after_stat(component)),
                     geom = "area", linewidth = 1, alpha = 0.25) +
  stat_distrmix_eq(aes(colour = after_stat(component)))

ggplot(faithful, aes(x = waiting)) +
  stat_distrmix_line(aes(colour = after_stat(component),
                         fill = after_stat(component)),
                     geom = "area", linewidth = 1, alpha = 0.25,
                     components = "members") +
 stat_distrmix_eq(aes(colour = after_stat(component)),
                     components = "members")

ggplot(faithful, aes(x = waiting)) +
  stat_distrmix_line(geom = "area", linewidth = 1, alpha = 0.25,
                     colour = "black", outline.type = "upper",
                     components = "sum", se = FALSE) +
 stat_distrmix_eq(components = "sum")

# special case of no mixture
ggplot(subset(faithful, waiting > 66), aes(x = waiting)) +
  stat_distrmix_line(k = 1) +
  stat_distrmix_eq(k = 1)

ggplot(subset(faithful, waiting > 66), aes(x = waiting)) +
  stat_distrmix_line(k = 1) +
  stat_distrmix_eq(k = 1, se = TRUE)

# Inspecting the returned data using geom_debug_group()
gginnards.installed <- requireNamespace("gginnards", quietly = TRUE)

if (gginnards.installed)
  library(gginnards)

if (gginnards.installed)
  ggplot(faithful, aes(x = waiting)) +
    stat_distrmix_line(geom = "debug_group", components = "all")
    stat_distrmix_eq(geom = "debug_group", components = "all")

if (gginnards.installed)
  ggplot(faithful, aes(x = waiting)) +
    stat_distrmix_eq(geom = "debug_group", components = "sum")

if (gginnards.installed)
  ggplot(faithful, aes(x = waiting)) +
    stat_distrmix_eq(geom = "debug_group", components = "members")

if (gginnards.installed)
  ggplot(faithful, aes(x = waiting)) +
    stat_distrmix_eq(geom = "debug_group",
                     components = "members",
                     fm.values = TRUE)

}
\seealso{
For the underlying computations see for mixes of two or more Normal
  distributions \code{\link[mixtools]{normalmixEM}()} and for a single Normal
  distribution \code{\link[MASS]{fitdistr}()}.

  \emph{statistics} from 'ggpmisc' for model fit annotations:
  \code{\link{stat_poly_eq}()}, \code{\link{stat_quant_eq}()},
  \code{\link{stat_ma_eq}()} and \code{\link{stat_distrmix_eq}()}, and for
  model fit predictions: \code{\link{stat_poly_line}()},
  \code{\link{stat_quant_line}()}, \code{\link{stat_quant_band}()},
  \code{\link{stat_ma_line}()} and \code{\link{stat_distrmix_line}()}.
}
\section{Aesthetics}{

\code{stat_distrmix_eq()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}} \emph{or} \code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → \code{after_stat(component)} \cr
 • \tab \code{hjust} \tab → \code{"inward"} \cr
 • \tab \code{label} \tab → \code{after_stat(eq.label)} \cr
 • \tab \code{npcx} \tab → \code{after_stat(npcx)} \cr
 • \tab \code{npcy} \tab → \code{after_stat(npcy)} \cr
 • \tab \code{vjust} \tab → \code{"inward"} \cr
}

\code{stat_distrmix_line()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}} \emph{or} \code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → \code{after_stat(component)} \cr
 • \tab \code{weight} \tab → \code{NULL} \cr
}

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}
