% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-fit-broom.R
\name{stat_fit_augment}
\alias{stat_fit_augment}
\title{Augment data with fitted values and statistics}
\usage{
stat_fit_augment(
  mapping = NULL,
  data = NULL,
  geom = "smooth",
  position = "identity",
  ...,
  method = "lm",
  method.args = list(formula = y ~ x),
  n.min = 2L,
  fit.seed = NA,
  augment.args = list(),
  level = 0.95,
  y.out = ".fitted",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}()}. Only needs
to be set at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points on this
layer}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}

\item{method}{character or function.}

\item{method.args, augment.args}{list of arguments to pass to \code{method}
and to to \code{broom::augment}.}

\item{n.min}{integer Minimum number of distinct values in the explanatory
variable (on the rhs of formula) for fitting to the attempted.}

\item{fit.seed}{RNG seed argument passed to
\code{\link[base:Random]{set.seed}()}. Defaults to \code{NA}, which means
that \code{set.seed()} will not be called.}

\item{level}{numeric Level of confidence interval to use (0.95 by default)}

\item{y.out}{character (or numeric) index to column to return as \code{y}.}

\item{na.rm}{logical indicating whether \code{NA} values should be stripped
before the computation proceeds.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}
}
\description{
\code{stat_fit_augment()} fits a model and returns a "tidy"
  version of the model's data with prediction added, using \code{augmnent()}
  methods from packages 'broom', 'broom.mixed', or other sources. The
  prediction can be added to the plot as a line.
}
\details{
\code{stat_fit_augment()} together with
  \code{\link{stat_fit_glance}()} and \code{\link{stat_fit_tidy}()}, based on
  package 'broom' can be used with a broad range of model fitting functions
  as supported at any given time by 'broom'. In contrast to
  \code{\link{stat_poly_eq}()} which can generate text or expression labels
  automatically, for these functions the mapping of aesthetic \code{label}
  needs to be explicitly supplied in the call, and labels built on the fly.

  A ggplot statistic receives as data a data frame that is not the one passed
  as argument by the user, but instead a data frame with the variables mapped
  to aesthetics. In other words, it respects the grammar of graphics and
  consequently within arguments passed through \code{method.args} names of
  aesthetics like _x_ and _y_ should be used instead of the original variable
  names, while data is automatically passed the data frame. This helps ensure
  that the model is fitted to the same data as plotted in other layers.
}
\note{
The statistic \code{stat_fit_augment} can be used only with
  \code{methods} that accept formulas under any formal parameter name and a
  \code{data} argument. Use \code{ggplot2::stat_smooth()} instead of
  \code{stat_fit_augment} in production code if the additional features are
  not needed.

Although arguments passed to parameter \code{augment.args} will be
  passed to \code{\link[generics]{augment}()} whether they are silently
  ignored or obeyed depends on each specialization of \code{augment()}, so do
  carefully read the documentation for the version of \code{augment()}
  corresponding to the \code{method} used to fit the model. Be aware that
  \code{se_fit = FALSE} is the default in these methods even when supported.
}
\section{Warning!}{
 Not all \code{augment()} method specializations are
  defined in package 'broom'. \code{augment()} specializations for mixed
  models fits of classes \code{"lme"}, \code{"nlme"}, \code{"lme4"} and many
  others are defined in package 'broom.mixed'.
}

\section{Handling of grouping}{
 \code{stat_fit_augment()} applies the function
  given by \code{method} separately to each group of observations; in
  'ggplot2' factors mapped to aesthetics generate a separate group for each
  level. Because of this, \code{stat_fit_augment()} is not useful for
  annotating plots with results from \code{t.test()} or ANOVA or ANCOVA
  (e.g., when a factor is mapped to the _x_ or _y_ aesthetics. In such cases
  use instead \code{stat_fit_tb()} which applies the model fitting per panel.
}

\section{Computed variables}{
 The output of \code{augment()} is
  returned as is, except for \code{y} which is set based on \code{y.out} and
  \code{y.observed} which preserves the \code{y} returned by the
  \code{generics::augment} methods. This renaming is needed so that the geom
  works as expected.

  To explore the values returned by this statistic, which vary depending
  on the model fitting function and model formula we suggest the use of
  \code{\link[gginnards]{geom_debug}}. An example is shown below.
}

\examples{
# Package 'broom' needs to be installed to run these examples.
# We check availability before running them to avoid errors.

broom.installed <- requireNamespace("broom", quietly = TRUE)
gginnards.installed <- requireNamespace("gginnards", quietly = TRUE)

if (broom.installed) {
  library(broom)
}

# Inspecting the returned data using geom_debug_group()
  if (gginnards.installed) {
    library(gginnards)
}

# Regression by panel, inspecting data
if (broom.installed & gginnards.installed) {
    ggplot(mtcars, aes(x = disp, y = mpg)) +
      geom_point(aes(colour = factor(cyl))) +
      stat_fit_augment(method = "lm",
                       method.args = list(formula = y ~ x),
                       geom = "debug_group",
                       dbgfun.data = colnames)
}

# Regression by panel example
if (broom.installed)
  ggplot(mtcars, aes(x = disp, y = mpg)) +
    geom_point(aes(colour = factor(cyl))) +
    stat_fit_augment(method = "lm",
                     method.args = list(formula = y ~ x))

if (broom.installed)
  ggplot(mtcars, aes(x = disp, y = mpg)) +
    geom_point(aes(colour = factor(cyl))) +
    stat_fit_augment(method = "lm",
                     augment.args = list(se_fit = TRUE),
                     method.args = list(formula = y ~ x + I(x^2)))

# Residuals from regression by panel example
if (broom.installed)
  ggplot(mtcars, aes(x = disp, y = mpg)) +
    geom_hline(yintercept = 0, linetype = "dotted") +
    stat_fit_augment(geom = "point",
                     method = "lm",
                     method.args = list(formula = y ~ x),
                     y.out = ".resid")

# Regression by group example
if (broom.installed)
  ggplot(mtcars, aes(x = disp, y = mpg, colour = factor(cyl))) +
    geom_point() +
    stat_fit_augment(method = "lm",
                     augment.args = list(se_fit = TRUE),
                     method.args = list(formula = y ~ x))

# Residuals from regression by group example
if (broom.installed)
  ggplot(mtcars, aes(x = disp, y = mpg, colour = factor(cyl))) +
    geom_hline(yintercept = 0, linetype = "dotted") +
    stat_fit_augment(geom = "point",
                     method.args = list(formula = y ~ x),
                     y.out = ".resid")

# Weighted regression example
if (broom.installed)
  ggplot(mtcars, aes(x = disp, y = mpg, weight = cyl)) +
    geom_point(aes(colour = factor(cyl))) +
    stat_fit_augment(method = "lm",
                     method.args = list(formula = y ~ x,
                                        weights = quote(weight)))

# Residuals from weighted regression example
if (broom.installed)
  ggplot(mtcars, aes(x = disp, y = mpg, weight = cyl)) +
    geom_hline(yintercept = 0, linetype = "dotted") +
    stat_fit_augment(geom = "point",
                     method.args = list(formula = y ~ x,
                                        weights = quote(weight)),
                     y.out = ".resid")

}
\seealso{
Package \code{\link[broom]{broom}} for details on how the tidying of
  the result of model fits is done.

Other ggplot statistics for model fits: 
\code{\link{stat_fit_deviations}()},
\code{\link{stat_fit_glance}()},
\code{\link{stat_fit_residuals}()},
\code{\link{stat_fit_tb}()},
\code{\link{stat_fit_tidy}()}
}
\concept{ggplot statistics for model fits}
\section{Aesthetics}{

\code{stat_fit_augment()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → inferred \cr
 • \tab \code{\link[ggplot2:aes_position]{ymax}} \tab → \code{after_stat(y + .se.fit * t.value)} \cr
 • \tab \code{\link[ggplot2:aes_position]{ymin}} \tab → \code{after_stat(y - .se.fit * t.value)} \cr
}

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}
