% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-fit-deviations.R, R/stat-fit-residuals.R
\name{stat_fit_deviations}
\alias{stat_fit_deviations}
\alias{stat_fit_fitted}
\alias{stat_fit_residuals}
\title{Residuals and fitted values from model fit}
\usage{
stat_fit_deviations(
  mapping = NULL,
  data = NULL,
  geom = "segment",
  position = "identity",
  ...,
  orientation = NA,
  method = "lm",
  method.args = list(),
  n.min = 2L,
  formula = NULL,
  fit.seed = NA,
  na.rm = FALSE,
  show.legend = TRUE,
  inherit.aes = TRUE
)

stat_fit_fitted(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  orientation = NA,
  ...,
  method = "lm",
  method.args = list(),
  n.min = 2L,
  formula = NULL,
  fit.seed = NA,
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = TRUE
)

stat_fit_residuals(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  orientation = NA,
  method = "lm",
  method.args = list(),
  n.min = 2L,
  formula = NULL,
  fit.seed = NA,
  resid.type = NULL,
  weighted = FALSE,
  na.rm = FALSE,
  show.legend = TRUE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}()}. Only needs to be set at the layer level if you
are overriding the plot defaults.}

\item{data}{A layer specific dataset, only needed if you want to override the
plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points on this
layer.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}

\item{orientation}{character Either "x" or "y" controlling the default for
\code{formula}. The letter indicates the aesthetic considered the
explanatory variable in the model fit.}

\item{method}{function or character If character, "lm", "rlm", "lmrob",
"lts", "gls", "ma", "sma", "segreg", "rq" or the name of a model fit
function are accepted, possibly followed by the fit function's
\code{method} argument separated by a colon (e.g. \code{"rlm:M"}). If a
function is different to \code{lm()}, \code{rlm()}, \code{ltsReg()},
\code{gls()}, \code{ma}, \code{sma}, it must have formal parameters named
\code{formula}, \code{data}, and \code{weights}. See Details.}

\item{method.args}{named list with additional arguments. Not \code{data}
or \code{weights} which are always passed through aesthetic mappings.}

\item{n.min}{integer Minimum number of distinct values in the explanatory
variable (on the rhs of formula) for fitting to the attempted.}

\item{formula}{a formula object. Using aesthetic names \code{x} and \code{y}
instead of original variable names.}

\item{fit.seed}{RNG seed argument passed to
\code{\link[base:Random]{set.seed}()}. Defaults to \code{NA}, indicating
that \code{set.seed()} should not be called.}

\item{na.rm}{a logical indicating whether NA values should be stripped before
the computation proceeds.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}

\item{resid.type}{character passed to \code{\link[stats]{residuals}()} as
argument for \code{type} (defaults to \code{"working"} except if
\code{weighted = TRUE} when it is forced to \code{"deviance"}).}

\item{weighted}{logical If true weighted residuals will be returned.}
}
\value{
The returned value is always a data frame with the same number of
  rows as the argument passed to \code{data}, except for the case failure of
  the model fitting, in which case a data frame with no rows is returned. The
  columns returned vary between the three statistics, and for each statistic
  depending on the orientation..

  To explore the values returned by statistics we suggest the use of
  \code{\link[gginnards]{geom_debug_group}()}.
}
\description{
Statistic \code{stat_fit_residuals()} fits a model and plots residuals vs.
\code{x}. Statistic \code{stat_fit_deviations()} fits a model and and
highlighting residuals as segments in a \emph{y} vs. \emph{x} plot. Statistic
\code{stat_fit_fitted()} plots the fitted values vs. \emph{x}.
}
\details{
\code{stat_fit_deviations()} can be used to highlight residuals as
  segments in a plot of a fitted model prediction. This statistic returns the
  original \code{x} and \code{y} values and the fitted \code{y} or \code{x}
  values depending on the \emph{orientation}, together with prior and
  posterior weights.

  \code{stat_fit_fitted()} can be used to highlight as points the fitted
  values.  This statistic returns the original \code{x} or \code{y} values
  and the fitted \code{y} or \code{x} values depending on the
  \emph{orientation}.

  \code{stat_fit_residuals()} plots residuals as points. It applies to the
  fitted model object methods \code{\link[stats]{residuals}()} or
  \code{\link[stats]{weighted.residuals}()} depending on the argument passed
  to parameter \code{weighted}. This statistic returns the original \code{x}
  and \code{y} values and residuals depending on the \emph{orientation},
  together with prior and posterior weights.
}
\note{
In the case of \code{method = "rq"} quantiles are fixed at \code{tau =
  0.5} unless \code{method.args} has length > 0. Parameter \code{orientation}
  is redundant as it only affects the default for \code{formula} but is
  included for consistency with \code{ggplot2}.
}
\section{Prior and posterior weights}{

  Two types of weights are possible: prior ones supplied in the call, and
  posterior weights (called "robustness weights" in robust regression
  methods) implicitly or explicitly used by fit methods to address
  heterogeneity of error variance, including the presence of outlier
  observations . Not all the supported methods accepts prior weights and
  \code{gls()} returns posterior weights that are not in 0..1 like in the
  case of most other fits. When not accessible weights are set to 1 when
  known to be equal to 1, which is the most frequent case, or to \code{NA}
  otherwise.

  How weights are applied to residuals depends on the method used to fit the
  model. For ordinary least squares (OLS), weights are applied to the squares
  of the residuals, so the weighted residuals are obtained by multiplying the
  "deviance" residuals by the square root of the weights. When residuals are
  penalized differently to fit a model, the weighted residuals need to be
  computed accordingly.
}

\section{Variables returned by \code{stat_fit_residuals()}}{


  \describe{
  \item{x}{x coordinates of observations}
  \item{y}{y coordinates of observations}
  \item{x.resid}{x residuals from fitted values}
  \item{y.resid}{y residuals from fitted values}
  \item{weights}{the weights
  passed as input to \code{lm()}, \code{rlm()}, \code{lmrob()},
  or to other model fit functions
  using aesthetic weight. More generally the value returned by
  method \code{weights()} applied to the model fit object}
  \item{posterior.weights}{the "weights"
  of the applied minimization criterion relative to those of OLS in
  \code{rlm()} or \code{lmrob()} or the divisor weights from
   \code{gls()}, \code{lme()} or \code{nlme()}}
  }
}

\section{Variables returned by \code{stat_fit_deviations()}}{


  \describe{
  \item{x}{x coordinates of observations}
  \item{y}{y coordinates of observations}
  \item{x.fitted}{x coordinates of fitted values}
  \item{y.fitted}{y coordinates of fitted values}
  \item{weights}{the weights passed as input to \code{lm()}, \code{rlm()}, or \code{lmrob()},
  using aesthetic weight. More generally the value returned by
  \code{weights()}}
  \item{posterior.weights}{the "weights"
  of the applied minimization criterion relative to those of OLS in
  \code{rlm()}, or \code{lmrob()}}
  }
}

\section{Variables returned by \code{stat_fit_fitted()}}{


  \describe{
  \item{x}{x coordinates of observations or fitted}
  \item{y}{y coordinates of observations or fitted}
  }
}

\section{Model formula and model fitting}{

  A ggplot statistic receives as \code{data} a data frame that is not the one
  passed as argument by the user, but instead a data frame with the variables
  mapped to aesthetics. In \code{stat_poly_eq()} the compute function is
  applied by group, each call "seeing" the subset of \code{data} for an
  individual group. As supported models are for regression lines,
  variables mapped to \code{x} and \code{y} should both be continuous, i.e.,
  numeric or date time and model formulas defined using \code{x} and \code{y}
  as variable names.

  The interpretation of the argument passed to \code{formula} is enhanced
  compared to \code{stat_smooth()}. Formulas with \code{x} as explanatory
  variable work as in \code{stat_smooth()} but formulas with \code{y} as
  explanatory variable are also accepted. \code{orientation} is set
  automatically based on which explanatory variable appears in the formula.
  Spline-based smoothers are only partially supported.
}

\section{Model fit methods supported}{

  Several model fit functions are supported explicitly (see tables), and some
  of their differences smoothed out. Compatibility is checked late, based on
  the class of the returned fitted model object. This makes it possible to
  use wrapper functions that do model selection or other adjustments to the
  fit procedure on a per panel or per group basis. Moreover, if the value
  returned as model fit object is \code{NULL} or \code{NA}, plotting is
  skipped on a per group within panel basis.

  \emph{In the case of fitted model objects of classes not explicitly
  supported, an attempt is made to find the usual accessors and/or fitted
  object members, and if found, either complete or partial support is
  frequently achieved. In this case a message is issued encouraging users to
  check the validity of the values extracted as the structure of fitted model
  objects belonging to different classes and the values returned by their
  accessors can vary, potentially resulting in decoding errors leading to the
  return of wrong values for estimates.}

  The argument to parameter \code{method} can be either the name of a
  function object, possibly using double colon notation in case its package
  is not attached, or a character string matching the function name for
  functions in the search path. This approach makes it possible to support
  model fit functions that are not dependencies of 'ggpmisc'. Either by
  attaching the package where the function is defined and passing it by name
  or as string, or using double colon notation when passing the name of the
  function.

  User-defined functions can be passed as argument to parameter \code{method}
  as long as they have parameters \code{formula}, \code{data} \code{subset}
  and possibly \code{weights}. Additional arguments can be passed to any
  method as a named list through parameter \code{method.args}. As in
  \code{\link[ggplot2:geom_smooth]{stat_smooth}()} prior \code{weights} are
  passed to the model fit functions' \code{weights} (plural!) parameter by
  mapping a numeric variable to plot aesthetic \code{weight} (singular!).

  Tables 1 lists natively supported model fit functions, with the
  caveat that only some 'broom' methods' specializations have been actually
  tested with statistics from 'ggpmisc'. In addition, the statistics based
  on 'broom' methods require the user to tailor their behaviour by passing
  additional arguments in the call and occasionally some detective work to
  find out the names of variables in the returned data frame as these names
  are set by methods from 'broom'.

  \strong{Table 1.} Model fit methods supported by the different statistics
  available in package 'ggpmisc'. Column \eqn{f} indicates whether
  computations are done by group (G) or by plot panel (P).
  \tabular{lcl}{
  \strong{Statistic} \tab \eqn{f} \tab \strong{Supported model fit methods} \cr
  \code{\link{stat_poly_line}()} \tab G \tab "lm", "rlm", "lts", "sma", "ma", "gls", "nls", "onls", \emph{others with methods} \code{\link[stats]{predict}()} or \code{\link[stats]{fitted}()} \cr
  \code{\link{stat_poly_eq}()}   \tab G \tab "lm", "rlm", "lts", "sma", "ma", "gls", "nls", "onls", \emph{others with needed accesors} \cr
  \code{\link{stat_quant_line}()} \tab G \tab "rq", "rqss" \cr
  \code{\link{stat_quant_band}()} \tab G \tab "rq", "rqss" \cr
  \code{\link{stat_quant_eq}()} \tab G \tab "rq", "rqss" \cr
  \code{\link{stat_ma_line}()} \tab G \tab "SMA", "MA", "RMA", "OLS" \cr
  \code{\link{stat_ma_eq}()} \tab G \tab "SMA", "MA", "RMA", "OLS" \cr
  \code{\link{stat_fit_residuals}()} \tab G \tab "lm", "rlm", "lts", "sma", "ma", "gls", "rq", "rqss", "nls", "onls", \emph{others with method} \code{\link[stats]{residuals}()} \cr
  \code{\link{stat_fit_fitted}()} \tab G \tab "lm", "rlm", "lts", "gls", "rq", "rqss", "nls", "onls", \emph{others with method} \code{\link[stats]{fitted}()} \cr
  \code{\link{stat_fit_deviations}()} \tab G \tab "lm", "rlm", "lts", "gls", "rq", "rqss", "nls", "onls", \emph{others with methods} \code{\link[stats]{fitted}()} and \code{\link[stats]{weights}()} \cr
  \code{\link{stat_fit_augment}()} \tab G \tab \emph{any with 'broom' method} \code{\link[broom]{augment}()} \cr
  \code{\link{stat_fit_glance}()} \tab G \tab \emph{any with 'broom' method} \code{\link[broom]{glance}()} \cr
  \code{\link{stat_fit_tidy}()} \tab G \tab \emph{any with 'broom' method} \code{\link[broom]{tidy}()} \cr
  \code{\link{stat_fit_tb}()} \tab P \tab \emph{any with 'broom' method} \code{\link[broom]{tidy}()} \cr
  }

  The single colon notation is based on parsing
  the name and is available when passing the name of the fit method as a
  character string. In a string such as "head:tail" the "head" gives the name
  of the model fit function and the "tail" gives the argument to pass it's
  \code{method} parameter. This is only a convenience, as \code{method.args}
  can be also used. In some methods, i.e., splines, the default
  \code{formula = y ~ x} needs to be overridden by the user.

  Table 2 lists the correspondence of pre-defined \emph{method names}
  to model fit method functions. As mentioned above, these are only
  a subset of the model fit methods that are expected to work. When using
  these names there is no need for users to attach additional packages but
  the packages must be available (installed).

  \strong{Table 2.} Available predefined method names, the model fit functions
  they call, the packages where the functions reside, the class of the
  returned fitted model object and the arguments that can be
  passed to their \code{method} parameter using single colon notation.
  \tabular{llll}{
  \strong{Predefined method names} \tab \strong{Model fit methods} \tab \strong{R package} \tab \strong{Object class} \cr
  "lm", "lm:qr" \tab \code{\link[stats]{lm}()} \tab 'stats' \tab "lm" \cr
  "rlm", "rlm:M", "rlm:MM" \tab \code{\link[MASS]{rlm}()} \tab 'MASS' \tab "rlm" ("lm") \cr
  "lts", "ltsReg" \tab \code{\link[robustbase]{ltsReg}()} \tab 'robustbase' \tab "lts" \cr
  "ma", "sma", "sma:SMA", "sma:MA", "sma:OLS" \tab \code{\link[smatr]{sma}()} \tab 'smatr' \tab "ma" or "sma" \cr
  "gls", "gls:REML", "gls:ML" \tab \code{\link[nlme]{gls}()} \tab 'nlme' \tab "gls" \cr
  "rq", "rq:sfn", "rq:sfnc", "rq:lasso" \tab \code{\link[quantreg]{rq}()} \tab 'quantreg' \tab "rq" \cr
  "rqss", "rqss:sfn", "rqss:sfnc", "rqss:lasso" \tab \code{\link[quantreg]{rqss}()} \tab 'quantreg' \tab "rqss" \cr
  "SMA", "MA", "RMA", "OLS" \tab \code{\link[lmodel2]{lmodel2}()} \tab 'lmodel2' \tab ("list") \cr
  }
}

\section{Which variables are available for mapping?}{

  Computed variables and their names vary following those used in the
  specializations of 'broom' methods. A message is issued by default in
  interactive R sessions, listing the names of all variables accessible by
  \code{after_stat()} within a call to \code{aes()} with at least some
  non-missing values. This default (\code{"colnames"}) can be changed by
  setting R option \code{"ggpmisc.stat.vars.message"} to \code{"none"}.
}

\section{Model equation label}{

  By default the equation label uses as symbols the names of the aesthetics,
  \code{x} and \code{y}. However, \code{"x"} and \code{"y"} can be
  substituted by providing a replacement character string for the
  right-hand-side and left-hand-side through \code{eq.x.rhs} and
  \code{eq.with.lhs}, respectively. For backward compatibility a logical is
  also accepted as argument for \code{eq.with.lhs}, with \code{FALSE}
  suppressing the left-hand-side.

  If the model \code{formula} includes a transformation of the explanatory
  variable in its right-hand-side (rhs), a matching argument should be passed
  to parameter \code{eq.x.rhs} as its default value would result in an
  equation label that does not reflect the applied transformation. In most
  cases, a transformation should not be applied within the left hand side
  (lhs) of the model formula, but instead in the mapping of the response
  variable within \code{aes}. In this case it may be necessary to also pass a
  matching argument to parameter \code{eq.with.lhs}.

  Parameter \code{orientation} is redundant as the orientation can be set
  by the \code{formula} but is included for consistency with
  \code{ggplot2::stat_smooth()}.
}

\section{Position of labels}{

  When data are grouped by mapping a factor to an aesthetic, e.g.,
  \code{colour}, \code{shape} and/or \code{linetype} the model is fitted
  separately to each group, and for each group a whole set of labels is
  generated. If the argument passed to \code{label.y} is a vector of length
  1, this value determines the position of the equation and/or other labels
  for the first group, and the positions of the labels for the remaining
  groups are generated by adding \code{vspace} based on the group number.
  If the argument passed to \code{label.y} is a vector of length > 1, it is
  used unchanged, possibly extended by recycling, ignoring \code{vstep}.

  If the labels are rotated by 90 degrees then the automatic stepping is
  best based on \code{hstep} with \code{vstep = 0}. Similarly as described
  above, if \code{label.x} is a vector of length > 1, it is
  used unchanged, possibly extended by recycling, ignoring \code{hstep}.

  When using facets and with a grouping that does not repeat in each panel,
  the automatic positioning in most cases will not be the desired one. Manual
  positioning using a vector of length > 1 for \code{label.x} and/or
  \code{label.y} is the currently available workaround.
}

\section{Range of the prediction line}{
 The range of the prediction line is
  controlled by parameters \code{fullrange} and \code{limit.to}.
  \code{fullrange} is backwards compatible both with earlier versions of
  'ggpmisc' and with \code{stat_smooth()} from 'ggplot2'; an argument passed
  to \code{limit.to} overrides \code{fullrange} making it possible to
  constrain the range to that of \code{x}, \code{y}, or both simultaneously,
  with \code{"x"}, \code{"y"}, or \code{"xy"}, respectively, as argument.
  \code{limit.to} also accepts a numeric vector of values to be used as
  \code{newdata} when computing the prediction. Limiting the range based on
  both aesthetics is the best approach for major axis regression (MA, SMA,
  RMA) but can occasionally be useful also with some other methods when
  slopes are very steep and error variance in the explanatory variable is
  large. A numeric vector can be used to predict the response at specific
  values of the explanatory variable. If a single or very few values are
  predicted, it can be necessary to override the default \code{geom =
  "smooth"} with \code{geom = "pointrange"}.
}

\section{Variables returned by \code{stat_poly_line()}}{


  \describe{ \item{y \strong{or} x}{predicted value}
  \item{ymin \strong{or} xmin}{lower confidence limit around the fitted line}
  \item{ymax \strong{or} xmax}{upper confidence limit around the fitted line}
  \item{se}{standard error} }

  If \code{fm.values = TRUE} is passed then columns based on the summary of
  the model fit are added, with the same value in each row within a group.
  This is wasteful and disabled by default, but provides a simple and robust
  approach to achieve effects like colouring or hiding of the model fit line
  based on \eqn{P}, \eqn{R^2}, \eqn{R_{adj}^2} or the number of
  observations in a fit.
}

\section{Variables returned by \code{stat_poly_eq()}}{


For all \code{output.type} arguments the following values are returned.
\describe{
  \item{x,npcx}{x position}
  \item{y,npcy}{y position}
  \item{coefs}{fitted coefficients, named numeric vector as a list member}
  \item{r.squared, rr.confint.level, rr.confint.low, rr.confint.high, adj.r.squared, f.value, f.df1, f.df2, p.value, AIC, BIC, n, knots, knots.se}{numeric values, from the model fit object}
  \item{grp.label}{Set according to mapping in \code{aes}.}
  \item{knots}{list containing a numeric vector of knot or "psi" \emph{x}-value for linear splines}
  \item{fm.method}{name of method used, character}
  \item{fm.class}{most derived class or the fitted model object, character}
  \item{fm.formula.chr}{formatted model formula, character}}

If \code{output.type} is not \code{"numeric"} the returned tibble contains in
addition to those above the columns listed below, each containing a single
character string. The markup used depends on the value of \code{output.type}.
\describe{
  \item{eq.label}{equation for the fitted polynomial as a character string to be parsed or \code{NA}}
  \item{rr.label}{\eqn{R^2} of the fitted model as a character string to be parsed}
  \item{adj.rr.label}{Adjusted \eqn{R^2} of the fitted model as a character string to be parsed}
  \item{rr.confint.label}{Confidence interval for \eqn{R^2} of the fitted model as a character string to be parsed}
  \item{f.value.label}{F value and degrees of freedom for the fitted model as a whole.}
  \item{p.value.label}{P-value for the F-value above.}
  \item{AIC.label}{AIC for the fitted model.}
  \item{BIC.label}{BIC for the fitted model.}
  \item{n.label}{Number of observations used in the fit.}
  \item{knots.label}{The knots or change points in segmented regression.}
  \item{grp.label}{Set according to mapping in \code{aes}.}
  \item{method.label}{Set according \code{method} used.}}

If output.type is \code{"numeric"} the returned tibble contains columns
listed below in addition to the base ones. If the model fit function used
does not return a value, the variable is set to \code{NA_real_}.
\describe{
  \item{coef.ls}{list containing the "coefficients" matrix from the summary of the fit object}
  \item{b_0.constant}{TRUE is polynomial is forced through the origin}
  \item{b_i}{One or more columns with the coefficient estimates}}

To explore the computed values returned for a given input we suggest the use
of \code{\link[gginnards]{geom_debug}()} as shown in the last examples below.
}

\section{Output types}{
 The formatting of character strings to be displayed
in plots are marked as mathematical equations. Depending on the geom used,
the mark-up needs to be encoded differently, or in some cases mark-up not
applied.
\describe{
  \item{\code{"expression"}}{The labels are encoded as character strings to be parsed into R's plotmath expressions.}
  \item{\code{"LaTeX", "TeX", "tikz", "latex"}}{The labels are encoded as 'LaTeX' maths equations, without the "fences" for switching in math mode.}
  \item{\code{"latex.eqn"}}{Same as \code{"latex"} but enclosed in single \code{$}, i.e., as in-line maths.}
  \item{\code{"latex.deqn"}}{Same as \code{"latex"} but enclosed in double \code{$$}, i.e., as display maths.}
  \item{\code{"markdown"}}{The labels are encoded as character strings using markdown syntax, with some embedded HTML.}
  \item{\code{"marquee"}}{The labels are encoded as character strings using markdown syntax, with 'marquee' supported spans.}
  \item{\code{"text"}}{The labels are plain ASCII character strings.}
  \item{\code{"numeric"}}{No labels are generated. This value is accepted by the statistics, but not by the label formatting functions.}
  \item{\code{NULL}}{The value used depends on the argument passed to \code{geom}.}}

If \code{geom = "latex"} (package 'xdvir') the output type used is
\code{"latex.eqn"}. If \code{geom = "richtext"} (package 'ggtext') or
\code{geom = "textbox"} (package 'ggtext') the output type used is
\code{"markdown"}. If \code{geom = "marquee"} (package 'marquee') the output
type used is \code{"marquee"}. For all other values of \code{geom} the default
is \code{"expression"}. Invalid values as argument trigger an error.
}

\examples{
# generate artificial data
set.seed(4321)
x <- 1:100
y <- (x + x^2 + x^3) + rnorm(length(x), mean = 0, sd = mean(x^3) / 4)
my.data <- data.frame(x, y)

# give a name to a formula
my.formula <- y ~ poly(x, 3, raw = TRUE)
my.y.formula <- x ~ poly(y, 3, raw = TRUE)

# plot residuals from linear model
ggplot(my.data, aes(x, y)) +
  stat_poly_line(method = "lm", formula = my.formula) +
  stat_fit_deviations(method = "lm", formula = my.formula, colour = "red") +
  geom_point()

# plot residuals from linear model with y as explanatory variable
ggplot(my.data, aes(x, y)) +
  stat_poly_line(method = "lm", formula = my.y.formula) +
  stat_fit_deviations(method = "lm", formula = my.y.formula, colour = "red") +
  geom_point()

# plot robust regression
ggplot(my.data, aes(x, y)) +
  stat_poly_line(formula = my.formula, method = "rlm") +
  stat_fit_deviations(formula = my.formula, method = "rlm", colour = "red") +
  geom_point()

# plot robust regression with weights indicated by colour
my.data.outlier <- my.data
my.data.outlier[6, "y"] <- my.data.outlier[6, "y"] * 5
ggplot(my.data.outlier, aes(x, y)) +
  stat_poly_line(method = MASS::rlm, formula = my.formula) +
  stat_fit_deviations(formula = my.formula, method = "rlm",
                      mapping = aes(colour = after_stat(posterior.weights)),
                      show.legend = TRUE) +
  scale_color_gradient(low = "red", high = "blue", limits = c(0, 1),
                       guide = "colourbar") +
  geom_point()

# plot quantile regression (= median regression)
ggplot(my.data, aes(x, y)) +
  stat_quantile(formula = my.formula, quantiles = 0.5) +
  stat_fit_deviations(formula = my.formula, method = "rq", colour = "red") +
  geom_point()

# plot quantile regression (= "quartile" regression)
ggplot(my.data, aes(x, y)) +
  stat_quantile(formula = my.formula, quantiles = 0.75) +
  stat_fit_deviations(formula = my.formula, colour = "red",
                      method = "rq", method.args = list(tau = 0.75)) +
  geom_point()

# plot residuals from linear model
ggplot(my.data, aes(x, y)) +
  geom_hline(yintercept = 0, linetype = "dashed") +
  stat_fit_residuals(formula = my.formula)

# plot residuals from linear model with y as explanatory variable
ggplot(my.data, aes(x, y)) +
  geom_vline(xintercept = 0, linetype = "dashed") +
  stat_fit_residuals(formula = my.y.formula) +
  coord_flip()

ggplot(my.data, aes(x, y)) +
  geom_hline(yintercept = 0, linetype = "dashed") +
  stat_fit_residuals(formula = my.formula, resid.type = "response")

# plot residuals with weights indicated by colour
my.data.outlier <- my.data
my.data.outlier[6, "y"] <- my.data.outlier[6, "y"] * 5
ggplot(my.data.outlier, aes(x, y)) +
  stat_fit_residuals(formula = my.formula, method = "rlm",
                      mapping = aes(colour = after_stat(posterior.weights)),
                      show.legend = TRUE) +
  scale_color_gradient(low = "red", high = "blue", limits = c(0, 1),
                       guide = "colourbar")

# plot weighted residuals with weights indicated by colour
ggplot(my.data.outlier) +
  stat_fit_residuals(formula = my.formula, method = "rlm",
                     mapping = aes(x = x,
                                   y = stage(start = y, after_stat = y * weights),
                                   colour = after_stat(posterior.weights)),
                     show.legend = TRUE) +
  scale_color_gradient(low = "red", high = "blue", limits = c(0, 1),
                       guide = "colourbar")

# inspecting the returned data
gginnards.installed <- requireNamespace("gginnards", quietly = TRUE)

if (gginnards.installed)
  library(gginnards)

# plot, using geom_debug_group() to explore the after_stat data
if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    stat_fit_deviations(formula = my.formula,
                        geom = "debug_group")

if (gginnards.installed)
  ggplot(my.data.outlier, aes(x, y)) +
    stat_fit_deviations(formula = my.formula, method = "rlm",
                        geom = "debug_group")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
   stat_fit_residuals(formula = my.formula, resid.type = "working",
                      geom = "debug_group")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    stat_fit_residuals(formula = my.formula, method = "rlm",
                       geom = "debug_group")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
   stat_fit_fitted(formula = my.formula,
                   geom = "debug_group")

}
\references{
Originally written as an answer to question 7549694 at
  Stackoverflow but enhanced based on suggestions from several users and my
  own needs.
}
\seealso{
\code{\link[stats]{residuals}()} and \code{\link[stats]{weights}()}
  and their specializations for the \code{method} used.

  Please, see the articles at
  \href{https://docs.r4photobiology.info/ggpmisc/}{online-only documentation}
  for additional use examples and guidance.
}
\concept{statistics for display of model fit residuals}
\section{Aesthetics}{

\code{stat_fit_residuals()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → inferred \cr
}

\code{stat_fit_deviations()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → inferred \cr
 • \tab \code{\link[ggplot2:aes_position]{xend}} \tab → \code{after_stat(x.fitted)} \cr
 • \tab \code{\link[ggplot2:aes_position]{yend}} \tab → \code{after_stat(y.fitted)} \cr
}

\code{stat_fit_fitted()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → inferred \cr
}

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}
