% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-fit-broom.R
\name{stat_fit_tb}
\alias{stat_fit_tb}
\title{Return a tibble within a tibble with fitted parameter estimates.}
\usage{
stat_fit_tb(mapping = NULL, data = NULL, geom = "table", method = "lm",
  method.args = list(formula = y ~ x), tb.type = "fit.summary",
  digits = 3, label.x.npc = "center", label.y.npc = "top",
  label.x = NULL, label.y = NULL, position = "identity", na.rm = FALSE,
  show.legend = FALSE, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_string}}. Only needs
to be set at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{method}{character.}

\item{method.args}{list of arguments to pass to \code{method}.}

\item{tb.type}{character One of "fit.summary", "fit.anova" or "fit.coefs".}

\item{digits}{integer indicating the number of significant digits to be used.}

\item{label.x.npc, label.y.npc}{\code{numeric} with range 0..1 or character.
Coordinates to be used for positioning the output, expressed in "normalized
parent coordinates" or character string. If too short they will be recycled.}

\item{label.x, label.y}{\code{numeric} Coordinates (in data units) to be used
for absolute positioning of the output. If too short they will be recycled.}

\item{position}{The position adjustment to use for overlapping points on this
layer}

\item{na.rm}{a logical indicating whether NA values should be stripped
before the computation proceeds.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}
}
\description{
\code{stat_fit_tb} fits a model and returns a "tidy" version
  of the model's summary, using package 'broom'.
}
\section{Computed variables}{
 The output of \code{\link[broom]{tidy}} is
  returned as a single "cell" in a tibble (i.e. a tibble nested within a
  tibble). The returned \code{data} object contains a single row, containing
  the
  result from a single model fit to all data in a panel. If grouping is
  present, it is ignored.
}

