% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-ma-eq.R
\name{stat_ma_eq}
\alias{stat_ma_eq}
\title{Equation, p-value, R^2 of major axis regression}
\usage{
stat_ma_eq(
  mapping = NULL,
  data = NULL,
  geom = "text_npc",
  position = "identity",
  ...,
  orientation = NA,
  formula = NULL,
  method = "lmodel2:MA",
  method.args = list(),
  n.min = 2L,
  range.y = NULL,
  range.x = NULL,
  nperm = 99,
  fit.seed = NA,
  eq.with.lhs = TRUE,
  eq.x.rhs = NULL,
  small.r = getOption("ggpmisc.small.r", default = FALSE),
  small.p = getOption("ggpmisc.small.p", default = FALSE),
  coef.digits = 3,
  coef.keep.zeros = TRUE,
  decreasing = getOption("ggpmisc.decreasing.poly.eq", FALSE),
  rr.digits = 2,
  theta.digits = 2,
  p.digits = max(1, ceiling(log10(nperm))),
  label.x = "left",
  label.y = "top",
  hstep = 0,
  vstep = NULL,
  output.type = NULL,
  na.rm = FALSE,
  parse = NULL,
  show.legend = FALSE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}}. Only needs to be set at the layer level if you
are overriding the plot defaults.}

\item{data}{A layer specific dataset, only needed if you want to override the
plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points on this
layer.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}

\item{orientation}{character Either "x" or "y" controlling the default for
\code{formula}. The letter indicates the aesthetic considered the
explanatory variable in the model fit.}

\item{formula}{a formula object. Using aesthetic names \code{x} and \code{y}
instead of original variable names.}

\item{method}{function or character If character, "MA", "SMA" , "RMA" or
"OLS", alternatively "lmodel2" or the name of a model fit function are
accepted, possibly followed by the fit function's \code{method} argument
separated by a colon (e.g. \code{"lmodel2:MA"}). If a function different to
\code{lmodel2()}, it must accept arguments named \code{formula},
\code{data}, \code{range.y}, \code{range.x} and \code{nperm} and return a
model fit object of class \code{lmodel2}.}

\item{method.args}{named list with additional arguments. Not \code{data}
or \code{weights} which are always passed through aesthetic mappings.}

\item{n.min}{integer Minimum number of distinct values in the explanatory
variable (on the rhs of formula) for fitting to the attempted.}

\item{range.y, range.x}{character Pass "relative" or "interval" if method
"RMA" is to be computed.}

\item{nperm}{integer Number of permutation used to estimate significance.}

\item{fit.seed}{RNG seed argument passed to
\code{\link[base:Random]{set.seed}()}. Defaults to \code{NA}, indicating
that \code{set.seed()} should not be called.}

\item{eq.with.lhs}{If \code{character} the string is pasted to the front of
the equation label before parsing or a \code{logical} (see note).}

\item{eq.x.rhs}{\code{character} this string will be used as replacement for
\code{"x"} in the model equation when generating the label before parsing
it.}

\item{small.r, small.p}{logical Flags to switch use of lower case r and p for
coefficient of determination and p-value.}

\item{coef.digits}{integer Number of significant digits to use for
the fitted coefficients.}

\item{coef.keep.zeros}{logical Keep or drop trailing zeros when formatting
the fitted coefficients and F-value.}

\item{decreasing}{logical It specifies the order of the terms in the
returned character string; in increasing (default) or decreasing powers.}

\item{rr.digits, theta.digits, p.digits}{integer Number of digits after the
decimal point to use for R^2, theta and P-value in labels. If \code{Inf},
use exponential notation with three decimal places.}

\item{label.x, label.y}{\code{numeric} with range 0..1 "normalized parent
coordinates" (npc units) or character if using \code{geom_text_npc()} or
\code{geom_label_npc()}. If using \code{geom_text()} or \code{geom_label()}
numeric in native data units. If too short they will be recycled.}

\item{hstep, vstep}{numeric in npc units, the horizontal and vertical step
used between labels for different groups.}

\item{output.type}{character One of "expression", "LaTeX", "text",
"markdown" or "numeric".}

\item{na.rm}{a logical indicating whether NA values should be stripped before
the computation proceeds.}

\item{parse}{logical Passed to the geom. If \code{TRUE}, the labels will be
parsed into expressions and displayed as described in \code{?plotmath}.
Default is \code{TRUE} if \code{output.type = "expression"} and
\code{FALSE} otherwise.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}
}
\value{
A data frame, with a single row and columns as described under
  \strong{Computed variables}. In cases when the number of observations is
  less than \code{n.min} a data frame with no rows or columns is returned
  rendered as an empty/invisible plot layer.
}
\description{
\code{stat_ma_eq} fits model II regressions. From the fitted model it
generates several labels including the equation, p-value,
coefficient of determination (R^2), and number of observations.
}
\details{
This stat can be used to automatically annotate a plot with
  \eqn{R^2}, \eqn{P}-value, \eqn{n} and/or the fitted model equation. It
  supports linear major axis (MA), standard major axis (SMA) and ranged major
  axis (RMA) regression by means of function \code{\link[lmodel2]{lmodel2}}.
  Formulas describing a straight line and including an intercept are the
  only ones currently supported. Please see the documentation, including the
  vignette of package 'lmodel2' for details. The parameters in
  \code{stat_ma_eq()} follow the same naming as in function \code{lmodel2()}.

  It is important to keep in mind that although the fitted line does not
  depend on whether the \eqn{x} or \eqn{y} appears on the rhs of the model
  formula, the numeric estimates for the parameters do depend on this.

  A ggplot statistic receives as \code{data} a data frame that is not the one
  passed as argument by the user, but instead a data frame with the variables
  mapped to aesthetics. \code{stat_ma_eq()} mimics how \code{stat_smooth()}
  works, except that Model II regressions can be fitted. Similarly to
  \code{stat_smooth()} the model is fitted separately to data from each
  group, so the variables mapped to \code{x} and \code{y} should both be
  continuous rather than discrete as well as the corresponding scales.

  The minimum number of observations with distinct values can be set through
  parameter \code{n.min}. The default \code{n.min = 2L} is the smallest
  possible value. However, model fits with very few observations are of
  little interest and using a larger number for \code{n.min} than the default
  is usually wise. As model fitting functions can depend on
  the RNG, \code{fit.seed} if different to \code{NA} is used as argument in a
  call to \code{\link[base:Random]{set.seed}()} immediately ahead of model
  fitting.
}
\note{
For backward compatibility a logical is accepted as argument for
  \code{eq.with.lhs}. If \code{TRUE}, the default is used, either
  \code{"x"} or \code{"y"}, depending on the argument passed to \code{formula}.
  However, \code{"x"} or \code{"y"} can be substituted by providing a
  suitable replacement character string through \code{eq.x.rhs}.
  Parameter \code{orientation} is redundant as it only affects the default
  for \code{formula} but is included for consistency with
  \code{ggplot2::stat_smooth()}.

  Methods in \code{\link[lmodel2]{lmodel2}} are all computed always except
  for RMA that requires a numeric argument to at least one of \code{range.y}
  or \code{range.x}. The results for specific methods are extracted a
  posteriori from the model fit object. When a function is passed as argument
  to \code{method}, the method can be passed in a list to \code{method.args}
  as member \code{method}. More easily, the name of the function can be
  passed as a character string together with the \code{lmodel2}-supported
  method.

  R option \code{OutDec} is obeyed based on its value at the time the plot
  is rendered, i.e., displayed or printed. Set \code{options(OutDec = ",")}
  for languages like Spanish or French.

\code{stat_ma_eq} understands \code{x} and \code{y}, to
  be referenced in the \code{formula} while the \code{weight} aesthetic is
  ignored. Both \code{x} and \code{y} must be mapped to \code{numeric}
  variables. In addition, the aesthetics understood by the geom
  (\code{"text"} is the default) are understood and grouping respected.

  \emph{Transformation of \code{x} or \code{y} within the model formula
  is not supported by \code{stat_ma_eq()}. In this case, transformations
  should not be applied in the model formula, but instead in the mapping
  of the variables within \code{aes} or in the scales.}
}
\section{User-defined methods}{
 User-defined functions can be passed as
  argument to \code{method}. The requirements are 1) that the signature is
  similar to that of function \code{lmodel2()} and 2) that the value returned
  by the function is an object as returned by \code{lmodel2()} or an atomic
  \code{NA} value. Thus, user-defined methods can implement conditional
  skipping of labelling.
}

\section{Computed variables}{

If \code{output.type} is different from \code{"numeric"} the returned tibble
contains columns listed below. If the fitted model does not contain a given
value, the label is set to \code{character(0L)}.
\describe{
  \item{x,npcx}{x position}
  \item{y,npcy}{y position}
  \item{eq.label}{equation for the fitted polynomial as a character string to be parsed}
  \item{rr.label}{\eqn{R^2} of the fitted model as a character string to be parsed}
  \item{p.value.label}{P-value if available, depends on \code{method}.}
  \item{theta.label}{Angle in degrees between the two OLS lines for lines estimated from \code{y ~ x} and \code{x ~ y} linear model (\code{lm}) fits.}
  \item{n.label}{Number of observations used in the fit.}
  \item{grp.label}{Set according to mapping in \code{aes}.}
  \item{method.label}{Set according \code{method} used.}
  \item{r.squared, theta, p.value, n}{numeric values, from the model fit object}}

If output.type is \code{"numeric"} the returned tibble contains columns
listed below. If the model fit function used does not return a value,
the variable is set to \code{NA_real_}.
\describe{
  \item{x,npcx}{x position}
  \item{y,npcy}{y position}
  \item{coef.ls}{list containing the "coefficients" matrix from the summary of the fit object}
  \item{r.squared, theta, p.value, n}{numeric values, from the model fit object}
  \item{grp.label}{Set according to mapping in \code{aes}.}
  \item{b_0.constant}{TRUE is polynomial is forced through the origin}
  \item{b_i}{One or two columns with the coefficient estimates}}

To explore the computed values returned for a given input we suggest the use
of \code{\link[gginnards]{geom_debug}} as shown in the last examples below.
}

\section{Output types}{
 The formatting of character strings to be displayed
in plots are marked as mathematical equations. Depending on the geom used,
the mark-up needs to be encoded differently, or in some cases mark-up not
applied.
\describe{
  \item{\code{"expression"}}{The labels are encoded as character strings to be parsed into R's plotmath expressions.}
  \item{\code{"LaTeX", "TeX", "tikz", "latex"}}{The labels are encoded as 'LaTeX' maths equations, without the "fences" for switching in math mode.}
  \item{\code{"latex.eqn"}}{Same as \code{"latex"} but enclosed in single \code{$}, i.e., as in-line maths.}
  \item{\code{"latex.deqn"}}{Same as \code{"latex"} but enclosed in double \code{$$}, i.e., as display maths.}
  \item{\code{"markdown"}}{The labels are encoded as character strings using markdown syntax, with some embedded HTML.}
  \item{\code{"marquee"}}{The labels are encoded as character strings using markdown syntax, with 'marquee' supported spans.}
  \item{\code{"text"}}{The labels are plain ASCII character strings.}
  \item{\code{"numeric"}}{No labels are generated. This value is accepted by the statistics, but not by the label formatting functions.}
  \item{\code{NULL}}{The value used, \code{expression}, \code{latex.eqn} or \code{markup} depends on the argument passed to \code{geom}.}}

If \code{geom = "latex"} (package 'xdvir') the output type used is
\code{"latex.eqn"}. If \code{geom = "richtext"} (package 'ggtext') or
\code{geom = "textbox"} (package 'ggtext') the output type used is
\code{"markdown"}. If \code{geom = "marquee"} (package 'marquee') the output
type used is \code{"marquee"}. For all other values of \code{geom} the default
is \code{"expression"} unless the user passes an argument. Invalid values as
argument trigger an Error.
}

\section{Model fit methods supported}{
 Several model fit functions are
  supported explicitly (see tables), and some of their differences smoothed
  out. Compatibility is checked late, based on the class of the returned
  fitted model object. This makes it possible to use wrapper functions that
  do model selection or other adjustments to the fit procedure on a per panel
  or per group basis. Moreover, if the value returned as model fit object is
  \code{NULL} no layer is added to the plot on a per group within panel
  basis.

  In the case of fitted model objects of classes not explicitly supported an
  attempt is made to find the usual accessors and/or fitted object members,
  and if found, either complete or partial support is frequently achieved.
  In this case a message is issued encouraging users to check the valisdity
  of the values extracted.

  The argument to parameter \code{method} can be either the name of a
  function object, possibly using double colon notation, or a character
  string matching the function name. This approach makes it possible to
  support model fit functions that are not dependencies of 'ggpmisc'. Either
  by attaching the package where the function is defined and passing it by
  name or as string, or using double colon notation when passing the name of
  the function. User-defined functions can be passed as argument to parameter
  \code{method} as long as they have parameters \code{formula}, \code{data}
  \code{subset} and possibly \code{weights}. Additional arguments can be
  passed to any method as a named list as an argument to parameter
  \code{method.args}. As in \code{\link[ggplot2:geom_smooth]{stat_smooth}()}
  prior \code{weights} are passed to the model fit functions' \code{weights}
  (plural!) parameter by mapping a numeric variable to plot aesthetic
  \code{weight} (singular!).

  The table below lists natively supported model fit functions, with the
  caveat that only some 'broom' methods' specializations have been actually
  tested with statistics from 'ggpmisc'. In addition, the statistics based
  on 'broom' methods require the user to tailor their behaviour by passing
  additional arguments in the call.

  \tabular{lcl}{
  \strong{Statistic} \tab \eqn{f} \tab \strong{Supported model fit methods} \cr
  \code{\link{stat_poly_line}()} \tab G \tab "lm", "rlm", "lts", "sma", "ma", "gls", \emph{others with methods} \code{\link[stats]{predict}()} or \code{\link[stats]{fitted}()} \cr
  \code{\link{stat_poly_eq}()}   \tab G \tab "lm", "rlm", "lts", "sma", "ma", "gls",  \emph{others with needed accesors} \cr
  \code{\link{stat_quant_line}()} \tab G \tab "rq", "rqss" \cr
  \code{\link{stat_quant_band}()} \tab G \tab "rq", "rqss" \cr
  \code{\link{stat_quant_eq}()} \tab G \tab "rq", "rqss" \cr
  \code{\link{stat_ma_line}()} \tab G \tab "SMA", "MA", "RMA", "OLS" \cr
  \code{\link{stat_ma_eq}()} \tab G \tab "SMA", "MA", "RMA", "OLS" \cr
  \code{\link{stat_fit_residuals}()} \tab G \tab "lm", "rlm", "lts", "sma", "ma", "gls", "rq", "rqss" \emph{others with method} \code{\link[stats]{residuals}()} \cr
  \code{\link{stat_fit_fitted}()} \tab G \tab "lm", "rlm", "lts", "gls", "rq", "rqss" \emph{others with method} \code{\link[stats]{fitted}()} \cr
  \code{\link{stat_fit_deviations}()} \tab G \tab "lm", "rlm", "lts", "gls", "rq", "rqss" \emph{others with methods} \code{\link[stats]{fitted}()} and \code{\link[stats]{weights}()} \cr
  \code{\link{stat_fit_augment}()} \tab G \tab \emph{any with 'broom' method} \code{\link[broom]{augment}()} \cr
  \code{\link{stat_fit_glance}()} \tab G \tab \emph{any with 'broom' method} \code{\link[broom]{glance}()} \cr
  \code{\link{stat_fit_tidy}()} \tab G \tab \emph{any with 'broom' method} \code{\link[broom]{tidy}()} \cr
  \code{\link{stat_fit_tb}()} \tab P \tab \emph{any with 'broom' method} \code{\link[broom]{tidy}()} \cr
  }

  The table below lists the names for fit methods coded in the statistics
  as given in the table above. The single colon notation is based on parsing
  the name and is available whenever passing the name of the fit method as a
  character string. In a string such as "head:tail" the "head" gives the name
  of the model fit function and the "tail" gives the argument to pass it's
  \code{method} parameter. In some cases the default \code{formula = y ~ x}
  needs to be overridden with an explicit argument.

  \tabular{llll}{
  \strong{Predefined method names} \tab \strong{Model fit methods} \tab \strong{R package} \tab \strong{Object class} \cr
  "lm", "lm:qr" \tab \code{\link[stats]{lm}()} \tab 'stats' \tab "lm" \cr
  "rlm", "rlm:M", "rlm:MM" \tab \code{\link[MASS]{rlm}()} \tab 'MASS' \tab "rlm" ("lm") \cr
  "lts", "ltsReg" \tab \code{\link[robustbase]{ltsReg}()} \tab 'robustbase' \tab "lts" \cr
  "ma", "sma", "sma:SMA", "sma:MA", "sma:OLS" \tab \code{\link[smatr]{sma}()} \tab 'smatr' \tab "ma" or "sma" \cr
  "gls", "gls:REML", "gls:ML" \tab \code{\link[nlme]{gls}()} \tab 'nlme' \tab "gls" \cr
  "rq", "rq:sfn", "rq:sfnc", "rq:lasso" \tab \code{\link[quantreg]{rq}()} \tab 'quantreg' \tab "rq" \cr
  "rqss", "rqss:sfn", "rqss:sfnc", "rqss:lasso" \tab \code{\link[quantreg]{rqss}()} \tab 'quantreg' \tab "rqss" \cr
  "SMA", "MA", "RMA", "OLS" \tab \code{\link[lmodel2]{lmodel2}()} \tab 'lmodel2' \tab \cr
  }
}

\examples{
# generate artificial data
set.seed(98723)
my.data <- data.frame(x = rnorm(100) + (0:99) / 10 - 5,
                      y = rnorm(100) + (0:99) / 10 - 5,
                      group = c("A", "B"))

# using defaults (major axis regression)
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line() +
  stat_ma_eq()

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line() +
  stat_ma_eq(mapping = use_label("eq"))

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line() +
  stat_ma_eq(mapping = use_label("eq"), decreasing = TRUE)

# use_label() can assemble and map a combined label
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line(method = "MA") +
  stat_ma_eq(mapping = use_label("eq", "R2", "P"))

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line(method = "MA") +
  stat_ma_eq(mapping = use_label("R2", "P", "theta", "method"))

# using ranged major axis regression
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line(method = "RMA",
               range.y = "interval",
               range.x = "interval") +
  stat_ma_eq(mapping = use_label("eq", "R2", "P"),
             method = "RMA",
             range.y = "interval",
             range.x = "interval")

# No permutation-based test
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line(method = "MA") +
  stat_ma_eq(mapping = use_label("eq", "R2"),
             method = "MA",
             nperm = 0)

# explicit formula "x explained by y"
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line(formula = x ~ y) +
  stat_ma_eq(formula = x ~ y,
             mapping = use_label("eq", "R2", "P"))

# modifying both variables within aes()
ggplot(my.data, aes(log(x + 10), log(y + 10))) +
  geom_point() +
  stat_poly_line() +
  stat_poly_eq(mapping = use_label("eq"),
               eq.x.rhs = "~~log(x+10)",
               eq.with.lhs = "log(y+10)~~`=`~~")

# grouping
ggplot(my.data, aes(x, y, color = group)) +
  geom_point() +
  stat_ma_line() +
  stat_ma_eq()

# labelling equations
ggplot(my.data,
       aes(x, y,  shape = group, linetype = group, grp.label = group)) +
  geom_point() +
  stat_ma_line(color = "black") +
  stat_ma_eq(mapping = use_label("grp", "eq", "R2")) +
  theme_classic()

# Inspecting the returned data using geom_debug_group()
# This provides a quick way of finding out the names of the variables that
# are available for mapping to aesthetics with after_stat().

gginnards.installed <- requireNamespace("gginnards", quietly = TRUE)

if (gginnards.installed)
  library(gginnards)

# default is output.type = "expression"
if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    geom_point() +
    stat_ma_eq(geom = "debug_group")

\dontrun{
if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    geom_point() +
    stat_ma_eq(mapping = aes(label = after_stat(eq.label)),
               geom = "debug_group",
               output.type = "markdown")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    geom_point() +
    stat_ma_eq(geom = "debug_group", output.type = "text")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    geom_point() +
    stat_ma_eq(geom = "debug_group", output.type = "numeric")
}

}
\seealso{
The major axis regression model is fitted with function
  \code{\link[lmodel2]{lmodel2}}, please consult its documentation. Statistic
  \code{stat_ma_eq()} can return different ready formatted labels depending
  on the argument passed to \code{output.type}.

Other ggplot statistics for major axis regression: 
\code{\link{stat_ma_line}()}
}
\concept{ggplot statistics for major axis regression}
\section{Aesthetics}{

\code{stat_ma_eq()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → inferred \cr
 • \tab \code{grp.label} \tab   \cr
 • \tab \code{hjust} \tab → \code{"inward"} \cr
 • \tab \code{label} \tab → \code{after_stat(rr.label)} \cr
 • \tab \code{npcx} \tab → \code{after_stat(npcx)} \cr
 • \tab \code{npcy} \tab → \code{after_stat(npcy)} \cr
 • \tab \code{vjust} \tab → \code{"inward"} \cr
}

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}
