% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-ma-line.R
\name{stat_ma_line}
\alias{stat_ma_line}
\title{Predicted line from major axis linear fit}
\usage{
stat_ma_line(
  mapping = NULL,
  data = NULL,
  geom = "smooth",
  position = "identity",
  ...,
  orientation = NA,
  method = "lmodel2:MA",
  method.args = list(),
  n.min = 2L,
  formula = NULL,
  range.y = NULL,
  range.x = NULL,
  se = TRUE,
  fit.seed = NA,
  fm.values = FALSE,
  n = 80,
  nperm = 99,
  fullrange = FALSE,
  level = 0.95,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}()}. Only needs to be set at the layer level if you
are overriding the plot defaults.}

\item{data}{A layer specific dataset, only needed if you want to override the
plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points on this
layer.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}

\item{orientation}{character Either "x" or "y" controlling the default for
\code{formula}. The letter indicates the aesthetic considered the
explanatory variable in the model fit.}

\item{method}{function or character If character, "MA", "SMA" , "RMA" or
"OLS", alternatively "lmodel2" or the name of a model fit function are
accepted, possibly followed by the fit function's \code{method} argument
separated by a colon (e.g. \code{"lmodel2:MA"}). If a function different to
\code{lmodel2()}, it must accept arguments named \code{formula},
\code{data}, \code{range.y}, \code{range.x} and \code{nperm} and return a
model fit object of class \code{lmodel2}.}

\item{method.args}{named list with additional arguments. Not \code{data}
or \code{weights} which are always passed through aesthetic mappings.}

\item{n.min}{integer Minimum number of distinct values in the explanatory
variable (on the rhs of formula) for fitting to the attempted.}

\item{formula}{a formula object. Using aesthetic names \code{x} and \code{y}
instead of original variable names.}

\item{range.y, range.x}{character Pass "relative" or "interval" if method
"RMA" is to be computed.}

\item{se}{logical Return confidence interval around smooth? (`TRUE` by
default, see `level` to control.)}

\item{fit.seed}{RNG seed argument passed to
\code{\link[base:Random]{set.seed}()}. Defaults to \code{NA}, indicating
that \code{set.seed()} should not be called.}

\item{fm.values}{logical Add metadata and parameter estimates extracted from
the fitted model object; \code{FALSE} by default.}

\item{n}{Number of points at which to predict with the fitted model.}

\item{nperm}{integer Number of permutation used to estimate significance.}

\item{fullrange}{Should the fit span the full range of the plot, or just the
range of the data group used in each fit?}

\item{level}{Level of confidence interval to use (only 0.95 currently).}

\item{na.rm}{a logical indicating whether NA values should be stripped before
the computation proceeds.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}
}
\value{
The value returned by the statistic is a data frame, that will have
  \code{n} rows of predicted values and their confidence limits. Optionally
  it will also include additional values related to the model fit.
}
\description{
Predicted values and a confidence band are computed and, by default, plotted.
\code{stat_ma_line()} behaves similarly to \code{\link{stat_poly_line}()}
except for fitting the model with \code{lmodel2::lmodel2()} with \code{"MA"}
as default for \code{method}.
}
\details{
Statistics \code{stat_ma_line()} and \code{stat_ma_eq} fit major
  axis (\code{"MA"}) and other model II regressions with function
  \code{\link[lmodel2]{lmodel2}}. They support linear major axis (MA),
  standard major axis (SMA) and ranged major axis (RMA) regression.

  \code{stat_ma_line()} adds the predicted line and confidence band based on
  the uncertainty of the slope estimate. considering). \code{stat_ma_line()}
  can add annotations with the fitted model equation and other parameter
  estimates.

  Model II regression is called for when both \code{x} and \code{y} are
  subject to random variation and the intention is not to predict \code{y}
  from \code{x} by means of the model but rather to study the relationship
  between two independent variables. A frequent case in biology are
  allometric relationships among body parts.

  As the fitted line is the same whether \code{x} or \code{y} is on the rhs
  of the model equation, \code{orientation} even if accepted does not have an
  effect on the fitted line.

  The minimum number of observations with distinct values can be set through
  parameter \code{n.min}. The default \code{n.min = 2L} is the smallest
  possible value. However, model fits with very few observations are of
  little interest and using a larger number for \code{n.min} than the default
  is wise. As model fitting functions could depend on the RNG,
  \code{fit.seed} if different to \code{NA} is used as argument in a call to
  \code{\link[base:Random]{set.seed}()} immediately ahead of model fitting.

  In \code{\link[lmodel2]{lmodel2}()} MA, SMA and OLS fits always computed
  while RMA requires a numeric argument to at least one of \code{range.y}
  or \code{range.x}. The statistics extract estimates for one of the methods
  based on the argument for \code{method}.

  Package 'lmodel2' implements a model fit function and fitted model object
  that differ from the usual approach of R. Thus, their use was implemented
  as a separate pair of statistics. Use of package 'smatr' is implemented in
  \code{\link{stat_poly_line}()} and \code{\link{stat_poly_eq}()}.
}
\section{Computed variables}{
 `stat_ma_line()` provides the following
  variables, some of which depend on the orientation:

  \describe{ \item{y \strong{or} x}{predicted value}
  \item{ymin \strong{or} xmin}{lower pointwise confidence interval around the mean}
  \item{ymax \strong{or} xmax}{upper pointwise confidence interval around the mean}
  \item{se}{standard error}
  }

  If \code{fm.values = TRUE} is passed then columns based on the summary of
  the model fit are added, with the same value in each row within a group.
  This is wasteful and disabled by default, but provides a simple and robust
  approach to achieve effects like colouring or hiding of the model fit line
  based on P-values, r-squared or the number of observations.
}

\section{Model formula and model fitting}{

  A ggplot statistic receives as \code{data} a data frame that is not the one
  passed as argument by the user, but instead a data frame with the variables
  mapped to aesthetics. In \code{stat_poly_eq()} the compute function is
  applied by group, each call "seeing" the subset of \code{data} for an
  individual group. As supported models are for regression lines,
  variables mapped to \code{x} and \code{y} should both be continuous, i.e.,
  numeric or date time and model formulas defined using \code{x} and \code{y}
  as variable names.

  The interpretation of the argument passed to \code{formula} is enhanced
  compared to \code{stat_smooth()}. Formulas with \code{x} as explanatory
  variable work as in \code{stat_smooth()} but formulas with \code{y} as
  explanatory variable are also accepted. \code{orientation} is set
  automatically based on which explanatory variable appears in the formula.
  Spline-based smoothers are only partially supported.
}

\section{Model fit methods supported}{
 Several model fit functions are
  supported explicitly (see tables), and some of their differences smoothed
  out. Compatibility is checked late, based on the class of the returned
  fitted model object. This makes it possible to use wrapper functions that
  do model selection or other adjustments to the fit procedure on a per panel
  or per group basis. Moreover, if the value returned as model fit object is
  \code{NULL}, no layer is added to the plot on a per group within panel
  basis.

  In the case of fitted model objects of classes not explicitly supported an
  attempt is made to find the usual accessors and/or fitted object members,
  and if found, either complete or partial support is frequently achieved.
  In this case a message is issued encouraging users to check the validity
  of the values extracted.

  The argument to parameter \code{method} can be either the name of a
  function object, possibly using double colon notation in case its package
  is not attached, or a character string matching the function name for
  functions in the search path. This approach makes it possible to support
  model fit functions that are not dependencies of 'ggpmisc'. Either by
  attaching the package where the function is defined and passing it by name
  or as string, or using double colon notation when passing the name of the
  function.

  User-defined functions can be passed as argument to parameter
  \code{method} as long as they have parameters \code{formula}, \code{data}
  \code{subset} and possibly \code{weights}. Additional arguments can be
  passed to any method as a named list as argument to parameter
  \code{method.args}. As in \code{\link[ggplot2:geom_smooth]{stat_smooth}()}
  prior \code{weights} are passed to the model fit functions' \code{weights}
  (plural!) parameter by mapping a numeric variable to plot aesthetic
  \code{weight} (singular!).

  Tables 1 lists natively supported model fit functions, with the
  caveat that only some 'broom' methods' specializations have been actually
  tested with statistics from 'ggpmisc'. In addition, the statistics based
  on 'broom' methods require the user to tailor their behaviour by passing
  additional arguments in the call and occasionally some detective work to
  find out the names of variables in the returned data frame.

  \strong{Table 1.} Model fit methods supported by the different statistics
  available in package 'ggpmisc'. Column \eqn{f} indicates whether
  computations are done by group (G) or by plot panel (P).
  \tabular{lcl}{
  \strong{Statistic} \tab \eqn{f} \tab \strong{Supported model fit methods} \cr
  \code{\link{stat_poly_line}()} \tab G \tab "lm", "rlm", "lts", "sma", "ma", "gls", \emph{others with methods} \code{\link[stats]{predict}()} or \code{\link[stats]{fitted}()} \cr
  \code{\link{stat_poly_eq}()}   \tab G \tab "lm", "rlm", "lts", "sma", "ma", "gls",  \emph{others with needed accesors} \cr
  \code{\link{stat_quant_line}()} \tab G \tab "rq", "rqss" \cr
  \code{\link{stat_quant_band}()} \tab G \tab "rq", "rqss" \cr
  \code{\link{stat_quant_eq}()} \tab G \tab "rq", "rqss" \cr
  \code{\link{stat_ma_line}()} \tab G \tab "SMA", "MA", "RMA", "OLS" \cr
  \code{\link{stat_ma_eq}()} \tab G \tab "SMA", "MA", "RMA", "OLS" \cr
  \code{\link{stat_fit_residuals}()} \tab G \tab "lm", "rlm", "lts", "sma", "ma", "gls", "rq", "rqss" \emph{others with method} \code{\link[stats]{residuals}()} \cr
  \code{\link{stat_fit_fitted}()} \tab G \tab "lm", "rlm", "lts", "gls", "rq", "rqss" \emph{others with method} \code{\link[stats]{fitted}()} \cr
  \code{\link{stat_fit_deviations}()} \tab G \tab "lm", "rlm", "lts", "gls", "rq", "rqss" \emph{others with methods} \code{\link[stats]{fitted}()} and \code{\link[stats]{weights}()} \cr
  \code{\link{stat_fit_augment}()} \tab G \tab \emph{any with 'broom' method} \code{\link[broom]{augment}()} \cr
  \code{\link{stat_fit_glance}()} \tab G \tab \emph{any with 'broom' method} \code{\link[broom]{glance}()} \cr
  \code{\link{stat_fit_tidy}()} \tab G \tab \emph{any with 'broom' method} \code{\link[broom]{tidy}()} \cr
  \code{\link{stat_fit_tb}()} \tab P \tab \emph{any with 'broom' method} \code{\link[broom]{tidy}()} \cr
  }

  The single colon notation is based on parsing
  the name and is available whenever passing the name of the fit method as a
  character string. In a string such as "head:tail" the "head" gives the name
  of the model fit function and the "tail" gives the argument to pass it's
  \code{method} parameter. This is only a convenience, as \code{method.args}
  can be also used. In some methods, i.e., splines, the default
  \code{formula = y ~ x} needs to be overridden with an explicit argument.

  Table 2 lists the correspondence of pre-defined \emph{method names}
  to model fit method functions. As mentioned above, these are only
  a subset of the model fit methods that are expected to work. When using
  these names there is no need for users to attach additional packages but
  the packages must be available (installed).

  \strong{Table 2.} Available predefined method names, the model fit functions
  they call, the packages where the functions reside, the class of the
  returned fitted model object and the arguments that can be
  passed to their \code{method} parameter using single colon notation.
  \tabular{llll}{
  \strong{Predefined method names} \tab \strong{Model fit methods} \tab \strong{R package} \tab \strong{Object class} \cr
  "lm", "lm:qr" \tab \code{\link[stats]{lm}()} \tab 'stats' \tab "lm" \cr
  "rlm", "rlm:M", "rlm:MM" \tab \code{\link[MASS]{rlm}()} \tab 'MASS' \tab "rlm" ("lm") \cr
  "lts", "ltsReg" \tab \code{\link[robustbase]{ltsReg}()} \tab 'robustbase' \tab "lts" \cr
  "ma", "sma", "sma:SMA", "sma:MA", "sma:OLS" \tab \code{\link[smatr]{sma}()} \tab 'smatr' \tab "ma" or "sma" \cr
  "gls", "gls:REML", "gls:ML" \tab \code{\link[nlme]{gls}()} \tab 'nlme' \tab "gls" \cr
  "rq", "rq:sfn", "rq:sfnc", "rq:lasso" \tab \code{\link[quantreg]{rq}()} \tab 'quantreg' \tab "rq" \cr
  "rqss", "rqss:sfn", "rqss:sfnc", "rqss:lasso" \tab \code{\link[quantreg]{rqss}()} \tab 'quantreg' \tab "rqss" \cr
  "SMA", "MA", "RMA", "OLS" \tab \code{\link[lmodel2]{lmodel2}()} \tab 'lmodel2' \tab ("list") \cr
  }
}

\examples{
# generate artificial data
set.seed(98723)
my.data <- data.frame(x = rnorm(100) + (0:99) / 10 - 5,
                      y = rnorm(100) + (0:99) / 10 - 5,
                      group = c("A", "B"))

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line()

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line(method = "MA")

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line(method = "SMA")

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line(method = "RMA",
               range.y = "interval", range.x = "interval")

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line(method = "OLS")

# plot line to the ends of range of data (the default)
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line(fullrange = FALSE) +
  expand_limits(x = c(-10, 10), y = c(-10, 10))

# plot line to the limits of the scales
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line(fullrange = TRUE) +
  expand_limits(x = c(-10, 10), y = c(-10, 10))

# plot line to the limits of the scales
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line(orientation = "y", fullrange = TRUE) +
  expand_limits(x = c(-10, 10), y = c(-10, 10))

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line(formula = x ~ y)

# Smooths are automatically fit to each group (defined by categorical
# aesthetics or the group aesthetic) and for each facet.

ggplot(my.data, aes(x, y, colour = group)) +
  geom_point() +
  stat_ma_line()

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_ma_line() +
  facet_wrap(~group)

# Inspecting the returned data using geom_debug_group()
gginnards.installed <- requireNamespace("gginnards", quietly = TRUE)

if (gginnards.installed)
  library(gginnards)

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    stat_ma_line(geom = "debug_group")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    stat_ma_line(geom = "debug_group", fm.values = TRUE)

}
\seealso{
Other 'ggpmisc' statistics for model fits:
\code{\link[=stat_distrmix_eq]{stat_distrmix_eq()}},
\code{\link[=stat_distrmix_line]{stat_distrmix_line()}},
\code{\link[=stat_fit_deviations]{stat_fit_deviations()}},
\code{\link[=stat_fit_glance]{stat_fit_glance()}},
\code{\link[=stat_fit_residuals]{stat_fit_residuals()}},
\code{\link[=stat_fit_tb]{stat_fit_tb()}},
\code{\link[=stat_fit_tidy]{stat_fit_tidy()}},
\code{\link[=stat_ma_eq]{stat_ma_eq()}},
\code{\link[=stat_poly_eq]{stat_poly_eq()}},
\code{\link[=stat_poly_line]{stat_poly_line()}},
\code{\link[=stat_quant_band]{stat_quant_band()}},
\code{\link[=stat_quant_eq]{stat_quant_eq()}},
\code{\link[=stat_quant_line]{stat_quant_line()}}
}
\concept{'ggpmisc' statistics for model fits}
\section{Aesthetics}{

\code{stat_ma_line()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → inferred \cr
}

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}
