% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-multcomp.R
\name{stat_multcomp}
\alias{stat_multcomp}
\title{Labels for pairwise multiple comparisons}
\usage{
stat_multcomp(
  mapping = NULL,
  data = NULL,
  geom = NULL,
  position = "identity",
  ...,
  orientation = NA,
  formula = y ~ factor(x),
  method = "lm",
  method.args = list(),
  contrasts = "Tukey",
  p.adjust.method = NULL,
  fit.seed = NA,
  fm.cutoff.p.value = 1,
  mc.cutoff.p.value = 1,
  mc.critical.p.value = 0.05,
  small.p = getOption("ggpmisc.small.p", default = FALSE),
  adj.method.tag = 4,
  p.digits = NULL,
  label.type = "bars",
  bjust = 0.5,
  label.y = NULL,
  vstep = NULL,
  output.type = NULL,
  na.rm = FALSE,
  parse = NULL,
  show.legend = FALSE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}()}. Only needs to be set at the layer level if you
are overriding the plot defaults.}

\item{data}{A layer specific dataset, only needed if you want to override the
plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points on this
layer.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}

\item{orientation}{character The orientation of the layer. The default
(\code{NA}) automatically determines the orientation from the aesthetic
mapping. In the rare event that this fails it can be given explicitly by
setting orientation to either "x" or "y".}

\item{formula}{a formula object. Using aesthetic names \code{x} and \code{y}
instead of original variable names. The rhs must include a call to
\code{factor()} even if the variable mapped to the \code{x} aesthetic is
a factor! \emph{In both flipped and not flipped plots, \code{x} should be
in the rhs}.}

\item{method}{function or character If character, "lm", "rlm", "lmrob",
"lts", "gls", "ma", "sma", "segreg", "rq" or the name of a model fit
function are accepted, possibly followed by the fit function's
\code{method} argument separated by a colon (e.g. \code{"rlm:M"}). If a
function is different to \code{lm()}, \code{rlm()}, \code{ltsReg()},
\code{gls()}, \code{ma}, \code{sma}, it must have formal parameters named
\code{formula}, \code{data}, and \code{weights}. See Details.}

\item{method.args}{named list with additional arguments. Not \code{data}
or \code{weights} which are always passed through aesthetic mappings.}

\item{contrasts}{character vector of length one or a numeric matrix. If
character, one of "Tukey" or "Dunnet". If a matrix, one column per level
of the factor mapped to \code{x} and one row per \strong{pairwise}
contrast.}

\item{p.adjust.method}{character As the argument for parameter \code{type} of
function \code{adjusted()} passed as argument to parameter \code{test} of
\code{\link[multcomp]{summary.glht}}. Accepted values are "single-step",
"Shaffer", "Westfall", "free", "holm", "hochberg", "hommel", "bonferroni",
"BH", "BY", "fdr", "none".}

\item{fit.seed}{RNG seed argument passed to
\code{\link[base:Random]{set.seed}()}. Defaults to \code{NA}, indicating
that \code{set.seed()} should not be called.}

\item{fm.cutoff.p.value}{numeric [0..1] The \emph{P}-value for the main
effect of factor \code{x} in the ANOVA test for the fitted model above
which no pairwise comparisons are computed or labels generated. Be aware
that recent literature tends to recommend to consider which testing
approach is relevant to the problem at hand instead of requiring the
significance of the main effect before applying multiple comparisons'
tests. The default value is 1, imposing no restrictions.}

\item{mc.cutoff.p.value}{numeric [0..1] The \emph{P}-value for the individual
contrasts above which no labelled bars are generated. Default is 1,
labelling all pairwise contrasts tested.}

\item{mc.critical.p.value}{numeric The critical \emph{P}-value used for tests
when encoded as letters.}

\item{small.p}{logical Flag to switch use of lower case p for \emph{p}-value.}

\item{adj.method.tag}{numeric, character or function If \code{numeric}, the
length in characters of the abbreviation of the method used to adjust
\emph{p}-values. A value of zero, adds no label and a negative value uses
as starting point for the abbreviation the word "adjusted". If
\code{character} its value is used as subscript. If a \code{function}, the
value used is the value returned by the function when passed
\code{p.adjust.method} as its only argument.}

\item{p.digits}{integer Number of digits after the decimal point to use for
\emph{P}-value in labels. If \code{Inf}, use exponential notation with
three decimal places.}

\item{label.type}{character One of "bars", "letters" or "LETTERS", selects
how the results of the multiple comparisons are displayed. Only "bars" can
be used together with \code{contrasts = "Dunnet"}.}

\item{bjust}{numeric usually with range 0..1, the justification of the labels
relative to the bars.}

\item{label.y}{numeric with range 0..1 "normalized parent
coordinates" (npc units) or character if using \code{geom_text_npc()} or
\code{geom_label_npc()}. If using \code{geom_text()} or \code{geom_label()}
numeric in native data units. \emph{In flipped plots, it refers to \code{x}
rather than \code{y} position}.}

\item{vstep}{numeric in npc units, the vertical displacement step-size used
between labels for different contrasts when
\code{label.type = "bars"}. \emph{In flipped plots, it increases the
\code{x} rather than \code{y} position}.}

\item{output.type}{character One of "expression", "text", "markdown",
"marquee", "latex", "latex.eqn", "latex.deqn" or "numeric".}

\item{na.rm}{a logical indicating whether NA values should be stripped before
the computation proceeds.}

\item{parse}{logical Passed to the geom. If \code{TRUE}, the labels will be
parsed into expressions and displayed as described in
\code{\link[grDevices]{plotmath}}. Default is \code{TRUE} if
\code{output.type = "expression"} and \code{FALSE} otherwise.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}
}
\value{
A data frame with one row per comparison for \code{label.type =
  "bars"}, or a data frame with one row per factor level for
  \code{label.type = "letters"} and \code{label.type = "LETTERS"}.
  Variables (= columns) as described under \strong{Computed variables}.
}
\description{
\code{stat_multcomp} fits a linear model by default with \code{stats::lm()}
but alternatively using other model fit functions. The model is passed to
function \code{glht()} from package 'multcomp' to fit Tukey, Dunnet or other
\strong{pairwise} contrasts and generates labels based on adjusted
\emph{P}-values.
}
\details{
This statistic can be used to automatically annotate a plot with
  \emph{P}-values for \strong{pairwise} multiple comparison tests, based on
  Tukey contrasts (all pairwise), Dunnet contrasts (other levels against the
  first one) or a subset of all possible pairwise contrasts. See Meier (2022,
  Chapter 3) for an accessible explanation of multiple comparisons and
  contrasts with package 'multcomp', of which \code{stat_multcomp()} is
  mostly a wrapper. It supports most model fit methods supported by function
  \code{\link[multcomp]{glht}()}. The requirement is that the necessary query
  functions for the class of the fitted model object are available. If not
  supported, an error will be triggered by \code{glht()}. The computations
  are done in two stages. First a model is fitted to all the data in a plot
  panel, followed by the pairwise contrasts. A threshold \eqn{P}-value can
  the set for the main effect of "treatments" as a condition to apply the
  pairwise tests.

  One variable mapped to either \emph{x} or \emph{y} aesthetic must be a
  factor, and the other variable a continuous one. Mapping a factor creates
  the required grouping and determines the default orientation, i.e.,
  similarly as in \code{\link[ggplot2:geom_boxplot]{stat_boxplot}()}.

  Two ways of displaying the outcomes are implemented, and are selected by
  `"bars"`, `"letters"` or `"LETTERS"` as argument to parameter `label.type`.
  `"letters"` and `"LETTERS"` can be used only with Tukey contrasts, as
  otherwise the encoding is ambiguous. As too many bars clutter a plot, the
  maximum number of factor levels supported for `"bars"` together with Tukey
  contrasts is five, while together with Dunnet contrasts or contrasts
  defined by a numeric matrix, no limit is imposed. \code{label.y} determines
  the location of the \emph{letters} labels or the position of the first
  \emph{pairwise label}. If \code{label.y} is numeric and has length > 1,
  values are taken as positions for successive bar labels, and recycled if
  the vector is too short, ignoring \code{vstep}.

  \code{stat_multcomp()} by default generates character labels ready to be
  parsed as R expressions but LaTeX (use TikZ device), markdown (use package
  'ggtext') and plain text are also supported, as well as numeric values for
  user-generated text labels. The value of \code{parse} is set automatically
  based on \code{output.type}, but if you assemble labels that need parsing
  from \code{numeric} output, the default needs to be overridden. This
  statistic only generates annotation labels and segments connecting the
  compared factor levels, or letter labels that discriminate significantly
  different groups.

  \emph{Currently, contrasts that involve more than two levels of a factor,
  such as the average of two treatment levels against a control level are not
  supported. Similarly, flipping is not yet functional with bar labels
  because of the lack of a working geom.}
}
\note{
R option \code{OutDec} is obeyed based on its value at the time the plot
  is rendered, i.e., displayed or printed. Set \code{options(OutDec = ",")}
  for languages like Spanish or French.
}
\section{Computed variables}{


  Computed variables and their names vary depending on the \code{method} used
  to fit a model or the \code{output.type} in use. They can also depend for a
  given \code{method} on the \code{contrasts}.

If \code{output.type = "numeric"} and
\code{label.type = "bars"} the returned tibble contains
columns listed below. In all cases if the model fit function used does not return a value,
the label is set to \code{character(0L)} and the numeric value to \code{NA}.
\describe{
  \item{x,xmin,xmax}{x position, numeric.}
  \item{y,ymin,ymax}{y position, numeric.}
  \item{coefficients}{Delta estimate from pairwise contrasts, numeric.}
  \item{contrasts}{Contrasts as two levels' ordinal "numbers" separated by a dash, character.}
  \item{tstat}{\emph{t}-statistic estimates for the pairwise contrasts, numeric.}
  \item{p.value}{\emph{P}-value for the pairwise contrasts.}
  \item{p.signif}{logical for the pairwise contrasts significance.}
  \item{fm.method}{Set according \code{method} used.}
  \item{fm.class}{Most derived class of the fitted model object.}
  \item{fm.formula}{Formula extracted from the fitted model object if available, or the formula argument.}
  \item{fm.formula.chr}{Formula extracted from the fitted model object if available, or the formula argument, formatted as character.}
  \item{mc.adjusted}{The method used to adjust the \emph{P}-values.}
  \item{mc.contrast}{The type of contrast used for multiple comparisons.}
  \item{n}{The total number of observations or rows in data.}
  \item{default.label}{text label, always included, but possibly NA.}
  }

If output.type is not \code{"numeric"} the returned data frame includes in
addition the following labels:

\describe{
  \item{stars.label}{\emph{P}-value for the pairwise contrasts encoded as "starts", character.}
  \item{p.value.label}{\emph{P}-value for the pairwise contrasts, character.}
  \item{delta.label}{The coefficient or estimate for the difference between compared pairs of levels.}
  \item{t.value.label}{\emph{t}-statistic estimates for the pairwise contrasts, character.}
  }

If \code{output.type = "numeric"} and \code{label.type = "letters"} or
\code{label.type = "LETTERS"} the returned tibble contains columns listed
below.

\describe{
  \item{x,xmin,xmax}{x position, numeric.}
  \item{y,ymin,ymax}{y position, numeric.}
  \item{critical.p.value}{\emph{P}-value used in pairwise tests, numeric.}
  \item{fm.method}{Set according \code{method} used.}
  \item{fm.class}{Most derived class of the fitted model object.}
  \item{fm.formula}{Formula extracted from the fitted model object if available, or the formula argument.}
  \item{fm.formula.chr}{Formula extracted from the fitted model object if available, or the formula argument, formatted as character.}
  \item{mc.adjusted}{The method used to adjust the \emph{P}-values.}
  \item{mc.contrast}{The type of contrast used for multiple comparisons.}
  \item{n}{The total number of observations or rows in data.}
  \item{default.label}{text label, always included, but possibly NA.}
  }

If output.type is not \code{"numeric"} the returned data frame includes in
addition the following labels:

\describe{
  \item{letters.label}{Letters that distinguish levels based on significance from multiple comparisons test.}
  }
}

\section{Alternatives}{
 \code{stat_signif()} in package 'ggsignif' is
  an earlier and independent implementation of pairwise tests.
}

\section{Output types}{
 The formatting of character strings to be displayed
in plots are marked as mathematical equations. Depending on the geom used,
the mark-up needs to be encoded differently, or in some cases mark-up not
applied.
\describe{
  \item{\code{"expression"}}{The labels are encoded as character strings to be parsed into R's plotmath expressions.}
  \item{\code{"LaTeX", "TeX", "tikz", "latex"}}{The labels are encoded as 'LaTeX' maths equations, without the "fences" for switching in math mode.}
  \item{\code{"latex.eqn"}}{Same as \code{"latex"} but enclosed in single \code{$}, i.e., as in-line maths.}
  \item{\code{"latex.deqn"}}{Same as \code{"latex"} but enclosed in double \code{$$}, i.e., as display maths.}
  \item{\code{"markdown"}}{The labels are encoded as character strings using markdown syntax, with some embedded HTML.}
  \item{\code{"marquee"}}{The labels are encoded as character strings using markdown syntax, with 'marquee' supported spans.}
  \item{\code{"text"}}{The labels are plain ASCII character strings.}
  \item{\code{"numeric"}}{No labels are generated. This value is accepted by the statistics, but not by the label formatting functions.}
  \item{\code{NULL}}{The value used depends on the argument passed to \code{geom}.}}

If \code{geom = "latex"} (package 'xdvir') the output type used is
\code{"latex.eqn"}. If \code{geom = "richtext"} (package 'ggtext') or
\code{geom = "textbox"} (package 'ggtext') the output type used is
\code{"markdown"}. If \code{geom = "marquee"} (package 'marquee') the output
type used is \code{"marquee"}. For all other values of \code{geom} the default
is \code{"expression"}. Invalid values as argument trigger an error.
}

\section{Which variables are available for mapping?}{

  Computed variables and their names can vary depending on the \code{method}
  used to fit a model or the \code{output.type} in use. They can also depend
  for a given \code{method} on other arguments passed when fitting a model or
  extracting estimates and other computed values. In many cases, when values
  are not available, the variables are filled with \code{NA} values.

  In the statistics returning formatted strings for use as annotations, a
  message is issued by default in interactive R sessions, listing the short
  names for available formatted labels as recognized by functions
  \code{\link{use_label}()} and \code{\link{f_use_label}()}, except when
  \code{output.type = "numeric"} is passed, in which case the names of all
  variables accessible by \code{after_stat()} within a call to \code{aes()}
  are listed. This default (\code{"nicknames"}) can be changed by setting R option
  \code{"ggpmisc.stat.vars.message"} to one of \code{"names"},
  \code{"colnames"} or \code{"none"}.

  In the statistics that plot a prediction or more generally mainly return
  numeric variables, a message is issued by default in interactive R
  sessions, listing the names of all variables accessible by
  \code{after_stat()} within a call to \code{aes()} with at least some
  non-missing values. This default (\code{"colnames"}) can be changed by
  setting R option \code{"ggpmisc.stat.vars.message"} to \code{"none"}.
}

\examples{

p1 <- ggplot(mpg, aes(factor(cyl), hwy)) +
  geom_boxplot(width = 0.33)

## labeleld bars

p1 +
  stat_multcomp()

p1 +
  stat_multcomp(adj.method.tag = FALSE)

# test against a control, with first level being the control
# change order of factor levels in data to set the control group
p1 +
  stat_multcomp(contrasts = "Dunnet")

# arbitrary pairwise contrasts, in arbitrary order
p1 +
  stat_multcomp(contrasts = rbind(c(0, 0, -1, 1),
                                  c(0, -1, 1, 0),
                                  c(-1, 1, 0, 0)))

# different methods to adjust the contrasts
p1 +
  stat_multcomp(p.adjust.method = "holm")

p1 +
  stat_multcomp(p.adjust.method = "fdr")

# position of contrasts' bars (based on scale limits)
p1 +
  stat_multcomp(label.y = "bottom")

p1 +
  stat_multcomp(label.y = 11)

# use different labels: difference and P-value from hypothesis tests
p1 +
  stat_multcomp(use_label("Delta", "P"),
                size = 2.75)

# control smallest P-value displayed and number of digits
p1 +
  stat_multcomp(p.digits = 4)

# highlight significant differences
p1 +
  stat_multcomp(aes(alpha = after_stat(p.signif)),
                alpha.target = c("text", "segment", "box.line")) +
  scale_alpha_manual(values =  c(0.33, 1))

# label only significant differences
# but test and correct for all pairwise contrasts!
p1 +
  stat_multcomp(mc.cutoff.p.value = 0.01)

## letters as labels for test results

p1 +
  stat_multcomp(label.type = "letters")

# stricter critical p-value than default used for test
p1 +
  stat_multcomp(label.type = "letters",
                mc.critical.p.value = 0.01)

# justification of label in bar, and text anchor point
# for a compact display

p1 +
  stat_multcomp(geom = "text_pairwise",
                arrow = grid::arrow(ends = "both",
                                    length = unit(1.5, "mm")),
                bjust = 0,
                hjust = 1.05,
                vjust = 0.5,
                vstep = 0.05,
                label.y = 14,
                adj.method.tag = 0) +
  expand_limits(y = 0)

## Flipping is supported
p2 <- ggplot(mpg, aes(hwy, factor(cyl))) +
  geom_boxplot(width = 0.33)

p2 +
  stat_multcomp(label.type = "letters") +
  scale_y_discrete(expand = expansion(add = c(1.2, 0.5)))

p2 +
  stat_multcomp(label.type = "bars",
                geom = "text_pairwise",
                size = 3)

}
\references{
Meier, Lukas (2022) \emph{ANOVA and Mixed Models: A Short Introduction
Using R}. Chapter 3 Contrasts and Multiple Testing. The R Series. Boca Raton:
Chapman and Hall/CRC. ISBN: 9780367704209, \doi{10.1201/9781003146216}.
}
\seealso{
This statistic uses the implementation of Tests of General Linear
  Hypotheses in function \code{\link[multcomp]{glht}}. See
  \code{\link[multcomp]{summary.glht}} and \code{\link[stats]{p.adjust}}
  for the supported tests and the references therein for the theory
  behind them.

  Please, see the articles at
  \href{https://docs.r4photobiology.info/ggpmisc/}{online-only documentation}
  for additional use examples and guidance.
}
\concept{statistics for multiple comparison annotations}
\section{Aesthetics}{

\code{stat_multcomp()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → inferred \cr
 • \tab \code{hjust} \tab → \code{after_stat(just)} \cr
 • \tab \code{label} \tab → \code{after_stat(default.label)} \cr
 • \tab \code{\link[ggplot2:aes_linetype_size_shape]{size}} \tab → \code{2.5} \cr
 • \tab \code{weight} \tab → \code{1} \cr
 • \tab \code{\link[ggplot2:aes_position]{xmax}} \tab → \code{after_stat(ifelse(flipped_aes, NA_real_, xmax))} \cr
 • \tab \code{\link[ggplot2:aes_position]{xmin}} \tab → \code{after_stat(ifelse(flipped_aes, NA_real_, xmin))} \cr
 • \tab \code{\link[ggplot2:aes_position]{ymax}} \tab → \code{after_stat(ifelse(flipped_aes, ymax, NA_real_))} \cr
 • \tab \code{\link[ggplot2:aes_position]{ymin}} \tab → \code{after_stat(ifelse(flipped_aes, ymin, NA_real_))} \cr
}

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}
