% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-peaks.R
\name{stat_peaks}
\alias{stat_peaks}
\alias{stat_valleys}
\title{Find peaks and valleys.}
\usage{
stat_peaks(mapping = NULL, data = NULL, geom = "point", span = 5,
  ignore_threshold = 0, strict = FALSE, label.fmt = "\%.3g",
  x.label.fmt = label.fmt, y.label.fmt = label.fmt, position = "identity",
  na.rm = FALSE, show.legend = FALSE, inherit.aes = TRUE, ...)

stat_valleys(mapping = NULL, data = NULL, geom = "point", span = 5,
  ignore_threshold = 0, strict = FALSE, label.fmt = "\%.3g",
  x.label.fmt = label.fmt, y.label.fmt = label.fmt, position = "identity",
  na.rm = FALSE, show.legend = FALSE, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{span}{a peak is defined as an element in a sequence which is greater
than all other elements within a window of width span centered at that
element. The default value is 5, meaning that a peak is bigger than two
consequtive neighbors on each side. Default: 5.}

\item{ignore_threshold}{numeric value between 0.0 and 1.0 indicating the size
threshold below which peaks will be ignored.}

\item{strict}{logical flag: if TRUE, an element must be strictly greater than
all other values in its window to be considered a peak. Default: FALSE.}

\item{label.fmt}{character  string giving a format definition for converting
values into character strings by means of function \code{\link{sprintf}}.}

\item{x.label.fmt}{character  string giving a format definition for converting
$x$-values into character strings by means of function \code{\link{sprintf}}.}

\item{y.label.fmt}{character  string giving a format definition for converting
$y$-values into character strings by means of function \code{\link{sprintf}}.}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{na.rm}{a logical value indicating whether NA values should be
stripped before the computation proceeds.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2]{borders}}.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This can
include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}
}
\description{
\code{stat_peaks} finds at which x positions local maxima are located. If
you want find local minima, you can use \code{stat_valleys} instead.
}
\section{Computed variables}{

\describe{
  \item{x.label}{x-value at the peak}
}
}
\seealso{
\code{\link[photobiology]{find_peaks}}, which is used internally.

Other peaks.and.valleys.functions: \code{\link{find_peaks}}
}

