% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-eq-lm.R
\docType{data}
\name{stat_poly_eq}
\alias{StatPolyEq}
\alias{stat_poly_eq}
\title{Add a curve from a fitted linear model and a label to a plot.}
\usage{
stat_poly_eq(mapping = NULL, data = NULL, geom = "text", formula = NULL,
  position = "identity", na.rm = FALSE, show.legend = FALSE,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_string}}. Only needs
to be set at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{formula}{a formula object}

\item{position}{The position adjustment to use for overlapping points on this
layer}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}
}
\description{
\code{stat_poly_eq} fits a polynomial and generates a label with an equation
and/or coefficient of determination (R^2).
}
\section{Computed variables}{

  \describe{ \item{x}{x position for left edge}
  \item{y}{y position near upper edge}
  \item{eq.label}{equation for the
  fitted polynomial as a character string to be parsed}
  \item{rr.label}{\eqn{R^2} of the fitted model as a character string to be parsed}
  \item{adj.rr.label}{Adjusted \eqn{R^2} of the fitted model as a character string
  to be parsed}
  \item{hjust}{Set to zero to override the default of the "text" geom.}}
}
\keyword{datasets}

