% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-quant-band.r
\name{stat_quant_band}
\alias{stat_quant_band}
\title{Predicted band from quantile regression fits}
\usage{
stat_quant_band(
  mapping = NULL,
  data = NULL,
  geom = "smooth",
  position = "identity",
  ...,
  orientation = NA,
  quantiles = c(0.25, 0.5, 0.75),
  formula = NULL,
  fit.seed = NA,
  fm.values = FALSE,
  n = 80,
  method = "rq",
  method.args = list(),
  n.min = 3L,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}()}. Only needs to be set at the layer level if you
are overriding the plot defaults.}

\item{data}{A layer specific dataset, only needed if you want to override the
plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points on this
layer.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}

\item{orientation}{character Either "x" or "y" controlling the default for
\code{formula}. The letter indicates the aesthetic considered the
explanatory variable in the model fit.}

\item{quantiles}{A numeric vector of length 3, with unique values in
\eqn{0\ldots 1}. The three quantile regressions are mapped to \code{y},
\code{ymax} and \code{ymin} aesthetics, and by default plotted as a line
and band.}

\item{formula}{a formula object. Using aesthetic names \code{x} and \code{y}
instead of original variable names.}

\item{fit.seed}{RNG seed argument passed to
\code{\link[base:Random]{set.seed}()}. Defaults to \code{NA}, indicating
that \code{set.seed()} should not be called.}

\item{fm.values}{logical Add metadata and parameter estimates extracted from
the fitted model object; \code{FALSE} by default.}

\item{n}{Number of points at which to predict with the fitted model.}

\item{method}{function or character If character, "rq", "rqss" or the name of
a model fit function are accepted, possibly followed by the fit function's
\code{method} argument separated by a colon (e.g. \code{"rq:br"}). If a
function different to \code{rq()}, it must accept arguments named
\code{formula}, \code{data}, \code{weights}, \code{tau} and \code{method}
and return a model fit object of class \code{rq}, \code{rqs} or
\code{rqss}.}

\item{method.args}{named list with additional arguments passed to
\code{rq()}, \code{rqss()} or to another function passed as argument to
\code{method}.}

\item{n.min}{integer Minimum number of distinct values in the explanatory
variable (on the rhs of formula) for fitting to the attempted.}

\item{na.rm}{a logical indicating whether NA values should be stripped before
the computation proceeds.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}
}
\value{
The value returned by the statistic is a data frame, that will have
  \code{n} rows of predicted values for three quantiles as \code{y},
  \code{ymin} and \code{ymax}, plus \code{x}.
}
\description{
Predicted values are computed and plotted as a band plus an optional line
within. Using the default quantiles, the band plotted is similar to a boxplot
in its interpretation.
}
\details{
\code{stat_quant_band()} fits quantile regression and obtains
  predictions identically to \code{stat_quant_line()}.
  \code{stat_quant_band()} fits 2 or 3 quantiles in the same plot layer
  always display the area between the predicted regression lines for the
  extreme quantiles as a band. In contrast \code{stat_quant_line()} fits one
  or more regressions, adding a line for each. \code{stat_quant_line()}
  displays confidence bands for the regression lines when \code{se = TRUE}.

  While \code{stat_quant_eq()} supports only method \code{"rq"},
  \code{stat_quant_line()} and \code{stat_quant_band()} support both
  \code{"rq"} and \code{"rqss"}, In the case of \code{"rqss"} the model
  formula makes normally use of \code{qss()} to formulate the spline and its
  constraints.

  The minimum number of observations with distinct values in the explanatory
  variable can be set through parameter \code{n.min}. The default \code{n.min
  = 3L} is the smallest usable value. However, model fits with very few
  observations are of little interest and using larger values of \code{n.min}
  than the default is wise.

  There are multiple uses for double regression on \code{x} and \code{y}. For
  example, when two variables are subject to mutual constrains, it is useful
  to consider both of them as explanatory and interpret the relationship
  based on them. 'ggpmisc' (>= 0.4.1) supports \code{orientation} making it
  easy implement the approach described by Cardoso (2019) under the name of
  "Double quantile regression".
}
\section{Model formula and model fitting}{

  A ggplot statistic receives as \code{data} a data frame that is not the one
  passed as argument by the user, but instead a data frame with the variables
  mapped to aesthetics. In \code{stat_poly_eq()} the compute function is
  applied by group, each call "seeing" the subset of \code{data} for an
  individual group. As supported models are for regression lines,
  variables mapped to \code{x} and \code{y} should both be continuous, i.e.,
  numeric or date time and model formulas defined using \code{x} and \code{y}
  as variable names.

  The interpretation of the argument passed to \code{formula} is enhanced
  compared to \code{stat_smooth()}. Formulas with \code{x} as explanatory
  variable work as in \code{stat_smooth()} but formulas with \code{y} as
  explanatory variable are also accepted. \code{orientation} is set
  automatically based on which explanatory variable appears in the formula.
  Spline-based smoothers are only partially supported.
}

\section{Model fit methods supported}{
 Several model fit functions are
  supported explicitly (see tables), and some of their differences smoothed
  out. Compatibility is checked late, based on the class of the returned
  fitted model object. This makes it possible to use wrapper functions that
  do model selection or other adjustments to the fit procedure on a per panel
  or per group basis. Moreover, if the value returned as model fit object is
  \code{NULL}, no layer is added to the plot on a per group within panel
  basis.

  In the case of fitted model objects of classes not explicitly supported an
  attempt is made to find the usual accessors and/or fitted object members,
  and if found, either complete or partial support is frequently achieved.
  In this case a message is issued encouraging users to check the validity
  of the values extracted.

  The argument to parameter \code{method} can be either the name of a
  function object, possibly using double colon notation in case its package
  is not attached, or a character string matching the function name for
  functions in the search path. This approach makes it possible to support
  model fit functions that are not dependencies of 'ggpmisc'. Either by
  attaching the package where the function is defined and passing it by name
  or as string, or using double colon notation when passing the name of the
  function.

  User-defined functions can be passed as argument to parameter
  \code{method} as long as they have parameters \code{formula}, \code{data}
  \code{subset} and possibly \code{weights}. Additional arguments can be
  passed to any method as a named list as argument to parameter
  \code{method.args}. As in \code{\link[ggplot2:geom_smooth]{stat_smooth}()}
  prior \code{weights} are passed to the model fit functions' \code{weights}
  (plural!) parameter by mapping a numeric variable to plot aesthetic
  \code{weight} (singular!).

  Tables 1 lists natively supported model fit functions, with the
  caveat that only some 'broom' methods' specializations have been actually
  tested with statistics from 'ggpmisc'. In addition, the statistics based
  on 'broom' methods require the user to tailor their behaviour by passing
  additional arguments in the call and occasionally some detective work to
  find out the names of variables in the returned data frame.

  \strong{Table 1.} Model fit methods supported by the different statistics
  available in package 'ggpmisc'. Column \eqn{f} indicates whether
  computations are done by group (G) or by plot panel (P).
  \tabular{lcl}{
  \strong{Statistic} \tab \eqn{f} \tab \strong{Supported model fit methods} \cr
  \code{\link{stat_poly_line}()} \tab G \tab "lm", "rlm", "lts", "sma", "ma", "gls", \emph{others with methods} \code{\link[stats]{predict}()} or \code{\link[stats]{fitted}()} \cr
  \code{\link{stat_poly_eq}()}   \tab G \tab "lm", "rlm", "lts", "sma", "ma", "gls",  \emph{others with needed accesors} \cr
  \code{\link{stat_quant_line}()} \tab G \tab "rq", "rqss" \cr
  \code{\link{stat_quant_band}()} \tab G \tab "rq", "rqss" \cr
  \code{\link{stat_quant_eq}()} \tab G \tab "rq", "rqss" \cr
  \code{\link{stat_ma_line}()} \tab G \tab "SMA", "MA", "RMA", "OLS" \cr
  \code{\link{stat_ma_eq}()} \tab G \tab "SMA", "MA", "RMA", "OLS" \cr
  \code{\link{stat_fit_residuals}()} \tab G \tab "lm", "rlm", "lts", "sma", "ma", "gls", "rq", "rqss" \emph{others with method} \code{\link[stats]{residuals}()} \cr
  \code{\link{stat_fit_fitted}()} \tab G \tab "lm", "rlm", "lts", "gls", "rq", "rqss" \emph{others with method} \code{\link[stats]{fitted}()} \cr
  \code{\link{stat_fit_deviations}()} \tab G \tab "lm", "rlm", "lts", "gls", "rq", "rqss" \emph{others with methods} \code{\link[stats]{fitted}()} and \code{\link[stats]{weights}()} \cr
  \code{\link{stat_fit_augment}()} \tab G \tab \emph{any with 'broom' method} \code{\link[broom]{augment}()} \cr
  \code{\link{stat_fit_glance}()} \tab G \tab \emph{any with 'broom' method} \code{\link[broom]{glance}()} \cr
  \code{\link{stat_fit_tidy}()} \tab G \tab \emph{any with 'broom' method} \code{\link[broom]{tidy}()} \cr
  \code{\link{stat_fit_tb}()} \tab P \tab \emph{any with 'broom' method} \code{\link[broom]{tidy}()} \cr
  }

  The single colon notation is based on parsing
  the name and is available whenever passing the name of the fit method as a
  character string. In a string such as "head:tail" the "head" gives the name
  of the model fit function and the "tail" gives the argument to pass it's
  \code{method} parameter. This is only a convenience, as \code{method.args}
  can be also used. In some methods, i.e., splines, the default
  \code{formula = y ~ x} needs to be overridden with an explicit argument.

  Table 2 lists the correspondence of pre-defined \emph{method names}
  to model fit method functions. As mentioned above, these are only
  a subset of the model fit methods that are expected to work. When using
  these names there is no need for users to attach additional packages but
  the packages must be available (installed).

  \strong{Table 2.} Available predefined method names, the model fit functions
  they call, the packages where the functions reside, the class of the
  returned fitted model object and the arguments that can be
  passed to their \code{method} parameter using single colon notation.
  \tabular{llll}{
  \strong{Predefined method names} \tab \strong{Model fit methods} \tab \strong{R package} \tab \strong{Object class} \cr
  "lm", "lm:qr" \tab \code{\link[stats]{lm}()} \tab 'stats' \tab "lm" \cr
  "rlm", "rlm:M", "rlm:MM" \tab \code{\link[MASS]{rlm}()} \tab 'MASS' \tab "rlm" ("lm") \cr
  "lts", "ltsReg" \tab \code{\link[robustbase]{ltsReg}()} \tab 'robustbase' \tab "lts" \cr
  "ma", "sma", "sma:SMA", "sma:MA", "sma:OLS" \tab \code{\link[smatr]{sma}()} \tab 'smatr' \tab "ma" or "sma" \cr
  "gls", "gls:REML", "gls:ML" \tab \code{\link[nlme]{gls}()} \tab 'nlme' \tab "gls" \cr
  "rq", "rq:sfn", "rq:sfnc", "rq:lasso" \tab \code{\link[quantreg]{rq}()} \tab 'quantreg' \tab "rq" \cr
  "rqss", "rqss:sfn", "rqss:sfnc", "rqss:lasso" \tab \code{\link[quantreg]{rqss}()} \tab 'quantreg' \tab "rqss" \cr
  "SMA", "MA", "RMA", "OLS" \tab \code{\link[lmodel2]{lmodel2}()} \tab 'lmodel2' \tab ("list") \cr
  }
}

\examples{
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  stat_quant_band()

# If you need the fitting to be done along the y-axis set the orientation
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  stat_quant_band(orientation = "y")

ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  stat_quant_band(formula = y ~ x)

ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  stat_quant_band(formula = x ~ y)

ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  stat_quant_band(formula = y ~ poly(x, 3))

ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  stat_quant_band(formula = x ~ poly(y, 3))

# Instead of rq() we can use rqss() to fit an additive model:
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  stat_quant_band(method = "rqss",
                  formula = y ~ qss(x))

ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  stat_quant_band(method = "rqss",
                  formula = x ~ qss(y, constraint = "D"))

# Regressions are automatically fit to each group (defined by categorical
# aesthetics or the group aesthetic) and for each facet.

ggplot(mpg, aes(displ, hwy, colour = class)) +
  geom_point() +
  stat_quant_band(formula = y ~ x)

ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  stat_quant_band(formula = y ~ poly(x, 2)) +
  facet_wrap(~drv)

ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  stat_quant_band(linetype = "dashed", color = "darkred", fill = "red")

ggplot(mpg, aes(displ, hwy)) +
  stat_quant_band(color = NA, alpha = 1) +
  geom_point()

ggplot(mpg, aes(displ, hwy)) +
  stat_quant_band(quantiles = c(0, 0.1, 0.2)) +
  geom_point()

# Inspecting the returned data using geom_debug_group()
gginnards.installed <- requireNamespace("gginnards", quietly = TRUE)

if (gginnards.installed)
  library(gginnards)

if (gginnards.installed)
  ggplot(mpg, aes(displ, hwy)) +
    stat_quant_band(geom = "debug_group")

if (gginnards.installed)
  ggplot(mpg, aes(displ, hwy)) +
    stat_quant_band(geom = "debug_group", fm.values = TRUE)

}
\references{
Cardoso, G. C. (2019) Double quantile regression accurately assesses
  distance to boundary trade-off. Methods in ecology and evolution,
  10(8), 1322-1331.
}
\seealso{
\code{\link[quantreg]{rq}}, \code{\link[quantreg]{rqss}} and
  \code{\link[quantreg]{qss}}.

Other 'ggpmisc' statistics for model fits:
\code{\link[=stat_distrmix_eq]{stat_distrmix_eq()}},
\code{\link[=stat_distrmix_line]{stat_distrmix_line()}},
\code{\link[=stat_fit_deviations]{stat_fit_deviations()}},
\code{\link[=stat_fit_glance]{stat_fit_glance()}},
\code{\link[=stat_fit_residuals]{stat_fit_residuals()}},
\code{\link[=stat_fit_tb]{stat_fit_tb()}},
\code{\link[=stat_fit_tidy]{stat_fit_tidy()}},
\code{\link[=stat_ma_eq]{stat_ma_eq()}},
\code{\link[=stat_ma_line]{stat_ma_line()}},
\code{\link[=stat_poly_eq]{stat_poly_eq()}},
\code{\link[=stat_poly_line]{stat_poly_line()}},
\code{\link[=stat_quant_eq]{stat_quant_eq()}},
\code{\link[=stat_quant_line]{stat_quant_line()}}
}
\concept{'ggpmisc' statistics for model fits}
\section{Aesthetics}{

\code{stat_quant_band()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → inferred \cr
}

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}
