% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-quant-band.r, R/stat-quant-eq.R,
%   R/stat-quant-line.r
\name{stat_quant_band}
\alias{stat_quant_band}
\alias{stat_quant_eq}
\alias{stat_quant_line}
\title{Quantile regression predictions and annotations}
\usage{
stat_quant_band(
  mapping = NULL,
  data = NULL,
  geom = "smooth",
  position = "identity",
  ...,
  orientation = NA,
  quantiles = c(0.25, 0.5, 0.75),
  formula = NULL,
  fit.seed = NA,
  fm.values = FALSE,
  n = 80,
  fullrange = FALSE,
  limit.to = NULL,
  method = "rq",
  method.args = list(),
  n.min = 3L,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_quant_eq(
  mapping = NULL,
  data = NULL,
  geom = "text_npc",
  position = "identity",
  ...,
  orientation = NA,
  formula = NULL,
  quantiles = c(0.25, 0.5, 0.75),
  method = "rq:br",
  method.args = list(),
  n.min = 10L,
  fit.seed = NA,
  eq.with.lhs = TRUE,
  eq.x.rhs = NULL,
  coef.digits = 3,
  coef.keep.zeros = TRUE,
  decreasing = getOption("ggpmisc.decreasing.poly.eq", FALSE),
  rho.digits = 4,
  label.x = "left",
  label.y = "top",
  hstep = 0,
  vstep = NULL,
  output.type = NULL,
  na.rm = FALSE,
  parse = NULL,
  show.legend = FALSE,
  inherit.aes = TRUE
)

stat_quant_line(
  mapping = NULL,
  data = NULL,
  geom = "smooth",
  position = "identity",
  ...,
  orientation = NA,
  quantiles = c(0.25, 0.5, 0.75),
  formula = NULL,
  se = length(quantiles) == 1L,
  fit.seed = NA,
  fm.values = FALSE,
  n = 80,
  fullrange = FALSE,
  limit.to = NULL,
  method = "rq",
  method.args = list(),
  n.min = 3L,
  level = 0.95,
  type = "direct",
  interval = "confidence",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}()}. Only needs to be set at the layer level if you
are overriding the plot defaults.}

\item{data}{A layer specific dataset, only needed if you want to override the
plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points on this
layer.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}

\item{orientation}{character Either "x" or "y" controlling the default for
\code{formula}. The letter indicates the aesthetic considered the
explanatory variable in the model fit.}

\item{quantiles}{numeric vector Values in 0..1 indicating the quantiles.}

\item{formula}{a formula object. Using aesthetic names \code{x} and \code{y}
instead of original variable names.}

\item{fit.seed}{RNG seed argument passed to
\code{\link[base:Random]{set.seed}()}. Defaults to \code{NA}, indicating
that \code{set.seed()} should not be called.}

\item{fm.values}{logical Add metadata and parameter estimates extracted from
the fitted model object; \code{FALSE} by default.}

\item{n}{Number of points at which to predict with the fitted model.}

\item{fullrange}{logical Should the fit prediction span the full
range of the plot, or just the range of the explanatory variable?}

\item{limit.to}{character or numeric If character one of \code{""},
\code{"x"}, \code{"y"} or \code{"xy"}. Should the fit prediction be
constrained to the range of the variables mapped to \code{x} and/or
\code{y} in each data group? If numeric, the \emph{new data} values to use
for the explanatory variable when computing the predicted line and
confidence band. When set, \code{limit.to} silently overrides
\code{fullrange}!}

\item{method}{function or character If character, "rq", "rqss" or the name of
a model fit function are accepted, possibly followed by the fit function's
\code{method} argument separated by a colon (e.g. \code{"rq:br"}). If a
function different to \code{rq()}, it must accept arguments named
\code{formula}, \code{data}, \code{weights}, \code{tau} and \code{method}
and return a model fit object of class \code{rq}, \code{rqs} or
\code{rqss}.}

\item{method.args}{named list with additional arguments passed to
\code{rq()}, \code{rqss()} or to another function passed as argument to
\code{method}.}

\item{n.min}{integer Minimum number of distinct values in the explanatory
variable (on the rhs of formula) for fitting to the attempted.}

\item{na.rm}{a logical indicating whether NA values should be stripped before
the computation proceeds.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}

\item{eq.with.lhs}{If \code{character} the string is pasted to the front of
the equation label before parsing or a \code{logical} (see note).}

\item{eq.x.rhs}{\code{character} this string will be used as replacement for
\code{"x"} in the model equation when generating the label before parsing
it.}

\item{coef.digits, rho.digits}{integer Number of significant digits to use for
the fitted coefficients and rho in labels.}

\item{coef.keep.zeros}{logical Keep or drop trailing zeros when formatting
the fitted coefficients and F-value.}

\item{decreasing}{logical It specifies the order of the terms in the
returned character string; in increasing (default) or decreasing powers.}

\item{label.x, label.y}{\code{numeric} with range 0..1 "normalized parent
coordinates" (npc units) or character if using \code{geom_text_npc()} or
\code{geom_label_npc()}. If using \code{geom_text()} or \code{geom_label()}
numeric in native data units. If too short they will be recycled.}

\item{hstep, vstep}{numeric in npc units, the horizontal and vertical step
used between labels for different groups.}

\item{output.type}{character One of "expression", "LaTeX", "text",
"markdown" or "numeric".}

\item{parse}{logical Passed to the geom. If \code{TRUE}, the labels will be
parsed into expressions and displayed as described in
\code{\link[grDevices]{plotmath}}. Default is \code{TRUE} if
\code{output.type = "expression"} and \code{FALSE} otherwise.}

\item{se}{logical Passed to \code{quantreg::predict.rq()}.}

\item{level}{numeric in range [0..1] Passed to \code{quantreg::predict.rq()}.}

\item{type}{character Passed to \code{quantreg::predict.rq()}.}

\item{interval}{character Passed to \code{quantreg::predict.rq()}.}
}
\value{
\code{stat_quant_eq()} returns a data frame, with one row per
  quantile and columns as described below, while \code{stat_quant_line()}
  and \code{stat_quant_band()} return a data frame, with \code{n} rows per
  quantile and columns as described below. If the number of observations
  is less than \code{n.min} or if the model fit method returns \code{NA} or
  \code{NULL}, a data frame with no rows or columns is returned, resulting
  in an empty/invisible plot layer.
}
\description{
Statistics \code{stat_quant_line()}, \code{stat_quant_band()} and
\code{stat_quant_eq()} fit models by quantile regression. While
\code{stat_quant_line()} and \code{stat_quant_band()} add prediction lines and
bands, \code{stat_quant_eq()} adds textual labels to a plot.
}
\details{
While \code{stat_poly_line()} and \code{stat_poly_eq()} fit
  a single model per plot layer, \code{stat_quant_line()}, \code{stat_quant_band()}
  and \code{stat_quant_eq()} can fit multiple models sharing the same
  \code{method} and \code{formula} but differing in their
  probability. These probabilities are passed a vector argument to parameter
  \code{quantiles}.

  \code{stat_quant_line} fits one or more quantile regressions and obtains
  predictions similarly to \code{\link[ggplot2]{stat_quantile}()} from
  'ggplot2', but in addition it computes confidence regions for the
  prediction lines. By default each quantile is plotted as a line, with a
  confidence band when \code{se = TRUE}.

  \code{stat_quant_band()} fits quantile regressions and obtains predictions
  identically to \code{stat_quant_line()}. \code{stat_quant_band()} fits 2 or
  3 quantiles in the same plot layer and displays the area between the
  predicted regression lines for the extreme quantiles as a band.

  \code{stat_quant_eq()} fits quantile regressions and generates a set of
  labels for each regression line fitted. By default the labels are formatted
   as R's \code{\link[grDevices]{plotmath}} expressions, LaTeX and
   markdown are also supported.

  \code{stat_quant_eq()}, \code{stat_quant_line()} and
  \code{stat_quant_band()} support both \code{"rq"} and \code{"rqss"} as
  \code{method}. In the case of \code{"rqss"} the model formula makes
  normally use of \code{qss()} to formulate the spline and its constraints.
  User defined functions are supported as \code{method} as long as they
  accept arguments named \code{formula}, \code{data}, \code{weights},
  \code{tau} and \code{method} and return a model fit object of class
  \code{rq}, \code{rqs} or \code{rqss}. Such user-defined functions can
  implement model selection and/or method selection, or conditionally skip
  model fitting on a per data group basis.

  The minimum number of observations with distinct values in the explanatory
  variable can be set through parameter \code{n.min}. The default \code{n.min
  = 10L} is a bare minimum for quantile regression. Model fits with such a
  small number of observations are of little interest and using larger values
  of \code{n.min} than the default is wise.

  There are interesting uses for \emph{double quantile regression}, i.e., a
  pair of quantile regressions on \code{x} and \code{y} on the same data. For
  example, when two variables are subject to mutual constrains, it is useful
  to consider both of them as explanatory and interpret the relationship
  based on them considered as limiting. 'ggpmisc' (>= 0.4.1) supports
  \code{orientation} making it easy implement the approach described by
  Cardoso (2019) under the name of "Double quantile regression".
}
\section{Variables returned by \code{stat_quant_eq()}}{


If output.type is \code{"numeric"} the returned tibble contains columns
 in addition to a modified version of the original \code{group}:
\describe{
  \item{x,npcx}{x position}
  \item{y,npcy}{y position}
  \item{coef.ls}{list containing the "coefficients" matrix from the summary of the fit object}
  \item{rho, AIC, n}{numeric values extracted or computed from fit object}
  \item{rq.method}{character, method used.}
  \item{hjust, vjust}{Set to "inward" to override the default of the "text" geom.}
  \item{quantile}{Indicating the quantile  used for the fit}
  \item{quantile.f}{Factor with a level for each quantile}
  \item{b_0.constant}{TRUE is polynomial is forced through the origin}
  \item{b_i}{One or columns with the coefficient estimates}}

If output.type different from \code{"numeric"} the returned tibble contains
columns below in addition to a modified version of the original \code{group}:
\describe{
  \item{x,npcx}{x position}
  \item{y,npcy}{y position}
  \item{eq.label}{equation for the fitted polynomial as a character string to be parsed}
  \item{r.label, and one of cor.label, rho.label, or tau.label}{\eqn{rho} of the fitted model as a character string to be parsed}
  \item{AIC.label}{AIC for the fitted model.}
  \item{n.label}{Number of observations used in the fit.}
  \item{method.label}{Set according \code{method} used.}
  \item{rq.method}{character, method used.}
  \item{rho, n}{numeric values extracted or computed from fit object.}
  \item{hjust, vjust}{Set to "inward" to override the default of the "text" geom.}
  \item{quantile}{Numeric value of the quantile used for the fit}
  \item{quantile.f}{Factor with a level for each quantile}
  }

To explore the computed values returned for a given input we suggest the use
of \code{\link[gginnards]{geom_debug}} as shown in the example below.
}

\section{Variables returned by \code{stat_quant_line()}}{


  \describe{
  \item{y \strong{or} x}{predicted value}
  \item{ymin \strong{or} xmin}{lower confidence limit around the fitted line}
  \item{ymax \strong{or} xmax}{upper confidence limit around the fitted line}
  }

  If \code{fm.values = TRUE} is passed then one column with the number of
  observations \code{n} used for each fit is also included, with the same
  value in each row within a group. This is wasteful and disabled by default,
  but provides a simple and robust approach to achieve effects like colouring
  or hiding of the model fit line based on the number of observations.
}

\section{Variables returned by \code{stat_quant_band()}}{


  \describe{
  \item{y \strong{or} x}{Regression prediction for the middle quantile, if three quantiles are passed as argument}
  \item{ymin \strong{or} xmin}{Regression prediction for the smallest quantile}
  \item{ymax \strong{or} xmax}{Regression prediction for the largest quantile}
  }

  If \code{fm.values = TRUE} is passed then one column with the number of
  observations \code{n} used for each fit is also included, with the same
  value in each row within a group. This is wasteful and disabled by default,
  but provides a simple and robust approach to achieve effects like colouring
  or hiding of the model fit line based on the number of observations.
}

\section{Output types}{
 The formatting of character strings to be displayed
in plots are marked as mathematical equations. Depending on the geom used,
the mark-up needs to be encoded differently, or in some cases mark-up not
applied.
\describe{
  \item{\code{"expression"}}{The labels are encoded as character strings to be parsed into R's plotmath expressions.}
  \item{\code{"LaTeX", "TeX", "tikz", "latex"}}{The labels are encoded as 'LaTeX' maths equations, without the "fences" for switching in math mode.}
  \item{\code{"latex.eqn"}}{Same as \code{"latex"} but enclosed in single \code{$}, i.e., as in-line maths.}
  \item{\code{"latex.deqn"}}{Same as \code{"latex"} but enclosed in double \code{$$}, i.e., as display maths.}
  \item{\code{"markdown"}}{The labels are encoded as character strings using markdown syntax, with some embedded HTML.}
  \item{\code{"marquee"}}{The labels are encoded as character strings using markdown syntax, with 'marquee' supported spans.}
  \item{\code{"text"}}{The labels are plain ASCII character strings.}
  \item{\code{"numeric"}}{No labels are generated. This value is accepted by the statistics, but not by the label formatting functions.}
  \item{\code{NULL}}{The value used, \code{expression}, \code{latex.eqn} or \code{markup} depends on the argument passed to \code{geom}.}}

If \code{geom = "latex"} (package 'xdvir') the output type used is
\code{"latex.eqn"}. If \code{geom = "richtext"} (package 'ggtext') or
\code{geom = "textbox"} (package 'ggtext') the output type used is
\code{"markdown"}. If \code{geom = "marquee"} (package 'marquee') the output
type used is \code{"marquee"}. For all other values of \code{geom} the default
is \code{"expression"} unless the user passes an argument. Invalid values as
argument trigger an Error.
}

\section{Model equation label}{

  By default the equation label uses as symbols the names of the aesthetics,
  \code{x} and \code{y}. However, \code{"x"} and \code{"y"} can be
  substituted by providing a replacement character string for the
  right-hand-side and left-hand-side through \code{eq.x.rhs} and
  \code{eq.with.lhs}, respectively. For backward compatibility a logical is
  also accepted as argument for \code{eq.with.lhs}, with \code{FALSE}
  suppressing the left-hand-side.

  If the model \code{formula} includes a transformation of the explanatory
  variable in its right-hand-side (rhs), a matching argument should be passed
  to parameter \code{eq.x.rhs} as its default value would result in an
  equation label that does not reflect the applied transformation. In most
  cases, a transformation should not be applied within the left hand side
  (lhs) of the model formula, but instead in the mapping of the response
  variable within \code{aes}. In this case it may be necessary to also pass a
  matching argument to parameter \code{eq.with.lhs}.

  Parameter \code{orientation} is redundant as the orientation can be set
  by the \code{formula} but is included for consistency with
  \code{ggplot2::stat_smooth()}.
}

\section{Position of labels}{

  When data are grouped by mapping a factor to an aesthetic, e.g.,
  \code{colour}, \code{shape} and/or \code{linetype} the model is fitted
  separately to each group, and for each group a whole set of labels is
  generated. If the argument passed to \code{label.y} is a vector of length
  1, this value determines the position of the equation and/or other labels
  for the first group, and the positions of the labels for the remaining
  groups are generated by adding \code{vspace} based on the group number.
  If the argument passed to \code{label.y} is a vector of length > 1, it is
  used unchanged, possibly extended by recycling, ignoring \code{vstep}.

  If the labels are rotated by 90 degrees then the automatic stepping is
  best based on \code{hstep} with \code{vstep = 0}. Similarly as described
  above, if \code{label.x} is a vector of length > 1, it is
  used unchanged, possibly extended by recycling, ignoring \code{hstep}.

  When using facets and with a grouping that does not repeat in each panel,
  the automatic positioning in most cases will not be the desired one. Manual
  positioning using a vector of length > 1 for \code{label.x} and/or
  \code{label.y} is the currently available workaround.
}

\section{Model formula and model fitting}{

  A ggplot statistic receives as \code{data} a data frame that is not the one
  passed as argument by the user, but instead a data frame with the variables
  mapped to aesthetics. In \code{stat_poly_eq()} the compute function is
  applied by group, each call "seeing" the subset of \code{data} for an
  individual group. As supported models are for regression lines,
  variables mapped to \code{x} and \code{y} should both be continuous, i.e.,
  numeric or date time and model formulas defined using \code{x} and \code{y}
  as variable names.

  The interpretation of the argument passed to \code{formula} is enhanced
  compared to \code{stat_smooth()}. Formulas with \code{x} as explanatory
  variable work as in \code{stat_smooth()} but formulas with \code{y} as
  explanatory variable are also accepted. \code{orientation} is set
  automatically based on which explanatory variable appears in the formula.
  Spline-based smoothers are only partially supported.
}

\section{Range of the prediction line}{
 The range of the prediction line is
  controlled by parameters \code{fullrange} and \code{limit.to}.
  \code{fullrange} is backwards compatible both with earlier versions of
  'ggpmisc' and with \code{stat_smooth()} from 'ggplot2'; an argument passed
  to \code{limit.to} overrides \code{fullrange} making it possible to
  constrain the range to that of \code{x}, \code{y}, or both simultaneously,
  with \code{"x"}, \code{"y"}, or \code{"xy"}, respectively, as argument.
  \code{limit.to} also accepts a numeric vector of values to be used as
  \code{newdata} when computing the prediction. Limiting the range based on
  both aesthetics is the best approach for major axis regression (MA, SMA,
  RMA) but can occasionally be useful also with some other methods when
  slopes are very steep and error variance in the explanatory variable is
  large. A numeric vector can be used to predict the response at specific
  values of the explanatory variable. If a single or very few values are
  predicted, it can be necessary to override the default \code{geom =
  "smooth"} with \code{geom = "pointrange"}.
}

\section{Model fit methods supported}{

  Several model fit functions are supported explicitly (see tables), and some
  of their differences smoothed out. Compatibility is checked late, based on
  the class of the returned fitted model object. This makes it possible to
  use wrapper functions that do model selection or other adjustments to the
  fit procedure on a per panel or per group basis. Moreover, if the value
  returned as model fit object is \code{NULL} or \code{NA}, plotting is
  skipped on a per group within panel basis.

  \emph{In the case of fitted model objects of classes not explicitly
  supported, an attempt is made to find the usual accessors and/or fitted
  object members, and if found, either complete or partial support is
  frequently achieved. In this case a message is issued encouraging users to
  check the validity of the values extracted as the structure of fitted model
  objects belonging to different classes and the values returned by their
  accessors can vary, potentially resulting in decoding errors leading to the
  return of wrong values for estimates.}

  The argument to parameter \code{method} can be either the name of a
  function object, possibly using double colon notation in case its package
  is not attached, or a character string matching the function name for
  functions in the search path. This approach makes it possible to support
  model fit functions that are not dependencies of 'ggpmisc'. Either by
  attaching the package where the function is defined and passing it by name
  or as string, or using double colon notation when passing the name of the
  function.

  User-defined functions can be passed as argument to parameter \code{method}
  as long as they have parameters \code{formula}, \code{data} \code{subset}
  and possibly \code{weights}. Additional arguments can be passed to any
  method as a named list through parameter \code{method.args}. As in
  \code{\link[ggplot2:geom_smooth]{stat_smooth}()} prior \code{weights} are
  passed to the model fit functions' \code{weights} (plural!) parameter by
  mapping a numeric variable to plot aesthetic \code{weight} (singular!).

  Tables 1 lists natively supported model fit functions, with the
  caveat that only some 'broom' methods' specializations have been actually
  tested with statistics from 'ggpmisc'. In addition, the statistics based
  on 'broom' methods require the user to tailor their behaviour by passing
  additional arguments in the call and occasionally some detective work to
  find out the names of variables in the returned data frame as these names
  are set by methods from 'broom'.

  \strong{Table 1.} Model fit methods supported by the different statistics
  available in package 'ggpmisc'. Column \eqn{f} indicates whether
  computations are done by group (G) or by plot panel (P).
  \tabular{lcl}{
  \strong{Statistic} \tab \eqn{f} \tab \strong{Supported model fit methods} \cr
  \code{\link{stat_poly_line}()} \tab G \tab "lm", "rlm", "lts", "sma", "ma", "gls", \emph{others with methods} \code{\link[stats]{predict}()} or \code{\link[stats]{fitted}()} \cr
  \code{\link{stat_poly_eq}()}   \tab G \tab "lm", "rlm", "lts", "sma", "ma", "gls",  \emph{others with needed accesors} \cr
  \code{\link{stat_quant_line}()} \tab G \tab "rq", "rqss" \cr
  \code{\link{stat_quant_band}()} \tab G \tab "rq", "rqss" \cr
  \code{\link{stat_quant_eq}()} \tab G \tab "rq", "rqss" \cr
  \code{\link{stat_ma_line}()} \tab G \tab "SMA", "MA", "RMA", "OLS" \cr
  \code{\link{stat_ma_eq}()} \tab G \tab "SMA", "MA", "RMA", "OLS" \cr
  \code{\link{stat_fit_residuals}()} \tab G \tab "lm", "rlm", "lts", "sma", "ma", "gls", "rq", "rqss" \emph{others with method} \code{\link[stats]{residuals}()} \cr
  \code{\link{stat_fit_fitted}()} \tab G \tab "lm", "rlm", "lts", "gls", "rq", "rqss" \emph{others with method} \code{\link[stats]{fitted}()} \cr
  \code{\link{stat_fit_deviations}()} \tab G \tab "lm", "rlm", "lts", "gls", "rq", "rqss" \emph{others with methods} \code{\link[stats]{fitted}()} and \code{\link[stats]{weights}()} \cr
  \code{\link{stat_fit_augment}()} \tab G \tab \emph{any with 'broom' method} \code{\link[broom]{augment}()} \cr
  \code{\link{stat_fit_glance}()} \tab G \tab \emph{any with 'broom' method} \code{\link[broom]{glance}()} \cr
  \code{\link{stat_fit_tidy}()} \tab G \tab \emph{any with 'broom' method} \code{\link[broom]{tidy}()} \cr
  \code{\link{stat_fit_tb}()} \tab P \tab \emph{any with 'broom' method} \code{\link[broom]{tidy}()} \cr
  }

  The single colon notation is based on parsing
  the name and is available when passing the name of the fit method as a
  character string. In a string such as "head:tail" the "head" gives the name
  of the model fit function and the "tail" gives the argument to pass it's
  \code{method} parameter. This is only a convenience, as \code{method.args}
  can be also used. In some methods, i.e., splines, the default
  \code{formula = y ~ x} needs to be overridden by the user.

  Table 2 lists the correspondence of pre-defined \emph{method names}
  to model fit method functions. As mentioned above, these are only
  a subset of the model fit methods that are expected to work. When using
  these names there is no need for users to attach additional packages but
  the packages must be available (installed).

  \strong{Table 2.} Available predefined method names, the model fit functions
  they call, the packages where the functions reside, the class of the
  returned fitted model object and the arguments that can be
  passed to their \code{method} parameter using single colon notation.
  \tabular{llll}{
  \strong{Predefined method names} \tab \strong{Model fit methods} \tab \strong{R package} \tab \strong{Object class} \cr
  "lm", "lm:qr" \tab \code{\link[stats]{lm}()} \tab 'stats' \tab "lm" \cr
  "rlm", "rlm:M", "rlm:MM" \tab \code{\link[MASS]{rlm}()} \tab 'MASS' \tab "rlm" ("lm") \cr
  "lts", "ltsReg" \tab \code{\link[robustbase]{ltsReg}()} \tab 'robustbase' \tab "lts" \cr
  "ma", "sma", "sma:SMA", "sma:MA", "sma:OLS" \tab \code{\link[smatr]{sma}()} \tab 'smatr' \tab "ma" or "sma" \cr
  "gls", "gls:REML", "gls:ML" \tab \code{\link[nlme]{gls}()} \tab 'nlme' \tab "gls" \cr
  "rq", "rq:sfn", "rq:sfnc", "rq:lasso" \tab \code{\link[quantreg]{rq}()} \tab 'quantreg' \tab "rq" \cr
  "rqss", "rqss:sfn", "rqss:sfnc", "rqss:lasso" \tab \code{\link[quantreg]{rqss}()} \tab 'quantreg' \tab "rqss" \cr
  "SMA", "MA", "RMA", "OLS" \tab \code{\link[lmodel2]{lmodel2}()} \tab 'lmodel2' \tab ("list") \cr
  }
}

\examples{
# generate artificial data
set.seed(4321)
x <- 1:100
y <- (x + x^2 + x^3) + rnorm(length(x), mean = 0, sd = mean(x^3) / 4)
y <- y / max(y)
my.data <- data.frame(x = x, y = y,
                      group = c("A", "B"),
                      y2 = y * c(1, 2) + max(y) * c(0, 0.1),
                      w = sqrt(x))

# Predictions as lines
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line()

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line(quantiles = 0.5, se = TRUE)

# Predictions as band
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_band()

# y as explanatory variable (orientation = y)
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_band(formula = x ~ y)

# Using splines
library(quantreg)

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line(method = "rqss",
                  formula = y ~ qss(x, constraint = "D"),
                  quantiles = 0.5, se = FALSE)

# Adding annotations
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line() +
  stat_quant_eq()

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line() +
  stat_quant_eq(mapping = use_label("eq"))

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line() +
  stat_quant_eq(mapping = use_label("eq"), decreasing = TRUE)

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line() +
  stat_quant_eq(mapping = use_label("eq", "method"))

# same formula as default
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line(formula = y ~ x) +
  stat_quant_eq(formula = y ~ x)

# explicit formula "x explained by y"
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line(formula = x ~ y) +
  stat_quant_eq(formula = x ~ y)

# using color
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line(mapping = aes(color = after_stat(quantile.f))) +
  stat_quant_eq(mapping = aes(color = after_stat(quantile.f))) +
  labs(color = "Quantiles")

# location and colour
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line(mapping = aes(color = after_stat(quantile.f))) +
  stat_quant_eq(mapping = aes(color = after_stat(quantile.f)),
                label.y = "bottom", label.x = "right") +
  labs(color = "Quantiles")

# give a name to a formula
formula <- y ~ poly(x, 3, raw = TRUE)

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line(formula = formula, linewidth = 0.5) +
  stat_quant_eq(formula = formula)

# angle
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line(formula = formula, linewidth = 0.5) +
  stat_quant_eq(formula = formula, angle = 90, hstep = 0.04, vstep = 0,
                label.y = 0.02, hjust = 0, size = 3) +
  expand_limits(x = -15) # make space for equations

# user set quantiles
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line(formula = formula, quantiles = 0.5) +
  stat_quant_eq(formula = formula, quantiles = 0.5)

ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_band(formula = formula,
                  quantiles = c(0.1, 0.5, 0.9)) +
  stat_quant_eq(formula = formula, parse = TRUE,
                quantiles = c(0.1, 0.5, 0.9))

# grouping
ggplot(my.data, aes(x, y2, color = group)) +
  geom_point() +
  stat_quant_line(formula = formula, linewidth = 0.5) +
  stat_quant_eq(formula = formula)

ggplot(my.data, aes(x, y2, color = group)) +
  geom_point() +
  stat_quant_band(formula = formula, linewidth = 0.75) +
  stat_quant_eq(formula = formula) +
  theme_bw()

# labelling equations
ggplot(my.data, aes(x, y2,  shape = group, linetype = group,
       grp.label = group)) +
  geom_point() +
  stat_quant_band(formula = formula, color = "black", linewidth = 0.75) +
  stat_quant_eq(mapping = use_label("grp", "eq", sep = "*\": \"*"),
                formula = formula) +
  expand_limits(y = 3) +
  theme_classic()

# modifying the explanatory variable within the model formula
# modifying the response variable within aes()
formula.trans <- y ~ I(x^2)
ggplot(my.data, aes(x, y + 1)) +
  geom_point() +
  stat_quant_line(formula = formula.trans) +
  stat_quant_eq(mapping = use_label("eq"),
               formula = formula.trans,
               eq.x.rhs = "~x^2",
               eq.with.lhs = "y + 1~~`=`~~")

# using weights
ggplot(my.data, aes(x, y, weight = w)) +
  geom_point() +
  stat_quant_line(formula = formula, linewidth = 0.5) +
  stat_quant_eq(formula = formula)

# no weights, quantile set to upper boundary
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line(formula = formula, quantiles = 0.95) +
  stat_quant_eq(formula = formula, quantiles = 0.95)

# manually assemble and map a specific label using paste() and aes()
ggplot(my.data, aes(x, y2, color = group, grp.label = group)) +
  geom_point() +
  stat_quant_line(method = "rq", formula = formula,
                  quantiles = c(0.05, 0.5, 0.95),
                  linewidth = 0.5) +
  stat_quant_eq(mapping = aes(label = paste(after_stat(grp.label), "*\": \"*",
                                            after_stat(eq.label), sep = "")),
                quantiles = c(0.05, 0.5, 0.95),
                formula = formula, size = 3)

# manually assemble and map a specific label using sprintf() and aes()
ggplot(my.data, aes(x, y2, color = group, grp.label = group)) +
  geom_point() +
  stat_quant_band(method = "rq", formula = formula,
                  quantiles = c(0.05, 0.5, 0.95)) +
  stat_quant_eq(mapping = aes(label = sprintf("\%s*\": \"*\%s",
                                              after_stat(grp.label),
                                              after_stat(eq.label))),
                quantiles = c(0.05, 0.5, 0.95),
                formula = formula, size = 3)

# geom = "text"
ggplot(my.data, aes(x, y)) +
  geom_point() +
  stat_quant_line(formula = formula, quantiles = 0.5) +
  stat_quant_eq(label.x = "left", label.y = "top",
                formula = formula,
                quantiles = 0.5)

# Inspecting the returned data using geom_debug_group()
# This provides a quick way of finding out the names of the variables that
# are available for mapping to aesthetics using after_stat().

gginnards.installed <- requireNamespace("gginnards", quietly = TRUE)

if (gginnards.installed)
  library(gginnards)

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    stat_quant_line(geom = "debug_group")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    stat_quant_band(geom = "debug_group")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    geom_point() +
    stat_quant_eq(formula = formula, geom = "debug_group")

\dontrun{
if (gginnards.installed)
  ggplot(mpg, aes(displ, hwy)) +
    stat_quant_line(geom = "debug_group", fm.values = TRUE)

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    stat_quant_band(geom = "debug_group", fm.values = TRUE)

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    geom_point() +
    stat_quant_eq(mapping = aes(label = after_stat(eq.label)),
                  formula = formula, geom = "debug_group",
                  output.type = "markdown")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    geom_point() +
    stat_quant_eq(formula = formula, geom = "debug_group", output.type = "text")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    geom_point() +
    stat_quant_eq(formula = formula, geom = "debug_group", output.type = "numeric")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    geom_point() +
    stat_quant_eq(formula = formula, quantiles = c(0.25, 0.5, 0.75),
                  geom = "debug_group", output.type = "text")

if (gginnards.installed)
  ggplot(my.data, aes(x, y)) +
    geom_point() +
    stat_quant_eq(formula = formula, quantiles = c(0.25, 0.5, 0.75),
                  geom = "debug_group", output.type = "numeric")
}

}
\references{
Cardoso, G. C. (2019) Double quantile regression accurately assesses
  distance to boundary trade-off. Methods in ecology and evolution,
  10(8), 1322-1331.
}
\seealso{
\code{\link[quantreg]{rq}}, \code{\link[quantreg]{rqss}} and
  \code{\link[quantreg]{qss}}.

Other 'ggpmisc' statistics for model fits:
\code{\link[=stat_distrmix_eq]{stat_distrmix_eq()}},
\code{\link[=stat_fit_deviations]{stat_fit_deviations()}},
\code{\link[=stat_fit_glance]{stat_fit_glance()}},
\code{\link[=stat_fit_tb]{stat_fit_tb()}},
\code{\link[=stat_fit_tidy]{stat_fit_tidy()}},
\code{\link[=stat_ma_eq]{stat_ma_eq()}},
\code{\link[=stat_poly_eq]{stat_poly_eq()}}
}
\concept{'ggpmisc' statistics for model fits}
\section{Aesthetics}{

\code{stat_quant_line()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → \code{after_stat(group)} \cr
 • \tab \code{weight} \tab → \code{1} \cr
}

\code{stat_quant_band()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → inferred \cr
}

\code{stat_quant_eq()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → inferred \cr
 • \tab \code{grp.label} \tab   \cr
 • \tab \code{hjust} \tab → \code{"inward"} \cr
 • \tab \code{label} \tab → \code{after_stat(eq.label)} \cr
 • \tab \code{npcx} \tab → \code{after_stat(npcx)} \cr
 • \tab \code{npcy} \tab → \code{after_stat(npcy)} \cr
 • \tab \code{vjust} \tab → \code{"inward"} \cr
}

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}
