% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-spikes.R
\name{stat_spikes}
\alias{stat_spikes}
\title{Local narrow maxima or minima (spikes)}
\usage{
stat_spikes(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  orientation = "x",
  height.threshold = 20,
  z.threshold = 7,
  k = 20,
  spike.direction = "both",
  label.fmt = NULL,
  x.label.fmt = label.fmt,
  y.label.fmt = NULL,
  extract.spikes = NULL,
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. Only needs to be
set at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points on this
layer}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}

\item{orientation}{character The orientation of the layer can be set to
either \code{"x"}, the default, or \code{"y"}.}

\item{height.threshold}{numeric The minimum height of spikes expressed
relative to the median amplitude of the baseline local variation of
\code{x}.}

\item{z.threshold}{numeric Modified local \eqn{Z} values larger than
\code{z.threshold} are detected as boundaries of spikes.}

\item{k}{integer width of median window used for smoothing; must be odd}

\item{spike.direction}{character One of \code{"up"}, \code{"down"},
\code{"both"} or \code{"skip"}, indicating which spikes are to be returned,
if any.}

\item{label.fmt, x.label.fmt, y.label.fmt}{character strings giving a format
definition for construction of character strings labels with function
\code{\link{sprintf}} from \code{x} and/or \code{y} values.}

\item{extract.spikes}{If \code{TRUE} only the rows containing
spikes are returned. If \code{FALSE} the whole of \code{data} is
returned but with labels set to \code{""} in rows not containing spikes.
If \code{NULL}, the default, \code{TRUE}, is used unless the argument
passed to \code{geom} is \code{"text_repel"}, \code{"label_repel"}
or \code{"marquee_repel"}.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}
}
\value{
A data frame with one row for each spike found in the data
  extracted from the input \code{data} or all rows in data. Added columns
  contain the labels.
}
\description{
\code{stat_spikes()} tags or extracts rows in \code{data} containing local
\code{y} narrow maxima and/or minima with very steep shoulders. It makes it
possible to highlight and label spikes based on their \code{x} and/or
\code{y} coordinates. Orientations flipping as well as dates and times are
supported.
}
\details{
Spikes are detected based on a modified \eqn{Z} score calculated
  from the differenced spectrum. The \eqn{Z} threshold used should be
  adjusted to the characteristics of the input and desired sensitivity. The
  lower the threshold the more stringent the test becomes, with shorter
  spikes being detected.

  The algorithm uses running differences to detect abrupt changes in value,
  compared to an estimate of the baseline variation of the differences,
  approximating a baseline \eqn{Z} from MAD and a baseline value from the
  median differences. Currently, a single estimate of MAD is used but running
  medians, when posisble, as baseline. This comparison detects running
  differences that are unusually large, in most cases signalling a transition
  between values near the baseline and far from it, in both directions.

  Transitions into- and out of spikes are distinguished based on the median
  of the non-differenced values, as a descriptor of the data baseline. As for
  the median of the differences, a running median is used when possible.

  This function thus detects the start and end of each spike, and
  distinguishes upward and downward spikes.

  \code{k} is the width in number of observations of the window used for
  running median smoothing to extract the baseline. A value several times the
  width of the broader spike but narrow enough to track broader peaks needs
  to be manually set in most cases.

  With \code{na.rm = TRUE}, \code{NA} values are omitted before searching for
  spikes and set to \code{0L} in the returned vector.

  If all spikes are guaranteed to be one observation-wide and either going up
  or down from the baseline, it is possible to detect them based purely on
  the \code{z.threshold} by passing \code{height.threshold = NA} and either
  \code{spike.direction = "up"} or \code{spike.direction = "down"}, which
  ensures very fast computation.
}
\section{Computed and copied variables in the returned data frame}{

\describe{
  \item{x}{x-values at the spikes as numeric.}
  \item{y}{y-values at the spikes  as numeric.}
  \item{x.label}{x-values at the spikes formatted as character.}
  \item{y.label}{y-values at the spikes formatted as character.}
  \item{is.spike}{integer vector of \code{0}, \code{1} or \code{-1}.}
}
}

\section{Label positioning and formatting}{
 \code{stat_peaks()},
  \code{stat_valleys()} and \code{stat_spikes()} work nicely together with
  geoms \code{geom_text_repel()}, \code{geom_label_repel()}, and
  \code{geom_marquee_repel()} from package \code{\link[ggrepel]{ggrepel}} to
  solve the problem of overlapping labels by displacing them. If using
  \code{geom_text()}, discard overlapping labels using
  \code{check_overlap = TRUE}.

  By default the labels are character values ready to be ploted as plain
  text, but with a suitable \code{label.fmt} argument, labels formatted as
  \code{\link[grDevices]{plotmath}} expressions, markdown or LaTeX can be
  created (e.g., containing Greek letters or super or subscripts, maths or
  colour) can be generated for use with geoms from packages 'marquee',
  'ggtext' and 'xdvir'.

  The default is \code{geom = "point"} it is likely to work well in almost
  any situation. The default aesthetics mappings set by these stats allow
  their direct use with \code{geom_text()}, \code{geom_label()},
  \code{geom_line()}, \code{geom_rug()}, \code{geom_hline()} and
  \code{geom_vline()} by just passing an argument to \code{geom}.
}

\examples{
# lynx and Nile are time.series objects recognized by
# ggpp::ggplot.ts() and converted on-the-fly with a default mapping

n = 500
set.seed(45678)
my.data <- data.frame(x = 1:n,
                      y = rep(sin((0:19)/20 * 2 * pi), n / 20) +
                          stats::rnorm(n, sd = 0.5))
selector <- sample(seq_len(n), 5)
my.data$y[selector] <- my.data$y[selector] + 10

ggplot(my.data, aes(x, y)) +
  geom_line() +
  stat_spikes(colour = "orange")

ggplot(my.data, aes(x, -y)) +
  geom_line() +
  stat_spikes(colour = "orange")

ggplot(my.data, aes(x, y)) +
  geom_line() +
  stat_spikes(geom = "text", vjust = -0.5) +
  stat_spikes(geom = "rug", colour = "red")

ggplot(my.data, aes(x, y)) +
  geom_line() +
  stat_spikes(colour = "red", spike.direction = "up") +
  stat_spikes(colour = "blue", spike.direction = "down")

ggplot(my.data, aes(x, y)) +
  geom_line() +
  stat_spikes(colour = "red", spike.direction = "up")

ggplot(my.data, aes(x, y)) +
  geom_line() +
  stat_spikes(colour = "blue", spike.direction = "down")

ggplot(my.data, aes(x, y)) +
  geom_line() +
  stat_spikes(z.threshold = 2, colour = "orange")

ggplot(my.data, aes(x, y)) +
  geom_line() +
  stat_spikes(z.threshold = 20, colour = "orange")

ggplot(my.data, aes(x, y)) +
  geom_line() +
  stat_spikes(colour = "red",
              spike.direction = "up",
              height.threshold = NA)

# Inspecting the returned data using geom_debug_group()
# This provides a quick way of finding out the names of the variables that
# are available for mapping to aesthetics with after_stat().

gginnards.installed <- requireNamespace("gginnards", quietly = TRUE)

if (gginnards.installed)
  library(gginnards)

if (gginnards.installed)
ggplot(my.data, aes(x, y)) +
  geom_line() +
  stat_spikes(geom = "debug_group")

if (gginnards.installed)
ggplot(my.data, aes(x, y)) +
  geom_line() +
  stat_spikes(geom = "debug_group", extract.spikes = FALSE)

}
\references{
Whitaker, D. A.; Hayes, K. (2018) A simple algorithm for despiking Raman
spectra. Chemometrics and Intelligent Laboratory Systems, 179, 82-84.
\doi{10.1016/j.chemolab.2018.06.009}.
}
\seealso{
\code{\link{find_spikes}()}, for the function used to located the
  spikes.

Other \emph{statistics} for peak, valley and spikes annotation:
\code{\link[=stat_peaks]{stat_peaks()}}
}
\concept{\emph{statistics} for peak, valley and spikes annotation}
\section{Aesthetics}{

\code{stat_spikes()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → inferred \cr
 • \tab \code{label} \tab → \code{after_stat(x.label)} \cr
 • \tab \code{xintercept} \tab → \code{after_stat(x)} \cr
 • \tab \code{yintercept} \tab → \code{after_stat(y)} \cr
}

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}
