% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-continuous-logFC.r
\name{symmetric_limits}
\alias{symmetric_limits}
\title{Expand limits to be symetric}
\usage{
symmetric_limits(x)
}
\arguments{
\item{x}{numeric The automatic limits}
}
\value{
A numeric vector of length two with the new limits.
}
\description{
A simple function to expand scale limits to be symetric around zero. Can be
passed as argument to parameter \code{limits} of continuous scales from
packages 'ggplot2' or 'scales'.
}
\examples{

symmetric_limits(c(-1, 1.8))

}
