% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-label.R
\name{use_label}
\alias{use_label}
\alias{f_use_label}
\title{Assemble label and map it}
\usage{
use_label(..., labels = NULL, other.mapping = NULL, sep = "*\\", \\"*")

f_use_label(..., labels = NULL, other.mapping = NULL, format = NULL)
}
\arguments{
\item{...}{character Strings giving the names of at most six label components
in the order they will be included in the combined label.}

\item{labels}{character A vector with the name of at most six label
components. If provided, values passed through \code{...} are ignored.}

\item{other.mapping}{An unevaluated expression constructed with function
\code{\link[ggplot2]{aes}()} to be included in the returned value.}

\item{sep}{character A string used as separator when pasting the label
components together.}

\item{format}{character A string with as many %s place-holders as
labels being mapped. \strong{The format string must have mark-up matching the
\code{output.type} used!} With the default geometries,
\code{output.type = "expression"} requiring that the character string can}
}
\value{
A mapping to the \code{label} aesthetic and optionally additional
  mappings as an unevaluated R expression, built using function
  \code{\link[ggplot2]{aes}()}, ready to be passed as argument to the
  \code{mapping} parameter of the supported statistics.
}
\description{
Assemble model-fit-derived text or expressions and map them to
the \code{label} aesthetic.
}
\details{
Statistics \code{\link{stat_poly_eq}()}, \code{\link{stat_ma_eq}()},
  \code{\link{stat_quant_eq}()}, \code{\link{stat_distrmix_eq}()} and
  \code{\link{stat_correlation}()} return multiple text strings to be used
  individually or assembled into longer character strings depending on the
  labels actually desired. Assembling and mapping them requires verbose R
  code and familiarity with the syntax of the markup language used, by
  default R \emph{\link[grDevices]{plotmath}} expression syntax. Function
  \code{use_label()} automates these two tasks and accepts abbreviated
  familiar names for the parameters in addition to the name of the columns in
  the data object returned by the statistics. The default separator is
  suitable for \code{\link[grDevices]{plotmath}} expressions.

  These four statistics return several \code{character} variables with names
  ending in \code{.label}. This ending can be omitted, as well as
  \code{.value} for \code{f.value.label}, \code{t.value.label},
  \code{z.value.label}, \code{S.value.label} and \code{p.value.label}.
  \code{R2} can be used in place of \code{rr}. Furthermore, case is ignored.
  Thus, \code{use_label("eq", "R2")} is equivalent to
  \code{aes(label = paste(after_stat(eq.label), after_stat(rr.label),
  sep = "*\", \"*"))}.

  Function \code{use_label()} calls \code{aes()} to create a mapping for
  the \code{label} aesthetic to a text string assembled by calling
  \code{\link{paste}()}. Function \code{f_use_label()} uses
  \code{\link{sprintf}()} instead of \code{paste()} and expects a format
  string with the correct number of place-holders for the strings.
  Both functions can, when needed, combine this mapping to the \code{label}
  aesthetic with other mappings created with \code{aes()}.

  **When the number of labels does not match the number of fields in the
  format, or the label names do not match the names of variables in
  \code{data}, errors are triggered when the mapping is constructed, i.e.,
  at the time the plot is rendered.**
}
\examples{
# generate artificial data
set.seed(4321)
x <- 1:100
y <- (x + x^2 + x^3) + rnorm(length(x), mean = 0, sd = mean(x^3) / 4)
my.data <- data.frame(x = x,
                      y = y * 1e-5,
                      group = c("A", "B"),
                      y2 = y * 1e-5 + c(2, 0))

# give a name to a formula
formula <- y ~ poly(x, 3, raw = TRUE)

# default label constructed by use_label()
ggplot(data = my.data,
       mapping = aes(x = x, y = y2, colour = group)) +
  geom_point() +
  stat_poly_line(formula = formula) +
  stat_poly_eq(mapping = use_label(),
               formula = formula)

# user specified label components and default for sep
ggplot(data = my.data,
       mapping = aes(x = x, y = y2, colour = group)) +
  geom_point() +
  stat_poly_line(formula = formula) +
  stat_poly_eq(mapping = use_label("eq", "F"),
               formula = formula)

# user specified label components and separator
ggplot(data = my.data,
       mapping = aes(x = x, y = y2, colour = group)) +
  geom_point() +
  stat_poly_line(formula = formula) +
  stat_poly_eq(mapping = use_label("R2", "F", sep = "*\" with \"*"),
               formula = formula)

# user specified label components and format string
ggplot(data = my.data,
       mapping = aes(x = x, y = y2, colour = group)) +
  geom_point() +
  stat_poly_line(formula = formula) +
  stat_poly_eq(mapping =
                 f_use_label("R2", "F",
                             format = "\"Estimates: \"*\%s*\" with \"*\%s*\" works!\""),
               formula = formula)

# combine the mapping to the label aesthetic with other mappings
ggplot(data = my.data,
       mapping = aes(x = x, y = y2)) +
  geom_point(mapping = aes(colour = group)) +
  stat_poly_line(mapping = aes(colour = group), formula = formula) +
  stat_poly_eq(mapping = use_label("grp", "eq", "F",
                                   aes(grp.label = group)),
              formula = formula)

# combine other mappings with default labels
ggplot(data = my.data,
       mapping = aes(x = x, y = y2)) +
  geom_point(mapping = aes(colour = group)) +
  stat_poly_line(mapping = aes(colour = group), formula = formula) +
  stat_poly_eq(mapping = use_label(aes(colour = group)),
              formula = formula)

# example with other available components
ggplot(data = my.data,
       mapping = aes(x = x, y = y2, colour = group)) +
  geom_point() +
  stat_poly_line(formula = formula) +
  stat_poly_eq(mapping = use_label("eq", "adj.R2", "n"),
               formula = formula)

# multiple labels
ggplot(data = my.data,
       mapping = aes(x, y2, colour = group)) +
  geom_point() +
  stat_poly_line(formula = formula) +
  stat_poly_eq(mapping = use_label("R2", "F", "P", "AIC", "BIC"),
               formula = formula) +
  stat_poly_eq(mapping = use_label(c("eq", "n")),
               formula = formula,
               label.y = "bottom",
               label.x = "right")

# quantile regression
ggplot(data = my.data,
       mapping = aes(x, y)) +
  stat_quant_band(formula = formula) +
  stat_quant_eq(mapping = use_label("eq", "n"),
                formula = formula) +
  geom_point()

# major axis regression
ggplot(data = my.data, aes(x = x, y = y)) +
  stat_ma_line() +
  stat_ma_eq(mapping = use_label("eq", "n")) +
  geom_point()

# correlation
ggplot(data = my.data,
       mapping = aes(x = x, y = y)) +
  stat_correlation(mapping = use_label("r", "t", "p")) +
  geom_point()

}
\seealso{
Functions \code{use_label()} and \code{f_use_label()} are wrappers
  on function \code{\link[ggplot2]{aes}()} and can be used to create
  mappings to the \code{label} aesthetic. These functions are matched to
  \emph{statistics} for annotations from package 'ggpmisc'. Please,
  see their documentation for the formatted character labels they generate.

  Please, see the articles at
  \href{https://docs.r4photobiology.info/ggpmisc/}{online-only documentation}
  for additional use examples and guidance.
}
