% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{volcano_example.df}
\alias{volcano_example.df}
\title{Example gene expression data}
\format{A \code{data.frame} object with 1218 rows and 5 variables}
\usage{
volcano_example.df
}
\description{
A dataset containing reshaped and simplified output from an
  analysis of data from RNAseq done with package edgeR. Original data from
  gene expression in the plant species \emph{Arabidopsis thaliana}.
}
\examples{
colnames(volcano_example.df)
head(volcano_example.df)

}
\seealso{
Other Transcriptomics data examples: 
\code{\link{quadrant_example.df}}
}
\concept{Transcriptomics data examples}
\keyword{datasets}
