% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome2factor.r
\name{xy_outcomes2factor}
\alias{xy_outcomes2factor}
\title{Convert two numeric ternary outcomes into a factor}
\usage{
xy_outcomes2factor(x, y)
}
\arguments{
\item{x, y}{numeric vectors of -1, 0, and +1 values, indicating down
regulation, uncertain response or upregulation.}
}
\description{
Convert two numeric ternary outcomes into a factor
}
\details{
This function converts the numerically encoded values into a factor
  with the four levels \code{"xy"}, \code{"x"}, \code{"y"} and \code{"none"}, or
  into a factor with two levels \code{de} and \code{uncertain} as expected
  by scales \code{scale_colour_outcome()}, \code{scale_fill_outcome()} and
  \code{scale_shape_outcome()}.
}
\note{
When \code{n.levels = 2} both -1 and +1 are merged to the same level
  with label \code{"de"}.
}
\examples{

xy_outcomes2factor(c(-1, 0, 0, 1, -1), c(0, 1, 0, 1, -1))

}
\seealso{
Other scales for omics data: \code{\link{outcome2factor}},
  \code{\link{scale_colour_outcome}},
  \code{\link{scale_shape_outcome}},
  \code{\link{scale_x_logFC}}, \code{\link{scale_y_Pvalue}}
}
\concept{scales for omics data}
